% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{addCaption}
\alias{addCaption}
\title{Add caption to plot, and print it to the console}
\usage{
addCaption(x, caption)
}
\arguments{
\item{x}{a ggplot object}

\item{caption}{a character string}
}
\description{
\code{addCaption} appends a string to a ggplot object as attribute 'caption',
and prints it to the console. No copy is made.
}
\keyword{internal}

