% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{address}
\alias{address}
\title{Address in RAM of a variable}
\usage{
address(x)
}
\arguments{
\item{x}{anything}
}
\value{
A character vector length 1.
}
\description{
\code{address} returns the pointer address of its argument.
}
\details{
Sometimes useful in determining whether a value has been copied or 
not, programatically.
}
\note{
This is just an imported and re-exported function (without any 
modification) from \bold{data.table}.
}

