% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anova.R
\name{anovaTfce}
\alias{anovaTfce}
\title{TFCE-correction of the test statistic}
\usage{
anovaTfce(x, target_dim, has_neg, nr_chan, nr_time, tfce)
}
\arguments{
\item{x}{numeric matrix or array of test statistic
(chan_time X other_dimensions or chan X time X other_dimenions)}

\item{target_dim}{the dimension(s) which hold chan and time values (usually
1L if x is a matrix and 1:2L if x is an array)}

\item{has_neg}{logical scalar whether x might have negative values (TRUE for
a t-statistic and FALSE for an F-statistic)}

\item{nr_chan}{number of channels}

\item{nr_time}{number of time points}

\item{tfce}{a list of TFCE parameters}
}
\description{
\code{anovaTfce} is called internally by \code{arrayAnova} and
\code{arrayTtest} and performs TFCE-correction of the given test statistic
}
\keyword{internal}

