% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{argumentDeparser}
\alias{argumentDeparser}
\title{Pass \code{arg = .(key1 = value1, key2 = value2)} function arguments}
\usage{
argumentDeparser(arg, replace_dot, transform_true = TRUE,
  null_params = NULL)
}
\arguments{
\item{arg}{an unevaluated call, see Details}

\item{replace_dot}{a character string which defines the function to which
\code{arg} should be passed to}

\item{transform_true}{logical; if TRUE (default), a single logical
\code{arg} argument is treated specially (see Details)}

\item{null_params}{a list of parameters which is passed to the parameter
setter function (see 'replace_dot') if \code{argumentDeparser} would return
NULL}
}
\description{
\code{argumentDeparser} passes \code{arg = .(key1 = value1, key2 = value2)}
function arguments to the appropriate function
}
\details{
This function is not intended for direct use. It allows
\code{method = .(key = value)} argument definition in high-level functions
by substituting \code{.} to the appriopriate \code{methodParams}
function. If \code{transform_true} is TRUE, the call
\code{method = TRUE} is transformed to \code{method = methodParams()},
thereby it returns the default parameter setting.
}
\examples{
mymethodParams <- function(x = 3, y = 4) {
    list(x = x, y = y)
}
tempfn <- function(my_method = NULL) {
    args <- as.list(match.call()[-1])
    argumentDeparser(args$my_method, "mymethodParams")
}
stopifnot(is.null(tempfn()))
stopifnot(identical(FALSE, tempfn(my_method = FALSE)))
stopifnot(identical(tempfn(my_method = TRUE),
                    tempfn(my_method = mymethodParams())))
new_y = 1:5
stopifnot(identical(tempfn(my_method = .(y = new_y)),
                    tempfn(my_method = mymethodParams(y = new_y))))
}
\keyword{internal}

