% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/array_transform.R
\name{array2mat}
\alias{array2mat}
\title{Reshape array to matrix with specified row dimension}
\usage{
array2mat(dat, row_dim, return_attributes = TRUE, keep_dimnames = TRUE)
}
\arguments{
\item{dat}{an array to reshape}

\item{row_dim}{name or index of dimension which should be the row dimension
of the returned matrix}

\item{return_attributes}{logical value; if TRUE (default), the attributes
of the array is saved as an additional attribute ("array_atributes") of
the returned matrix including the row_dim parameter}

\item{keep_dimnames}{logical value; if TRUE (default), the function not
only reshapes the array but retains the dimension names separated by "|"}
}
\value{
A matrix
}
\description{
\code{array2mat} reshapes an array to a matrix
}
\seealso{
\code{\link{mat2array}} which is the inverse of array2mat
}

