% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/array_transform.R
\name{array_}
\alias{array_}
\title{Fast in-place transformation to an array (without copy)}
\usage{
array_(x, dim, dimnames = NULL, force_length = TRUE, arg_check = TRUE)
}
\arguments{
\item{x}{a data vector, matrix or array}

\item{dim}{the dim attribute for the array to be created, that is an integer
vector of length one or more giving the maximal indices in each dimension}

\item{dimnames}{either NULL or the names for the dimensions. This must a
list (or it will throw an error) with one component for each dimension,
either NULL or a character vector of the length given by dim for that
dimension. The list can be named, and the list names will be used as names
for the dimensions. If the list is shorter than the number of dimensions, it
is extended by NULLs to the length required.}

\item{force_length}{logical. If TRUE (the default), \code{array_} checks if
length(x)==nrow*ncol. If not, x is recycled or subsetted to the desired
length.}

\item{arg_check}{logical indicating if argument checks should be performed
(TRUE, the default). Do not set to FALSE unless you really know what you
are doing!}
}
\value{
This function (invisibly) returns an array (or a matrix if
\code{length(dim)==2L}) without duplicating the input values.
}
\description{
\code{array_} transforms its data argument to an array by reference. No
copy is made at all, and it invisibly returns the array.
}
\note{
Use \code{array_} with extra care because it modifies in place all
objects which x refers to. See \code{\link{matrix_}} for further hints.
}
\seealso{
\code{\link{matrix_}} for in-place transformation of x to a matrix
(without copy) and \code{\link{array}} for creating a new array without
modifying the original input
}

