% Generated by roxygen2 (4.0.2): do not edit by hand
\name{arrayAnovaSub}
\alias{arrayAnovaSub}
\title{Workhorse function of arrayAnova}
\usage{
arrayAnovaSub(a_dat, f_def, d_names, f_dat, verbose = TRUE)
}
\arguments{
\item{a_dat}{a numeric matrix of responses}

\item{f_def}{a named list of factor definitions, containing the following
elements:
\itemize{
\item{between:}{character vector of between-subject factors (default: NULL)}
\item{within:}{character vector of within-subject factors (default: NULL)}
\item{w_id:}{name of the dimension which identifies the subjects
(default: "id")}
}}

\item{d_names}{(list of) dimension names of the original array in
\code{arrayAnova}}

\item{f_dat}{a data.frame which contains the identification codes
(f_def$w_id) and the between- and/or within-subject factors}

\item{verbose}{logical value indicating if p-values and effect sizes should
be computed for the traditional ANOVA results}
}
\description{
\code{arrayAnovaSub} computes F-values in ANOVA designs, and provides
additional p-values, degrees of freedom, generalized effect sizes if
requested.
}
\keyword{internal}

