% Generated by roxygen2 (4.0.2): do not edit by hand
\name{arrayIP}
\alias{arrayIP}
\title{Fast in-place transformation to an array (without copy)}
\usage{
arrayIP(x, dim, dimnames = NULL, force_length = TRUE)
}
\arguments{
\item{x}{a data vector, matrix or array}

\item{dim}{the dim attribute for the array to be created, that is an integer
vector of length one or more giving the maximal indices in each dimension}

\item{dimnames}{either NULL or the names for the dimensions. This must a
list (or it will throw an error) with one component for each dimension,
either NULL or a character vector of the length given by dim for that
dimension. The list can be named, and the list names will be used as names
for the dimensions. If the list is shorter than the number of dimensions, it
is extended by NULLs to the length required.}

\item{force_length}{logical. If TRUE (the default), \code{arrayIP} checks if
length(x)==nrow*ncol. If not, x is recycled or subsetted to the desired
length.}
}
\value{
An array (invisibly) or a matrix if length(dim)==2L.
}
\description{
\code{arrayIP} transforms its data argument to an array by reference. No
copy is made at all, and it invisibly returns the array.
}
\note{
Use \code{arrayIP} with extra care because it modifies in place all
objects which x refers to. See \code{\link{matrixIP}} for further hints.
}
\seealso{
\code{\link{matrixIP}} for in-place transformation of x to a matrix
(without copy) and \code{\link{array}} for creating a new array without
modifying the original input
}

