% Generated by roxygen2 (4.0.2): do not edit by hand
\name{arrayTtest}
\alias{arrayTtest}
\title{Point-to-point t-tests (potentially with TFCE correction) on arrays}
\usage{
arrayTtest(arraydat, arraydat2, paired = FALSE, groups = NULL, mu = 0,
  var_equal = TRUE, id_dim = "id", verbose = TRUE, nperm = 999L,
  useparallel = TRUE, par_method = c("default", "snow", "multicore"),
  ncores = NULL, cl = NULL, usetfce = TRUE, tfce_options = NULL,
  seed = NULL)
}
\arguments{
\item{arraydat}{a numeric array with named dimnames containing EEG (or
other) data. Missing values are not allowed. Must have at least three
dimensions with names "chan", "time", and "id" (see also id_dim)}

\item{arraydat2}{a numeric array with named dimnames containing EEG (or
other) data. If provided, see the parameter \code{paired} for running
dependent or independent-samples t-tests}

\item{paired}{logical scalar, only used if arraydat2 is provided. If paired
is FALSE (default), the function computes independent samples t-tests,
otherwise paired samples t-tests are performed}

\item{groups}{provides an alternative (and more efficient) way to perform
independent samples t-tests; a character, factor, or integer vector which
defines group membership. Groups is ignored if arraydat2 is not missing.
NA values code subjects to drop.}

\item{mu}{a numeric scalar indicating the true value of the mean (or
difference between means, if two-sample tests are performed)}

\item{var_equal}{a logical scalar whether the variances are equal (only
relevant for independent-samples t-tests). If TRUE (default), the pooled
variance is used to estimate the variance, otherwise the Welch (or
Satterthwaite) approximation to the degrees of freedom is used.}

\item{id_dim}{name of the dimension which identifies the subjects
(default: "id")}

\item{verbose}{logical value indicating if p-values should be computed for
the traditional t-test results}

\item{nperm}{integer value giving the number of permutations (default: 999L)}

\item{useparallel}{logical value; if TRUE (default), computations are done
in parallel}

\item{par_method}{parallelization method; can be set explicitly to "snow" or
"multicore" (ignored on Windows OS), or chosen automatically ("default").
Ignored if cl is provided.}

\item{ncores}{integer value corresponding to the number of cores;
if NULL (default), it is set to the maximum number of cores available.
Ignored if cl is provided.}

\item{cl}{a cluster definition; if NULL (default), it is set up automatically}

\item{usetfce}{logical value whether TFCE (threshold-free cluster enhancement)
correction should also be computed (default: TRUE)}

\item{tfce_options}{a named list containing the channel neighbourhood matrix
(named ChN, no default) and the vector of the E/H parameters (named EH,
defaults to c(0.66, 2))}

\item{seed}{an integer value which specifies a seed (default: NULL)}
}
\value{
A list object with t-values, TFCE-corrected t-values and
permutation-based p-values (if requested)
}
\description{
\code{arrayTtest} performs point-to-point t-tests on arrays.
Permutation-based p-values and Threshold-free Cluster Enhancement (TFCE)
correction can be requested.
}
\details{
The function assumes that the input array contains at least three
named dimensions: chan (corresponding to the channels [electrodes]) and time
(corresponding to time points), and \code{id_dim} (corresponding to subjects).
All other dimensions are treated in a similar way as chan and time, that is
separate t-tests are computed for each level of those dimensions.
}

