% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/erp_preprocess.R
\name{artrejOptions}
\alias{artrejOptions}
\title{Options for artifact rejection}
\usage{
artrejOptions(sampling_freq = 1000, channels = "all",
  apply_maxgrad = TRUE, maxgrad_limit = 50, maxgrad_mark = c(-200, 200),
  apply_diffrange = TRUE, diffrange_limit = c(0.5, 100),
  diffrange_mark = c(-200, 200), diffrange_interval = 200,
  apply_amplrange = TRUE, amplrange_limit = c(-200, 200),
  amplrange_mark = c(-200, 200))
}
\arguments{
\item{sampling_freq}{numeric value, the sampling frequency of the EEG-data}

\item{channels}{character vector containing the name or index of channels
which are subject to artifact rejection. If set to "all" (default), all
channels are included.}

\item{apply_maxgrad}{logical value, if set to TRUE (default), the maximum
gradient criterion is applied.}

\item{maxgrad_limit}{numeric value, the maximum gradient / millisecond
(default: 50)}

\item{maxgrad_mark}{numeric vector of length 2; the placement of the Bad
Interval mark in milliseconds before and after the occurence of maxgrad_limit
violation (default: c(-200, 200))}

\item{apply_diffrange}{logical value, if set to TRUE (default), the difference
range criterion is applied.}

\item{diffrange_limit}{numeric vector of length 2, the minimum and maximum
voltage difference in a given interval (default: 200)}

\item{diffrange_mark}{numeric vector of length 2; the placement of the Bad
Interval mark in milliseconds before and after the occurence of diffrange_limit
violation (default: c(-200, 200))}

\item{diffrange_interval}{numeric value, the length of interval for the
difference range criterion in milliseconds (default: 200)}

\item{apply_amplrange}{logical value, if set to TRUE (default), the amplitude
range criterion is applied.}

\item{amplrange_limit}{numeric vector of length 2, the minimum and maximum
voltage in the whole segment (default: c(-200, 200))}

\item{amplrange_mark}{numeric vector of length 2; the placement of the Bad
Interval mark in milliseconds before and after the occurence of
amplrange_limit violation (default: c(-200, 200))}
}
\value{
A list object with all parameters.
}
\description{
\code{artrejOptions} allows to set the parameters of the artifact rejection
methods.
}
\details{
The short definitions of the possible artifact rejection criteria
are as follows:
\itemize{
\item{Maximum gradient:}{The absolute difference between the voltages
measured at successive milliseconds.}
\item{Difference range:}{The minimum and maximum difference between the
maximum and minimum voltages in a given sampling interval.}
\item{Amplitude range:}{The minimum and maximum voltages in the segments.}
}
}
\note{
The algorithm takes care of the sampling frequency for all parameters
which are provided in milliseconds (or /ms) and makes adjustments if needed.
However, *_mark parameters are not used since only segmented data can be
analyzed in the present version of artifactRejection().
}

