% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is.R
\name{asNumeric}
\alias{asNumeric}
\alias{asNumeric_}
\title{Coerce to numeric}
\usage{
asNumeric(x, keep_dim = FALSE)

asNumeric_(x)
}
\arguments{
\item{x}{the object to coerce}

\item{keep_dim}{logical value whether \code{dim} and \code{dimnames}
attributes should be preserved}
}
\value{
\code{asNumeric} returns the original object (at the same memory
address) if it is numeric (that is, is.numeric(x) returns TRUE) or inherits
"Date" or "POSIXt" classes. Otherwise, it returns a double vector if
\code{as.numeric} is successful and stops with an error if
\code{as.numeric} results in a warning or error.
}
\description{
\code{asNumeric} is similar to \code{\link{as.numeric}} except that 1)
it results in an error for character strings for which as.numeric would
send a warning, 2) keeps integer, Date and POSIXt objects as they are (i.e.
does not convert them to double), 3) keeps 'dim' and 'dimnames'
attributes if requested, and 4) always copies the original object, even if no
actual coercion occurs. \code{asNumeric_} returns the original object if
possible and does not alter the shape of multidimensional inputs.
}
\section{Functions}{
\itemize{
\item \code{asNumeric_}: 
}}
\examples{
# if 'x' is double, asNumeric() simply returns it, just like as.numeric()
x <- rnorm(10)
stopifnot(identical(as.numeric(x), asNumeric(x)))

# if 'x' is integer, asNumeric() simply returns it,
# but as.numeric() converts to double
x <- 1:10
stopifnot(identical(x, asNumeric(x)))
stopifnot(all.equal(as.numeric(x), asNumeric(x))) # not fully identical!

# if 'x' is Date or POSIXt, asNumeric() simply returns it
x1 <- as.Date("2013-03-14")
x2 <- as.POSIXlt(x1)
stopifnot(identical(x1, asNumeric(x1)))
stopifnot(identical(x2, asNumeric(x2)))

# if 'x' contains only numeric-like characters (or NAs),
# asNumeric() returns the same as as.numeric()
x <- c(NA, as.character(1:10))
stopifnot(identical(as.numeric(x), asNumeric(x)))

# if 'x' contains any non-numeric characters (e.g. letters),
# asNumeric() returns an error
x <- c("a", as.character(1:10))
x_num <- try(asNumeric(x), silent = TRUE)
stopifnot(inherits(x_num, "try-error"))

# if 'x' is a matrix or array, asNumeric() converts it to a vector by
# default, just like as.numeric()
( y <- x <- matrix(rnorm(10), 5, 2) )
( x_num <- asNumeric(x) )
stopifnot(is.null(attr(x_num, "dim")))

# however, you can ask to keep the original shape
( x_num2 <- asNumeric(x, TRUE) )

# if you use asNumeric_() instead, it keeps the shape
( x_num3 <- asNumeric_(x) )
stopifnot(identical(x_num3, x))

# note that since 'x' is numeric, no coercion is needed, so
# so asNumeric_() does not create a copy, but asNumeric() does
address(x)
address(x_num2)
address(x_num3)
stopifnot(identical(address(x), address(x_num3)))
}

