% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{assert0}
\alias{assert0}
\title{Assertion with NULL allowed}
\usage{
assert0(x, fun, ..., .var.name = NULL)
}
\arguments{
\item{x}{argument to check}

\item{fun}{a function from the checkmate package (or similar)}

\item{...}{arguments to fun}

\item{.var.name}{name for \code{x}. See \code{\link[checkmate]{assert}}}
}
\description{
\code{assert0} allows NULL besides other assertion conditions.
}
