% Generated by roxygen2 (4.0.2): do not edit by hand
\name{assignList}
\alias{assignList}
\title{Assign the elements of a named list to the enclosing environment}
\usage{
assignList(listdat, verbose = TRUE, overwriteGlobal = FALSE)
}
\arguments{
\item{listdat}{a list with named elements}

\item{verbose}{a logical variable (default: TRUE) which determines if
a warning should be sent to the console}

\item{overwriteGlobal}{a logical variable (default FALSE) which determines
if the function is allowed to write to the global environment}
}
\value{
The function is invoked for its side effect, which is assigning list
elements to the enclosing environment
}
\description{
\code{assignList} assigns the elements of a named list to the enclosing environment.
Be aware that the function does not check if an object with the same name exists in
the enclosing environment - if it exists, it will be overwritten.
}

