% Generated by roxygen2 (4.0.1): do not edit by hand
\name{avgDims}
\alias{avgDims}
\title{Collapse data in array format across specific dimensions}
\usage{
avgDims(dat, dims, na_rm = TRUE)
}
\arguments{
\item{dat}{matrix or array to average}

\item{dims}{character vector of dimension names or numeric vector of
dimension indices across which averaging should ocuur. If character vector,
dat must have named dimnames.}

\item{na_rm}{logical variable; if TRUE (default), missing values are removed
before averaging}
}
\value{
An array (or matrix) with averaged data (the number of dimensions of
dat is decreased by the length of dims)
}
\description{
\code{avgDims} collapses data in array format across specific dimensions by
simple averaging
}

