% Generated by roxygen2 (4.0.1): do not edit by hand
\name{baselineCorr}
\alias{baselineCorr}
\title{Baseline correction}
\usage{
baselineCorr(dat, basedim = "chan", baseind = NULL)
}
\arguments{
\item{dat}{numeric array containing the ERPs}

\item{basedim}{character vector identifying the dimensions of dat along which
separate baseline averaging should occur (default: "chan")}

\item{baseind}{numeric or character vector identifying time points which
form the baseline}
}
\value{
A numeric array with the same attributes as dat
}
\description{
\code{baselineCorr} performs baseline correction
}

