% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bindArrays}
\alias{bindArrays}
\title{Combine several arrays into one large array}
\usage{
bindArrays(..., along = NULL, rev.along = NULL, new.names = NULL,
  force.array = TRUE, make.names = use.anon.names, use.anon.names = FALSE,
  use.first.dimnames = FALSE, hier.names = FALSE, along_name = NULL)
}
\arguments{
\item{along_name}{a character version of \code{along} in
\code{\link{abind::abind}}. Only considered if input arrays have named
dimension names, and if conflicts with \code{along}, \code{along_name}
overrides \code{along}. If \code{along_name} is not among the dimension
names, it will be the name of the new dimension name.}
}
\description{
\code{bindArrays} is just a wrapper around \code{\link{abind::abind}}.
}
\details{
This function calls \code{\link{abind::abind}} and adds the names
of the dimension names of the arrays. If the inputs have named dimension
names, their dimensions are permuted before feeding to \code{abind}.
}
\seealso{
\code{\link{abind::abind}}
}

