% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{calculate_convex_hulls}
\alias{calculate_convex_hulls}
\title{Calculate convex hulls of grouped data in scatter plots}
\usage{
calculate_convex_hulls(df, xdim, ydim, group_by = NULL)
}
\arguments{
\item{df}{Data frame, containing at least two (continuous) dimensions}

\item{xdim}{Character singleton, name of x dim in which to calculate hulls}

\item{ydim}{Character singleton, name of y dim in which to calculate hulls}

\item{group_by}{Character singleton, name of grouping dimension}
}
\description{
This function takes a data frame and calculates the
convex hull (the outline) of the data in two dimensions (i.e. data
frame columns). Alternatively, you may provide a column \code{group_by}
to calculate the outline of each group individually. The functions
outputs a dataframe formatted specifically for use with \code{geom_polygon}
in the \code{ggplot2} package.
}
