% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erp_preprocess.R
\name{centroid}
\alias{centroid}
\title{Compute centroids}
\usage{
centroid(dat, ch_pos, proj2map = TRUE, proj_unitsphere = FALSE, ...)
}
\arguments{
\item{dat}{numeric vector of amplitudes at a given sampling point}

\item{ch_pos}{matrix or data.frame containing channel positions (in the same
order as dat), see \code{\link{coordinates}}}

\item{proj2map}{logical value; if TRUE (default), the centroids are projected
onto a 2D plane}

\item{proj_unitsphere}{logical value; if TRUE, and proj2map is also TRUE, the 
projection assumes unit radius}

\item{...}{additional parameters to \code{\link{project3dMap}}}
}
\value{
A data.frame containing the positions of the negative and positive 
centroids
}
\description{
\code{centroid} computes the centroids (separately for the negative and
positive values)
}

