% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changefs.R
\name{changefs}
\alias{changefs}
\title{Change Sampling Rate}
\usage{
changefs(ctd_obj, new_fs)
}
\arguments{
\item{ctd_obj}{object of class \code{'ctd'} containing continuous time domain
data.}

\item{new_fs}{positive scalar indicating the new sampling frequency.}
}
\value{
A \code{ctd} object containing the resampled data.
}
\description{
Change sampling rate of a \code{ctd} object.
}
\details{
Changing the sampling rate is done using the \code{resample} function in the
\code{gsignal} package.
}
\examples{
data(EEGdata)
down <- changefs(EEGdata, 100)
up <- changefs(EEGdata, 512)

}
\seealso{
\code{\link[gsignal]{resample}}
}
\author{
Geert van Boxtel, \email{G.J.M.vanBoxtel@gmail.com}
}
