% Generated by roxygen2 (4.0.1): do not edit by hand
\name{chanNb}
\alias{chanNb}
\title{Find channel neighbours}
\usage{
chanNb(ch_pos, alpha = NULL, ...)
}
\arguments{
\item{ch_pos}{electrode positions}

\item{alpha}{a numeric value which influences the allowed distance between
neighbouring electrodes}

\item{...}{parameters to \code{\link{sph2cart}}}
}
\value{
An electrode neighbourhood matrix
}
\description{
\code{chanNb} finds neighbouring channels
}

