% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is.R
\name{checkIsSubsetAttribute}
\alias{checkIsSubsetAttribute}
\title{Check the 'dim_index' attribute of the subset. argument}
\usage{
checkIsSubsetAttribute(subset)
}
\description{
\code{checkIsSubsetAttribute} is a tiny helper function to check the
'dim_index' attribute of the 'subset.' argument in \code{is*} functions.
}
\keyword{internal}

