% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{checkSubsetExpand}
\alias{checkSubsetExpand}
\title{Get 'dim_index' of the subset. and expand. arguments in the is* functions}
\usage{
checkSubsetExpand(subset., expand.)
}
\description{
\code{checkIsSubsetExpand} is a tiny helper function to check the
'dim_index' attributes of the 'subset.' and 'expand.' arguments in
\code{is*} functions. It also checks for potential dimension overlaps.
}
\keyword{internal}
