% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chunkify}
\alias{chunkify}
\title{Call a function on array chunks}
\usage{
chunkify(dat, fun, arg_list = NULL, chunks = NULL)
}
\arguments{
\item{dat}{matrix or array with named dimnames}

\item{fun}{name of the function (can be a character string). See details.}

\item{arg_list}{a list of parameters passed to \code{fun}}

\item{chunks}{a named list which defines the chunking scheme. See details.}
}
\value{
An array (or vector/matrix); its shape and dimension names are
identical to the return value of \code{fun} called on \code{dat} without
chunking
}
\description{
\code{chunkify} creates subsets of an array according to various chunking
schemes and performs the given function on these data chunks. The resulting
array is identical to the result of a function call on the whole array.
}
\details{
The main purpose of \code{chunkify} is twofold: 1) it can decrease
memory load if a memory-intensive function is called on a large array; 2) it
enables within-dimension computations (e.g. perform baseline correction
separately for various levels of the trial dimension). The parameter
\code{chunks} is a named list; each element must correspond to a named
dimension of \code{dat}. If the list element contains one integer, it gives
the number of chunks for the given dimension. If it contains as many elements
as the length of the given dimension, it defines chunk code; i.e.,
which chunk the given level of the dimension belongs to. Otherwise, the
list elements are expanded (recycled to the dimensions' length) and
interpreted as chunk codes. \code{fun} must be a function which operates on
arrays (or matrices), and returns a matrix or array with named dimensions.
Chunking can be based only on those dimensions which are not affected by
\code{fun}.
}

