% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matstat_utils.R
\name{collapse}
\alias{collapse}
\title{Collapse row or column elements of a matrix}
\usage{
collapse(x, col = TRUE)
}
\arguments{
\item{x}{vector, matrix or data.frame}

\item{col}{logical value; if TRUE (default), elements are collapsed 
column-wise.}
}
\value{
\code{collapse} always returns a character vector.
}
\description{
\code{collapse} is a fast version of 
\code{apply(x, along_dim, paste, collapse = "")}. It also works on vectors.
}
