% Generated by roxygen2 (4.0.2): do not edit by hand
\name{colorize}
\alias{colorize}
\title{Make image plots more colorful (only for ggplot objects)}
\usage{
colorize(obj, low = scales::muted("blue"), mid = "white",
  high = scales::muted("red"), ...)
}
\arguments{
\item{obj}{a ggplot object created by \code{\link{imageValues}}}

\item{low,mid,high}{colors to produce a low > mid > high color scale}
}
\value{
a ggplot object
}
\description{
\code{colorize} changes the color scale of \code{\link{imageValues}} plots
}

