% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{compF}
\alias{compF}
\title{Compute F-values}
\usage{
compF(obj, new_indices = NULL, attribs = FALSE, verbose = FALSE)
}
\arguments{
\item{obj}{a list object as returned by \code{\link{preAnova}}}

\item{new_indices}{if not NULL (default), it must be a vector of numeric
indices to be passed to the model matrix. This parameter is used for
permutation-based analyses.}

\item{attribs}{a logical value if dimension and verbose attributes should
be attached to the resulting array (default: FALSE)}

\item{verbose}{a logical value if p-values, effect sizes, and degrees of
freedom should be added to the returned array (as 'p_value', 'effect_size',
and 'Df' attributes)}
}
\value{
This function returns a numeric matrix of F-values if 'attribs' is
FALSE, and an array of F-values if 'attribs' is TRUE. If the verbose variant
was requested, the array has the following extra attributes: 'p_value' and
'effect_size', having the same dimension as the array of F-values, and
an integer matrix named 'Df' with term- and residual degrees of freedom for
each model term.
}
\description{
\code{compF} computes F-values. This is an internal function called by
the high-level function \code{\link{arrayAnova}}.
}
\keyword{internal}

