% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erp_preprocess.R
\name{compGfp}
\alias{compGfp}
\title{Compute Global Field Power}
\usage{
compGfp(dat, keep_channels = FALSE, channel_dim = "chan")
}
\arguments{
\item{dat}{numeric matrix or array, usually with named dimensions (one of 
which is "chan")}

\item{keep_channels}{logical value; if TRUE, the original channels are 
retained, and the GFP values are added with channel code "GFP"
(default = FALSE)}

\item{channel_dim}{a character value or numeric index indicating the
channel dimension of \code{dat} (default: "chan")}
}
\value{
The function returns a matrix or an array. Note that if
the original channels are not retained, the channel dimension is dropped.
}
\description{
\code{compGfp} computes Global Field Power (the standard deviation of 
channel values for each sampling point)
}

