% Generated by roxygen2 (4.0.2): do not edit by hand
\name{complexplot2dview}
\alias{complexplot2dview}
\title{Plot topography in complex layout}
\usage{
complexplot2dview(dat, ch_pos, timepoint, datgrid = NULL,
  layout_matrix = NULL, heights = NULL, widths = NULL, title_row = NULL,
  title_col = NULL, title_main = paste("Time:", timepoint, "ms"),
  plot_bar = TRUE, ampl_range = NULL, resolcol = 1000, gfp = NULL, ...)
}
\arguments{
\item{dat}{a list of numeric vectors or matrices, which can be plotted by
\code{\link{plot2dview}}}

\item{ch_pos}{channel position matrix}

\item{datgrid}{arrangement of the grid}
}
\description{
\code{complexplot2dview} calls plot2dview for the given groups/conditions
}
\seealso{
\code{\link{plot2dview}}
}

