% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anova.R
\name{compPvalueTanova}
\alias{compPvalueTanova}
\title{Compute p-values and length-corrected p-values in TANOVA}
\usage{
compPvalueTanova(effect_perm, pcrit)
}
\arguments{
\item{effect_perm}{an array of permuted effects (including the observed
effect as the first level of the 'perm' dimension)}

\item{pcrit}{a vector of p-value limits for consecutive length correction}
}
\value{
a list of two arrays: the p-values and the corrected p-values
}
\description{
\code{compPvalueTanova} is a helper function to compute p-values and
length-corrected p-values in the high-level \code{\link{tanova}} function.
}
\keyword{internal}

