% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anova.R
\name{compSumSq}
\alias{compSumSq}
\title{Compute sum of squares}
\usage{
compSumSq(obj, new_indices = NULL)
}
\arguments{
\item{obj}{a list object as returned by \code{\link{preAnova}}}

\item{new_indices}{if not NULL (default), it must be a vector of numeric
indices to be passed to the model matrix. This parameter is used for
permutation-based analyses.}
}
\value{
a numeric matrix of the sum-of-squares with an extra attribute for
the degrees of freedom ('Df'), or a list of such matrices
}
\description{
\code{compSumSq} computes sum of squares which are needed by
\code{\link{compF}} and \code{\link{permPvalues}}.
}
\keyword{internal}

