% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anova.R
\name{compTanovaEffect}
\alias{compTanovaEffect}
\title{Compute generalized GFP-effects}
\usage{
compTanovaEffect(obj, new_indices = NULL, attribs = FALSE,
  extra_attribs = character())
}
\arguments{
\item{obj}{a list object as returned by \code{\link{preAnova}}}

\item{new_indices}{if not NULL (default), it must be a vector of numeric
indices to be passed to the model matrix. This parameter is used for
permutation-based analyses.}

\item{attribs}{a logical value if dimension attributes should be attached to
the resulting array (default: FALSE)}

\item{extra_attribs}{a character vector of extra attributes which should be
attached to the resulting array (currently only "residuals" is supported).
The attribute names can be abbreviated.}
}
\value{
This function returns a numeric matrix of effects if 'attribs' is
FALSE, and an array of effects if 'attribs' is TRUE.
}
\description{
\code{compTanovaEffect} computes the generalized effect size measure as
given in Koenig & Melie-Garcia, 2009, p177. This is an internal function
called by the high-level function \code{\link{tanova}}.
}
\keyword{internal}

