% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{compZTanova}
\alias{compZTanova}
\title{Compute effect size in TANOVA}
\usage{
compZTanova(effect_perm, label = "")
}
\arguments{
\item{effect_perm}{an array of permuted effects (including the observed
effect as the first level of the 'perm' dimension)}

\item{label}{the label of the observed test statistic}
}
\description{
\code{compZTanova} is a helper function to compute corrected test statistic
in the high-level \code{\link{tanova}} function. It is actually the z-scored 
generalized dissimilarity statistic.
}
\keyword{internal}
