% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matstat_utils.R
\name{consectrue}
\alias{consectrue}
\title{Maximum number of consecutive TRUE values in each column/row of a matrix}
\usage{
consectrue(x, col = TRUE, any_NA = NULL)
}
\arguments{
\item{x}{logical vector, matrix or data.frame}

\item{col}{logical value if the search shall be performed column-wise (TRUE,
the default). Set to FALSE for row-wise counting.}
}
\description{
\code{consectrue} is helper function of \code{\link{tanova}}. It 
computes the longest streak of TRUE values in each column (or row) of a 
two-dimensional input.
}
\keyword{internal}
