% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chan_pos.R
\name{coordinates}
\alias{cart2geo}
\alias{cart2sph}
\alias{coordinates}
\alias{geo2cart}
\alias{geo2sph}
\alias{sph2cart}
\alias{sph2geo}
\title{Transformation between spherical, cartesian and geographical coordinates}
\usage{
sph2cart(ch_pos, r = 1, deg = TRUE)

cart2sph(ch_pos, deg = TRUE)

sph2geo(ch_pos, r = 1, deg = TRUE, long360 = TRUE,
  orient = c("northpole", "equatorial"))

geo2sph(ch_pos, r = 1, deg = TRUE)

geo2cart(ch_pos, r = 1, deg = TRUE)

cart2geo(ch_pos, deg = TRUE)
}
\arguments{
\item{ch_pos}{a data frame or matrix containing the spherical, cartesian,
or geographical coordinates of the electrode positions. It should contain
at least the following (named) columns, respectively: theta and phi;
x, y, and z; or lat and long.}

\item{r}{radius (default = 1)}

\item{deg}{logical variable indicating whether spherical or geographical
coordinates are or should be given in degrees (TRUE, default)}

\item{long360}{logical variable; if TRUE (default), longitudes range from
0 to 360 (otherwise from -180 to 180)}

\item{orient}{a character value specifying the orientation in the
geographical coordinate system; "northpole" (default) or "equatorial"}
}
\value{
A data.frame with converted coordinates
}
\description{
\code{sph2cart}, \code{sph2geo}, \code{cart2sph}, \code{cart2geo},
\code{geo2sph}, \code{geo2cart} transform spherical (sph), cartesian (cart),
or geographical (geo) coordinates into the respective coordinate system.
}

