% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cosAngle}
\alias{cosAngle}
\title{Cosine of the angles between electrodes}
\usage{
cosAngle(x, y = NULL, coords = TRUE, units_in_rows = TRUE,
  check_params = TRUE)
}
\arguments{
\item{x}{a matrix or data.frame with at least two columns and rows}

\item{y}{If not NULL (default), a matrix or data.frame with at least two
columns and rows}

\item{coords}{a logical variable; if TRUE (default), data in x are
coordinates. Coordinates should be named as "x", "y", "z" or "theta", "phi",
and if names are present, only coordinate columns are used in the
computations.}

\item{units_in_rows}{a logical variable indicating whether the units (vectors)
are in the rows of x (TRUE, default) or in the columns.}

\item{check_params}{logical; if TRUE (default), the appropriateness of
input data are checked before computing the cosine of angles. Set it to
FALSE if you really know what you are doing.}
}
\value{
A symmetric matrix
}
\description{
\code{cosAngle} computes the cosine of angles between electrodes or
between any N-dimensional vectors, or between two sets of electrodes or
any N-dimensional vectors.
}

