% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cprep.R
\name{cprep}
\alias{cprep}
\title{cprep}
\usage{
cprep(
  x,
  fs = 1,
  dtsens = setdiff(colnames(x), "t"),
  detrend = c("highpass", "linear", "mean", "constant", "none"),
  hpcf = 1,
  hptbw = 1,
  hpdev = 0.01,
  lnsens = dtsens,
  lfreq = 50,
  fbw = 2,
  w = 4,
  overlap = 0,
  tbw = 2,
  k = round(tbw * w - 1),
  p = 0.01,
  tau = 100,
  MaxIterLn = 10,
  rrsens = lnsens,
  oldref = NULL,
  newref = "robust",
  saveref = FALSE,
  interp = c("post-reference", "pre-reference", "none"),
  estmean = c("median", "huber", "mean"),
  sl = getLocationsfromLabels(colnames(x[, rrsens])),
  MaxIterRR = 4,
  report = paste(getwd(), "cprep_report.pdf", sep = "/")
)
}
\arguments{
\item{x}{input time series, specified as a numeric matrix or vector. In case
of a vector it represents a single signal; in case of a matrix each column
is a signal. Alternatively, an object of class \code{\link{ctd}}}

\item{fs}{sampling frequency of \code{x} in Hz. Default: 1. Overruled if
\code{x} is a \code{ctd} object, in which case the sampling frequency is
\code{fs(x)}}

\item{dtsens}{sensors to detrend, specified as positive integers indicating
sensors numbers, or as sensor names (colnames of \code{x}). Default: all
sensors names in \code{x} excluding a variable \code{t} (if present).}

\item{detrend}{character string indicating which type of trend removal
is performed for line noise removal bad channel detection; one of: 
\describe{
  \item{\code{highpass}}{use a high-pass FIR filter with cutoff frequency
   specified by the parameter \code{hpcf} (default)}
  \item{\code{linear}}{remove polynomial linear trend}
  \item{\code{mean|constant}}{remove mean}
  \item{\code{none}}{no detrending}
}}

\item{hpcf}{high-pass cutoff frequency in Hz, specified as a positive numeric
value. Default: 1 Hz}

\item{hptbw}{high-pass transition bandwith in Hz, specified as a positive
numeric value. For instance, if the high-pass cutoff frequency is 1 Hz, and
the transition bandwith is also 1 Hz, the transition band of 1 Hz is
located between 0.5 and 1.5 Hz. Default: 1 Hz.}

\item{hpdev}{deviation from desired stop- and passband of the high-pass filter,
specified as a positive numeric value. Default: 0.01}

\item{lnsens}{sensors to remove line noise from, specified as positive
integers indicating sensors numbers, or as sensor names (colnames of
\code{x}). Default: idetical to \code{dtsens}.}

\item{lfreq}{line frequencies to remove, either specified as a single numeric
value, or as a character string. If a single numeric value is specified,
then this value will be expanded to include multiples of that frequency
(harmonics) up to the Nyquist frequency. If a character string is
specified, then the string will be converted to numeric values without
further expansion. Default: 50}

\item{fbw}{frequency bandwidth. Bandwidth centered on each \code{lfreq} to
scan for significant lines. Default: 2 Hz.}

\item{w}{length of sliding window (segments) in seconds. Default: 4}

\item{overlap}{proportion of overlap between the segments. Default: 0}

\item{tbw}{taper bandwidth, specified as a positive numeric value. Default: 2
Hz}

\item{k}{number of tapers to use, specified as a positive numeric value.
Default: \code{tbw * w - 1}}

\item{p}{significance level cutoff for F-test. Default 0.01}

\item{tau}{Window overlap smoothing factor. Default: 100}

\item{MaxIterLn}{Maximum times to iterate line noise removal. Default: 10}

\item{rrsens}{vector of sensor names or numbers that are used for the
referencing (hence they must contain EEG data). Default: lnsens.}

\item{oldref}{vector of sensor names or numbers that formed the original
reference. If specified, the old reference signal is reconstructed back
into the data object (added if it was not present in the data). Ignored if
\code{x} is an object of class \code{ctd} and contains a \code{"ref"}
attribute. Default: NULL.}

\item{newref}{signals that form the new reference, either specified as a
character vector of signal names (\code{colnames(ctd_obj)}), or as a vector
of numerical values. If more than one signal is specified, then the average
of these signals will form the new reference. Alternatively, specifying
\code{"robust"} (default) computes an average reference with iterative
detection and interpolation of bad channels, and \code{"average"} removes
the average of the reference channels with no interpolation.}

\item{saveref}{logical indication whether to save to new reference as a}

\item{interp}{For \code{newref = "robust"}, specifies whether to perform
final channel interpolation. If \code{"post-reference"} (Dafault), after a
final robust reference is computed, the channels are re-interpolated and
the reference is corrected. In \code{"pre-reference"}, reref incrementally
adds to the bad channel list and interpolates before computing the
reference. If the initial estimate of the reference is poor, this is not a
good approach. In the \code{"none"} option, reref removes the reference but
does not interpolate. Bad channels remain in the signal. You may choose to
remove them during post-processing.}

\item{estmean}{For \code{newref = "robust"}, statistic used for estimation of
the mean; one of \code{"median"} (default), \code{"huber"} or
\code{"mean"}. Median and mean are the obvious statistics; \code{"huber"}
uses the Huber M-estimator of location with median absolute deviation (MAD)
scale, which is computationally expensive.}

\item{sl}{For \code{newref = "robust"}, sensor locations of \code{x},
specified as a data frame according to \code{\link{sensorlocs}} format.
Alternatively, a matrix with named columns x, y, and z representing the
sensor locations. Default: obtained from the sensor names given by the
column names of \code{x}. Only used in case \code{"when"} is either
\code{"before"} or \code{"after"}.}

\item{MaxIterRR}{maximum number of iterations used for detecting noisy
channels and recomputing the reference. Default: 4.  Only used in case
\code{"newref"} is \code{"robust"}.}

\item{report}{name of PDF file to which a detailed processing report is
printed, or NULL to suppress the report. Default: "cprep_report.pdf", saved
to the current working directory}
}
\value{
A list consisting of the following elements:
  \describe{
    \item{y}{a \code{\link{ctd}} object containing the robust referenced data}
    \item{noisyPre}{list containing the output from running
      \code{\link{noisysensors}}} on the input data
    \item{noisyProst}{list containing the output from running
      \code{\link{noisysensors}}} on the output data
}
}
\description{
PREP pipeline for continuous data
}
\examples{
\dontrun{
  rr <- cprep(EEGdata[, 1:29], fs = fs(EEGdata))
}
  
}
\seealso{
\code{\link[gsignal]{detrend}}, \code{\link{removetrend}}
  \code{\link{cleanline}}, \code{link{noisysensors}},
  \code{\link{reref}}
}
\author{
Original Matlab code by Tim Mullen (2011); ported to R and
  adapted by Geert van Boxtel, \email{G.J.M.vanBoxtel@gmail.com},
}
