% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{ctd}
\alias{ctd}
\alias{ctd.default}
\alias{npts}
\alias{npts<-}
\alias{ns}
\alias{ns<-}
\alias{fs}
\alias{fs<-}
\alias{print.ctd}
\alias{summary.ctd}
\alias{print.summary.ctd}
\alias{plot.ctd}
\title{Continuous Time Domain (ctd) data}
\usage{
ctd(x, fs, add.t = TRUE, ...)

\method{ctd}{default}(x, fs, add.t = TRUE, ...)

npts(object)

npts(object) <- value

ns(object)

ns(object) <- value

fs(object)

fs(object) <- value

\method{print}{ctd}(x, ...)

\method{summary}{ctd}(object, ...)

\method{print}{summary.ctd}(x, ...)

\method{plot}{ctd}(x, sensors = 1:ns(x), xlim = c(0, 10),
  ylim = c(-50, 50), ...)
}
\arguments{
\item{x}{name of a matrix that contains the data, or an object that can be coerced to a matrix,
consisting of \code{npts} data points as rows, and \code{ns} signals as columns.}

\item{fs}{the sampling frequency in Hertz.}

\item{add.t}{flag indicating whether to add a time variable to the matrix}

\item{...}{other arguments}

\item{object}{Object, usually of class \code{ctd}}

\item{value}{the attribute \code{npts}, \code{ns}, or \code{fs} will be set to this value}

\item{sensors}{Numeric. Sensors to plot (default \code{1:ns})}

\item{xlim}{Numeric. Lower and upper limits of plot X-axis (default: 0,10)}

\item{ylim}{Numeric. Lower and upper limits of plot Y-axis (default: -50,50)}
}
\value{
A \code{ctd} object, a numeric matrix with attributes \code{fs}, \code{ns}, and \code{npts}.
If the input matrix does not have column names, the columns will be named \code{S1 ... Snpts}. An additional
variable \code{t} containing time points will be generated if not present already and \code{add.t = TRUE}
}
\description{
Define the \code{ctd} object and associated methods
}
\details{
The \code{ctd} object can be used to store \strong{c}ontinuous \strong{t}ime \strong{d}omain data. 
 A \code{ctd} object is a numeric matrix with attributes\cr
 \code{fs} - the sampling frequency,\cr
 \code{ns} - the number of signals/channels, (columns of the matrix), and \cr
 \code{npts} - the number of data points/samples (rows of the matrix).\cr
 Methods exist to print, summarize, and plot the data.
}
\examples{
# simulate some rather silly data that could look like EEG (2 channels)
# assume 2 seconds of data sampled at 100 Hz
C3 <- arima.sim(list(order = c(1,1,0), ar = 0.7), n = 199)+(10*rnorm(200))
C4 <- arima.sim(list(order = c(1,1,0), ar = 0.7), n = 199)+(10*rnorm(200))
eeg <- ctd(cbind(C3,C4), 100)
print(eeg)
summary(eeg)

}
\author{
Geert van Boxtel, \email{G.J.M.vanBoxtel@gmail.com}
}
