% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrenddata.R
\name{detrenddata}
\alias{detrenddata}
\title{Detrend Data}
\usage{
detrenddata(ctd_obj, p = 1)
}
\arguments{
\item{ctd_obj}{object of class \code{'ctd'} containing continuous time domain
data.}

\item{p}{Order of the polynomial. Default: 1. The order of the polynomial can
also be given as a string, in which case p must be either "constant"
(corresponds to p = 0) or "linear" (corresponds to p = 1).}
}
\value{
A \code{ctd} object containing the detrended data.
}
\description{
Remove polynomial trend in a \code{ctd} object.
}
\examples{

data(EEGdata)
det <- detrenddata(EEGdata, 'constant')
plot(det)

}
\author{
Geert van Boxtel, \email{G.J.M.vanBoxtel@gmail.com}
}
