% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dim2multidim}
\alias{dim2multidim}
\title{Transform a specific dimension of an array into a multidimensional array}
\usage{
dim2multidim(dat, whichdim, datfr)
}
\arguments{
\item{dat}{is the input matrix or array}

\item{whichdim}{is the target dimension to be expanded (can be a numerical
index, a character string specifying the dimension name or a logical vector).
If whichdim is a string, dat must have named dimnames.}

\item{datfr}{a data.frame or matrix with the combination of factor levels
(e.g., the result of a call to expand.grid)}
}
\value{
An array
}
\description{
\code{dim2multidim} transforms a user-specified dimension of an array into a
multidimensional array while keeping other dimensions intact. It can be
conceived of as the inverse of expand.grid on the specified dimension.
}
\seealso{
\code{\link{expand.grid}}
}

