% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtspec.R
\name{dum}
\alias{dum}
\title{Detrend Data}
\usage{
dum()
}
\description{
Remove polynomial trend in a \code{ctd} object.
}
\examples{

data(EEGdata)

}
\author{
Geert van Boxtel, \email{G.J.M.vanBoxtel@gmail.com}
}
