% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EEGdata.R
\docType{data}
\name{EEGdata}
\alias{EEGdata}
\title{EEGdata}
\format{
A \code{ctd} object with 30 seconds of data sampled at 200 Hz
  containing the following signals, all measured in microvolts, except
  respiration:
\describe{
  \item{AF7 ... O2}{28 EEG channels, referenced to M1 (left mastoid)}
  \item{M2}{right mastoid,referenced to M1}
  \item{EOGh, EOGl, EOGr}{bipolar horizontal and vertical left and right EOG channels}
  \item{Resp}{Respiration measured by a chest belt (arbitrary units)}
  \item{ECG}{Electrocardiogram}
  \item{AgL, AgR}{Agonist and Antagonist EMG activity (finger flexion and extension)}
}
}
\usage{
EEGdata
}
\description{
Sample EEG, EOG, EMG, ECG and respiration data.
}
\examples{
data(EEGdata)
plot(EEGdata)

}
\author{
Geert van Boxtel, \email{G.J.M.vanBoxtel@gmail.com}
}
\keyword{datasets}
