% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeginterp.R
\name{eeginterp}
\alias{eeginterp}
\title{eeginterp}
\usage{
eeginterp(
  x,
  sl = getLocationsfromLabels(colnames(x)),
  bad,
  method = "spherical"
)
}
\arguments{
\item{x}{input time series, specified as a numeric matrix or vector. In case
of a vector it represents a single signal; in case of a matrix each column
is a signal. Alternatively, an object of class \code{\link{ctd}}}

\item{sl}{sensor locations of \code{x}, specified as a data frame according
to \code{\link{sensorlocs}} format. Alternatively, a matrix with named
columns x, y, and z representing the sensor locations. Default: obtained
from the sensor names given by the column names of \code{x}.}

\item{bad}{sensor locations of the bad sensors (those to interpolate),
specified in one of the following forms:
\itemize{
  \item a data frame in \code{\link{sensorlocs}} format
  \item a matrix with named columns x, y, and z representing the sensor
  locations
  \item a character vector containing sensor names
  \item a numeric vector representing columns of x
}}

\item{method}{method used for interpolation, specified as a character string.
Currently, only "spherical" is supported.}
}
\value{
interpolated bad sensors, returned as an object of the same class and
  dimensions as the input
}
\description{
Spherical spline interpolation of EEG data
}
\details{
The data at bad sensors is replaced by interpolated data from remaining
sensors using spherical spline interpolation (Perrin et al., 1989).
}
\examples{
\dontrun{
x <- EEGdata[, 1:28]
y <- eeginterp(x, bad = "T7, T8")
plot(ctd(x, fs = fs(EEGdata)))
plot(ctd(y, fs = fs(EEGdata)))
}


}
\references{
Perrin, F., Pernier, J., Bertrand, O., & Echallier, J. F. (1989).
  Spherical splines for scalp potential and current density mapping.
  Electroencephalography and clinical neurophysiology, 72(2), 184-187.
}
\author{
Matlab code for EEGlab, Copyright (C) Arnaud Delorme, CERCO, 2006,
  \email{arno@salk.edu}. Ported to R and adapted by Geert van Boxtel,
  \email{G.J.M.vanBoxtel@gmail.com}
}
