\name{eegR-package}
\alias{eegR-package}
\alias{eegR}
\docType{package}
\title{
A package to analyze EEG signals
}
\description{
The eegR package has been developed to process electroencephalography (EEG) signals. Beyond common EEG signal processing functionalities, eegR provides advanced tools to analyze single-trial and averaged event related potentials (ERPs).
}
\details{
\tabular{ll}{
Package: \tab eegR\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2014-08-25\cr
License: \tab GPL-3\cr
}
}
\author{
Denes Toth <toth.denes@ttk.mta.hu>
}

