% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eegr.R
\docType{package}
\name{eegr}
\alias{eegr}
\title{eegr: A package for processing eeg data.}
\description{
The \code{eegr} package provides functions to process continuous or event-related electroencephalographic (EEG) or other
psychophysiological data.
}
\details{
The most common file format for the storage of psychophysiological data is EDF/BDF, and this package contains the function
\code{readEDF} to read data from files using this format.
}
\section{to be added}{
NA
}

\section{another section}{
NA
}

\author{
Geert an Boxtel \email{G.J.M.vanBoxtel@gmail.com}
}
