% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extract.arrayAnova}
\alias{extract.arrayAnova}
\title{Extract slots from an \code{arrayAnova} object}
\usage{
\method{extract}{arrayAnova}(obj, what = c("stat", "p", "es", "stat_corr",
  "p_corr", "means", "df"), time_window = NULL, term = NULL, chan = NULL,
  drop = FALSE)
}
\arguments{
\item{obj}{an object of class \code{arrayAnova}}

\item{what}{a character vector indicating the slots to extract}

\item{time_window}{numeric vector of length two indicating the time window
which should be extracted. If NULL (default), all time points are returned.}

\item{term}{character vector indicating the name of the model terms which
should be extracted. If NULL (default), all model terms are returned.}

\item{chan}{character or integer indices of channels to extract. If NULL
(default), all channels are returned.}

\item{drop}{logical value; if TRUE, singleton dimensions are dropped from
the arrays after subsetting on time, term, and/or chan (default: FALSE)}
}
\description{
Extractor for \code{\link{arrayAnova}} objects
}

