% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary.R
\name{extract.default}
\alias{extract.default}
\title{Default extractor}
\usage{
\method{extract}{default}(
  object,
  what,
  time_window = NULL,
  term = NULL,
  chan = NULL,
  drop = FALSE,
  ...
)
}
\description{
Default extractor which is behind \code{\link{extract.arrayTtest}},
\code{\link{arrayAnova}}, and \code{\link{tanova}}, and does nothing
(at the moment) for objects of other classes.
}
\keyword{internal}
