% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extract}
\alias{extract}
\title{Extract slots from arrayAnova, arrayTtest, or tanova object}
\usage{
extract(obj, ...)
}
\arguments{
\item{obj}{an object as returned by \code{\link{arrayTtest}},
\code{\link{arrayAnova}}, or \code{\link{tanova}}}

\item{...}{parameters to the correspondig extractor method
(e.g. \code{\link{extract.arrayAnova}})}
}
\value{
an array of the requested component or a named list of arrays if
multiple components were requested
}
\description{
\code{extract} extracts slots from arrayAnova, arrayTtest, or tanova objects,
each produced by the corresponding function.
}
\note{
Only \code{arrayAnova} is implemented at the moment.
}
\seealso{
See the examples of \code{\link{arrayAnova}}
}

