% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fastUnique}
\alias{fastUnique}
\title{Fast version of unique for matrices}
\usage{
fastUnique(x, units_in_rows = TRUE)
}
\arguments{
\item{x}{matrix}

\item{units_in_rows}{units are in rows (TRUE, default) or columns of x}
}
\value{
matrix without duplicated units
}
\description{
\code{fastUnique} finds the unique rows or columns of a matrix.
}

