% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{fill_}
\alias{fill_}
\title{Replace parts of an object without creating an interim copy}
\usage{
fill_(x, index, y, arg_check = TRUE)
}
\description{
Replace parts of an object without creating an interim copy
}
\keyword{internal}

