% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterdata.R
\name{filterdata}
\alias{filterdata}
\title{Filter Data}
\usage{
filterdata(ctd_obj, filt)
}
\arguments{
\item{ctd_obj}{object of class \code{'ctd'} containing continuous time domain
data.}

\item{filt}{filter object recognized by the \code{gsignal} package (of class
\code{Arma}, \code{Ma}, \code{Sos}, or \code{Zpg}, characterizing a FIR or
a IIR filter.)}
}
\value{
A \code{ctd} object containing the filtered data.
}
\description{
Filter data in a \code{ctd} object.
}
\details{
Filtering is done using the \code{filtfilt} function in the \code{gsignal}
package, which minimizes filter startup transients by pre- and postpending
reflected pieces of the input signal, which are tapered to zero. Forward and
reverse filtering the signals corrects for phase distortion (not perfect in
practice).
}
\examples{

data(EEGdata)
fs <- fs(EEGdata)
nyq <- fs / 2

## 10 Hz low-pass (Butterworth; maximally flat)
but <- gsignal::butter(3, 10 / nyq)
lpd <- filterdata(EEGdata, but)
plot(lpd)

## 8-12 Hz low-pass filter 
## note that selecting a subset of signal
filt <- gsignal::fir1(50, c(8, 12) / nyq, "pass")
alpha <- filterdata(selectdata(EEGdata, signals = 1:28), filt)
plot(alpha, ylim=c(-10,10))

}
\author{
Geert van Boxtel, \email{G.J.M.vanBoxtel@gmail.com}
}
