% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matstat.R
\name{findExtremaMatrix}
\alias{findExtremaMatrix}
\title{Find local extrema in column vectors}
\usage{
findExtremaMatrix(x, n, tail, constant, topN, has_NA)
}
\arguments{
\item{x}{a numeric matrix}

\item{n}{the number of neighbouring points (default: 1L) to the left and to 
the right of each data point; a data point is a local minimum/maximum if it 
is below/above all data points in its neighbourhood. See also Details.}

\item{tail}{character string defining how tails should be handled (might be
abbreviated): "if_plateau" (the default) means extrema at the tails of the 
vector are only valid if they are part of a plateau; "never" means no 
extrema at the tails; "do_not_care" means no special treatment of extrema 
at the tails}

\item{constant}{the value (an integer value between 0 and 3 or NA) which 
should be assigned to the elements of constant vectors (default: 3L)}

\item{topN}{a positive integer scalar; if not NULL (the default), the local 
extrema must be among the top N highest/lowest extrema to be marked as such. 
If there are ties, all of them are returned.}

\item{has_NA}{if FALSE, x is not checked for missing values, thereby speeding
up the computations; if has_NA is NULL (default), a fast check is performed 
and if x has missing values, special corrections are applied (see Details).}
}
\description{
\code{findExtremaMatrix} looks for local minima and maxima in the
columns of a matrix. This is an internal function which supports the
exported function \code{\link{findExtrema}}.
}
\seealso{
\code{\link{findExtrema}}
}
\keyword{internal}

