% Generated by roxygen2 (4.0.1): do not edit by hand
\name{findExtremes}
\alias{findExtremes}
\title{Find local peaks (maxima) and valleys (minima) of a vector}
\usage{
findExtremes(x)
}
\arguments{
\item{x}{a vector}
}
\value{
A vector of the same length as x, with values 1 (local maximum),
-1 (local minimum) and 0.
}
\description{
\code{findExtremes} identifies local peaks and valleys of a vector.
Endpoints are considered peaks/valleys if they are part of a plateau.
}

