% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frd.R
\name{frd}
\alias{frd}
\alias{frd.default}
\alias{npts.frd}
\alias{ns.frd}
\alias{fs.frd}
\alias{fu}
\alias{fu.frd}
\alias{type}
\alias{type.frd}
\alias{yscale}
\alias{yscale.frd}
\alias{print.frd}
\alias{plot.frd}
\title{Frequency Domain Data}
\usage{
frd(x, ...)

\method{frd}{default}(
  x,
  f,
  fu = c("Hz", "rad/s", "normalized"),
  fs = 1,
  type = c("spectrum", "cross-spectrum", "phase", "coherence"),
  yscale = c("linear", "log", "dB"),
  ...
)

\method{npts}{frd}(object)

\method{ns}{frd}(object)

\method{fs}{frd}(object)

fu(object)

\method{fu}{frd}(object)

type(object)

\method{type}{frd}(object)

yscale(object)

\method{yscale}{frd}(object)

\method{print}{frd}(x, ...)

\method{plot}{frd}(
  x,
  sensors = setdiff(colnames(x), "f"),
  yscale = c("linear", "log", "dB"),
  xlim = c(0, fs(x)/2),
  ylim = NULL,
  ...
)
}
\arguments{
\item{x}{object}

\item{...}{additional arguments passed to functions}

\item{f}{vector of frequencies}

\item{fu}{frequency units, \code{Hz} (default), \code{rad/s}, or
\code{normalized}}

\item{fs}{sampling frequency (default: 1)}

\item{type}{type of data; one of \code{'spectrum'}, \code{'cross-spectrum'},
\code{'phase'}, \code{'coherence'}}

\item{yscale}{scale of the data, one of \code{'linear'}, \code{'log'},
\code{'dB'}}

\item{object}{object}

\item{sensors}{Numeric. Sensors to plot (default \code{1:ns})}

\item{xlim}{Numeric. Lower and upper limits of plot X-axis (default: 0,10)}

\item{ylim}{Numeric. Lower and upper limits of plot Y-axis (default: -50,50)}
}
\value{
A \code{frd} object, a numeric matrix with attributes \code{fs},
  \code{ns}, \code{npts}, \code{type}, and \code{yscale}. If the input matrix
  does not have column names, the columns will be named \code{S1 ... Snpts}.
}
\description{
Define the \code{frd} object and associated methods.
}
\details{
The \code{frd} object can be used to store \strong{fr}equency \strong{d}omain
data.
 A \code{frd} object is a numeric matrix with at least a column named \code{f} containing
 the frequencies, one or more data columns (usually named), and the following attributes:
 \describe{
   \item{fs}{sampling frequency}
   \item{fu}{frequency units}
   \item{ns}{number of signals, (columns of the matrix)}
   \item{npts}{the number of data points/samples (rows of the matrix)}
   \item{type}{type of data: spectrum, cross-spectrum, phase, coherence}
   \item{yscale}{Y-scale of the data: linear, log, dB}
}
Methods exist to print, summarize, and plot the data.
}
\examples{
# coming up
f <- 1

# coming up

}
\author{
Geert van Boxtel, \email{G.J.M.vanBoxtel@gmail.com}
}
