% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparrow_utils.R
\name{get_sparrow_colours}
\alias{get_sparrow_colours}
\title{Get conventional sparrow species colours}
\usage{
get_sparrow_colours(species = c("PDOM"), accents = c(0))
}
\arguments{
\item{species}{Character vector of species codes ("PDOM", etc.)}

\item{accents}{Float vector of accent values between -1 and 1. 0 is default.}
}
\value{
A vector of hexadecimal colour codes.
}
\description{
This function returns appropriate colour values
for plotting Passer sparrow data. You pass in a vector of species
codes ("PDOM", "PMON", "PHIS", and "PITA" are currently supported),
and the function returns a vector of hexadecimals. You can
include each species code an arbitrary amount of times to get colours
for many individuals; the length of the output will equal that of your
input vector. You must also provide input "accents": a vector of numbers
between -1 and 1. This lightens (negative numbers) or darkens (positive
numbers) the colour to accent it. This will be applied to each individual
colour, not for all colours sharing the same species code.
}
