% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parallel.R
\name{getDoBackend}
\alias{getDoBackend}
\title{\code{getDoBackend} returns the internal data of the currently registered
foreach \%dopar\% backend.}
\usage{
getDoBackend()
}
\description{
\code{getDoBackend} returns the internal data of the currently registered
foreach \%dopar\% backend.
}
\author{
Renaud Gaujoux, Denes Toth
}
\keyword{internal}

