% Generated by roxygen2 (4.0.2): do not edit by hand
\name{imagePvalues}
\alias{imagePvalues}
\title{Image plot of p-values}
\usage{
imagePvalues(pvalues, pcrit = c(0.001, 0.01, 0.05), grid = NULL,
  wrap = NULL)
}
\arguments{
\item{pvalues}{numeric matrix or array of p-values with named dimensions (at
least chan and time)}

\item{pcrit}{numeric vector of significancy limits
(default: 0.001, 0.01, 0.05)}

\item{grid}{character vector or formula defining the layout of panels}

\item{wrap}{character vector or formula defining the dimension which
separates panels (only considered if grid is NULL)}
}
\value{
A ggplot object
}
\description{
\code{imagePvalues} creates an image plot from a matrix or array of p-values
}

