% Generated by roxygen2 (4.0.2): do not edit by hand
\name{imageValues}
\alias{imageValues}
\title{Image plot of channel values}
\usage{
imageValues(dat, grid = NULL, wrap = NULL, bar_title = "effect",
  time_label = "Time (ms)", channel_label = "Channels", raster = TRUE,
  cluster_order = FALSE)
}
\arguments{
\item{dat}{numeric matrix or array of values with named dimensions (at least
chan and time)}

\item{grid}{character vector or formula defining the layout of panels}

\item{wrap}{character vector or formula defining the dimension which
separates panels (only considered if grid is NULL)}

\item{bar_title}{character string; the title of the color bar}

\item{time_label}{character string; the label of the x (time) axis
(default: "Time (ms)")}

\item{channel_label}{character string; the label of the y (channel) axis
default: "Channels")}

\item{raster}{use raster image (TRUE, default) or not (FALSE)}

\item{cluster_order}{logical; if TRUE, channels are ordered with hierarchical
agglomerative clustering (default: FALSE)}
}
\value{
A ggplot object
}
\description{
\code{imageValues} creates an image plot from a matrix or array of channel x
time values
}
\seealso{
\code{\link{imagePvalues}} for plotting p-values,
\code{\link{colorize}} for setting color scale
}

