% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erp_preprocess.R
\name{importBVdat}
\alias{importBVdat}
\title{Import binary file exported from BrainVision}
\usage{
importBVdat(
  file_name,
  file_path = getwd(),
  id = "",
  import_options = importOptions()
)
}
\arguments{
\item{file_name}{character string; the name of the input files without
extensions}

\item{file_path}{character string; the path to the files if they are not in 
the working directory (default)}

\item{id}{character string denoting the identification code of the 
participant}

\item{import_options}{a list, which should be given by calling 
\code{\link{importOptions}}}
}
\value{
A list with three named elements: eeg (array), markers (data.frame), 
channels (data.frame)
}
\description{
\code{importBVdat} imports a binary file exported from the BrainVision 
software
}
\note{
This is a custom function tailored for the special datasets collected 
in our lab. Use it with extra care for general purposes!
}
