% Generated by roxygen2 (4.0.2): do not edit by hand
\name{importOptions}
\alias{importOptions}
\title{Import options for importBVdat}
\usage{
importOptions(eeg_ext = "dat", marker_ext = "vmrk", info_ext = "vhdr",
  marker_skip = 14, marker_segment = "Target",
  marker_badcat = "Bad Interval", marker_badstim = "resp_false",
  marker_keepstim = "*", marker_badchan = c("badchan_start",
  "badchan_stop"), marker_time0 = "Time 0", segment_dpoints = (-100):1023,
  marker_regexp = T, marker_ignorecase = T, marker_header = FALSE,
  marker_fill = TRUE, marker_asis = TRUE, marker_sep = ",")
}
\arguments{
\item{eeg_ext}{character value, the extension of the segmented EEG dataset
file (default: "dat")}

\item{marker_ext}{character value, the extension of the raw marker file
(default: "vmrk")}

\item{info_ext}{character value, the extension of the information file
(default: "vhdr")}

\item{marker_skip}{numeric value; number of rows to be skipped while
importing the marker file (default: 14)}

\item{marker_segment}{character vector; name(s) of stimuli defining the
the segments (default: "Target")}

\item{marker_badcat}{character string(s) identifying bad segments in the 1st
column of the marker file (default: "Bad Interval")}

\item{marker_badstim}{character string(s) identifying bad stimuli or
responses in the 2nd column of the marker file (default: "resp_false")}

\item{marker_badchan}{character vector of length 1 or 2; markers identifying
bad intervals of channels. If marker_badchan has only one element, "_start"
and "_stop" are automatically appended to the end of the string.}

\item{marker_time0}{character string identifying the marker name at time 0
in the imported EEG dataset}

\item{segment_dpoints}{numeric vector of data point indices defining a
segment (default: -100:1023)}

\item{marker_regexp}{logical value; should marker_segment/badcat/.../time0
strings be handled as regular expressions (default) or treated as they are}

\item{marker_ignorecase}{logical value; should the case of marker_segment/
.../time0 definitions be ignored (default)}

\item{keepstim}{character vector identifying which stimuli or responses
should be kept by checking the 2nd column of the marker file (default: "*")}

\item{{marker_header,marker_fill,marker_asis,marker_sep}}{parameters
to be passed to \code{\link{read.table}} while importing the marker file}
}
\value{
A list with named parameters
}
\description{
\code{importOptions} allows to set the raw data importation parameters
}

