% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{isCombineOperator}
\alias{isCombineOperator}
\title{Internal function which combines is* functions}
\usage{
isCombineOperator(lhs, rhs, op)
}
\arguments{
\item{lhs}{the function (of class "IsFunction") on the left-hand side}

\item{rhs}{the function (of class "IsFunction") on the right-hand side}

\item{op}{the operator (\code{`&`}, \code{`|`} or \code{xor})}
}
\description{
\code{isCombineOperator} is the workhorse function behind logical operators
for objects of class IsFunction.
}
\keyword{internal}
