% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{isExpandInto}
\alias{isExpandInto}
\title{Expansion in is* functions, with a fallback to repLen}
\usage{
isExpandInto(from, to, from_name)
}
\description{
\code{isExpandInto} is a wrapper around \code{expandInto} which is used
in the is* functions to transform the arguments like ref., tol., etc.
to match the input array.
}
\keyword{internal}

