% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{isMainCompFn}
\alias{isMainCompFn}
\title{Call the internal function in is* functions on subsets and expand}
\usage{
isMainCompFn(FUN., x, subset., expand., which_dims., options., negate., ...)
}
\arguments{
\item{FUN.}{the internal function whose first two arguments are the data (x)
and the options (options.). Additional arguments must be provided in dots
(...).}

\item{subset., expand., which.dims., options., negate.}{obligatory arguments, see
\code{is}}

\item{...}{arguments passed to FUN.}
}
\description{
\code{isMainCompFn} calls the internal function in the given \code{is*}
function on the input vector or on the appropriate slices of the input
array.
}
\keyword{internal}
