% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{isStrict}
\alias{isStrict}
\title{Check if an is* function is strict or optional}
\usage{
isStrict(...)
}
\arguments{
\item{...}{the functions to be tested. Can be a single list of functions as
well.}
}
\description{
\code{isStrict} is a helper function to check whether a function of class
\code{IsFunction} (see \code{\link[eegR]{is}}) is strict or optional. This
attribute is important if the tests are combined, see \code{\link{combine}}.
}
\seealso{
\code{\link[eegR]{is}}, \code{\link{combine}}
}

