% Generated by roxygen2 (4.0.2): do not edit by hand
\name{isTtest}
\alias{isTtest}
\title{Workhorse function of arrayTest (for independent samples t-test)}
\usage{
isTtest(dat, groups, id_dim = "id", mu = 0, var_equal = TRUE,
  verbose = FALSE, any_NA = NULL)
}
\description{
\code{indepTtest} computes t-values and additional p-values & degrees of
freedom if requested.
}

