% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ttest.R
\name{isTtestRandomGroups}
\alias{isTtestRandomGroups}
\title{Create random group memberships for independent samples t-tests}
\usage{
isTtestRandomGroups(obj, nperm, seed = NULL)
}
\arguments{
\item{obj}{a list object returned by \code{\link{preTtest}};
the \code{obj$groups} element is relevant indicating the original group
memberships}

\item{nperm}{integer indicating the number of permutations}

\item{seed}{an integer value which specifies a seed (default: NULL), or a
list of arguments passed to \code{\link{set.seed}}}
}
\value{
\code{isTtestRandomGroups} returns a matrix with as many rows as
the length of 'group' and as many columns as 'nperm'.
}
\description{
This function creates a random permutation of the original group
memberships for independent samples t-tests. Care is taken so that all
permutations are unique.
}
\examples{
\dontrun{
# this should fail: too many permutations on a small sample
temp <- try(isTtestRandomGroups(list(groups = rep(1:2, each = 4L), 1e3)),
            silent = TRUE)
stopifnot(inherits(temp, "try-error"))

# all permutations should be unique
temp <- isTtestRandomGroups(list(groups = rep(1:2, c(6, 4))), 500)
stopifnot(!anyDuplicated(temp, MARGIN = 2L))
temp <- isTtestRandomGroups(list(groups = rep(1:2, each = 20L)), 1e4)
stopifnot(!anyDuplicated(temp, MARGIN = 2L))
}
}
\keyword{internal}

