% Generated by roxygen2 (4.0.1): do not edit by hand
\name{listS}
\alias{listS}
\title{Create list with substituted names}
\usage{
listS(..., indices = NULL)
}
\arguments{
\item{...}{objects; if not named, listS is equilent to \code{\link{list}}.
Names which should be substituted should start with a dot (.) or INDICES has
to be provided.}

\item{indices}{character or numeric vector indicating the position of those
list elements whose name should be substituted. If provided, relevant names
in ... should not be dotted.}
}
\value{
A list with substituted names.
}
\description{
\code{listS} creates a named list where names are substituted with the
content of the referenced variable.
}

