% Generated by roxygen2 (4.0.1): do not edit by hand
\name{marginalMeans}
\alias{marginalMeans}
\title{Compute marginal means in an ANOVA design}
\usage{
marginalMeans(form, f_dat, a_dat, dimn, keep_term_order = FALSE,
  residualmean = FALSE, whichterm = NULL, no_icpt = FALSE)
}
\arguments{
\item{form}{formula of the model}

\item{f_dat}{data.frame of factors}

\item{a_dat}{matrix which contains the dependent variables. It must have as
many rows as f_dat.  Usually a_dat is the result of calling array2mat on the
orginal data array.}

\item{dimn}{the original dimension names for the array corresponding to the
column names of a_dat #### SHOULD BE SIMILAR TO MAT2ARRAY ###################}

\item{keep_term_order}{logival variable whether to keep the order of the
model terms as provided in form (TRUE) or all interaction terms should follow
all main effect terms (FALSE, default)}

\item{residualmean}{logical variable; if TRUE, each factor term is extracted
from the data after computing the given marginal mean (default: FALSE)}

\item{whichterm}{character, numeric or logical vector; indices of model terms
which should be analyzed (default: NULL, meaning all terms are included)}

\item{no_icpt}{logical value; if TRUE, global mean is subtracted before
computing the marginal means (default: FALSE)}
}
\value{
A named list containing the marginal means for each model terms
}
\description{
\code{marginalMeans} computes marginal means in ANOVA designs
}

