% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mat2array}
\alias{mat2array}
\title{(Back-)Transforms a matrix to an array. See array2mat as a related function.}
\usage{
mat2array(dat, dims = NULL, row_dim = NULL)
}
\arguments{
\item{dat}{a matrix to reshape}

\item{dims}{a numeric vector indicating the dimensions of the resulting array
or a (potentially named) list indicating the dimnames attribute of the
resulting array. If NULL (default), dat is assumed to be the result of a call
to array2mat().}

\item{row_dim}{name or index of dims which corresponds to the row dimension
of dat. If NULL (default), dat is assumed to be the result of a call to
array2mat().}
}
\description{
\code{mat2array} reshapes a matrix to an array. It is essentially the inverse
function of array2mat.
}
\seealso{
\code{\link{array2mat}}
}

