% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{matchArg}
\alias{matchArg}
\title{Argument verification if there might be an "all" option}
\usage{
matchArg(arg, choices = NULL, several_ok = TRUE)
}
\arguments{
\item{arg}{a character vector (of length one unless several_ok is TRUE) or 
NULL}

\item{choices}{a character vector of candidate values (if NULL [default], it 
is deduced from the formals of the calling function)}

\item{several_ok}{logical specifying whether arg should be allowed to have 
more than one element (default: TRUE)}
}
\value{
If 'arg' has an element 'all' (or 'ALL', 'All', etc.), all other 
choices are returned. Otherwise see \code{\link[base]{match.arg}}.
}
\description{
\code{matchArg} is a wrapper around \code{\link[base]{match.arg}} to check
arguments which have an "all" option in their choices. Note that the defaults
are not the same as in \code{\link[base]{match.arg}}.
}
\keyword{internal}
