% Generated by roxygen2 (4.0.2): do not edit by hand
\name{matrixIP}
\alias{matrixIP}
\title{Fast in-place transformation to a matrix (without copy)}
\usage{
matrixIP(x, nrow, ncol, dimnames = NULL, force_length = TRUE)
}
\arguments{
\item{x}{a data vector, matrix or array}

\item{nrow}{the desired number of rows}

\item{ncol}{the desired number of columns}

\item{dimnames}{A dimnames attribute for the matrix: NULL or a list of
length 2 giving the row and column names respectively. The list can be
named, and the list names will be used as names for the dimensions. An empty
list is treated as NULL, and a list of length one as row names.}

\item{force_length}{logical. If TRUE (the default), \code{matrixIP} checks if
length(x)==nrow*ncol. If not, x is recycled or subsetted to the desired
length.}
}
\value{
A matrix (invisibly)
}
\description{
\code{matrixIP} transforms its data argument to a matrix by reference. No
copy is made at all, and it invisibly returns the matrix.
}
\note{
Use \code{matrixIP} with extra care because it modifies in place all
objects which x refers to. If you want to avoid this, call
\code{x <- copy(x)} before calling \code{matrixIP} or use the standard way as
described in the Note section of \code{\link{matrix}}. However, for input
objects created on-the-fly (e.g. a temporary vector), \code{matrixIP} is safe
and more compact than the latter solution, and can be many times faster than
\code{matrix}.
}
\seealso{
\code{\link{arrayIP}} for in-place transformation of x to an array
(without copy) and \code{\link{matrix}} for creating a new matrix without
modifying the original input
}

