% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/array_transform.R
\name{mergeDims}
\alias{mergeDims}
\title{Reshape array by merging specific dimensions}
\usage{
mergeDims(dat, dims, return_attributes = TRUE, keep_dimnames = TRUE,
  sep = ".")
}
\arguments{
\item{dat}{array to reshape}

\item{dims}{list of dimension names or dimension indices to merge. Can be a vector if
only one group of dimensions should be merged. If any of the list elements is a
character vector, dat must have named dimnames.}

\item{return_attributes}{logical value (default: TRUE) whether original
attributes should be appended to the resulting array}

\item{keep_dimnames}{logical; if TRUE (default), dimension names are also
merged and attached as dimnames attribute}

\item{sep}{character which separates dimension names after merging (default ".")}
}
\value{
Array with merged dimensions
}
\description{
\code{mergeDims} reshapes an array by merging user-specified dimensions
}

