% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{modelplot}
\alias{modelplot}
\alias{modelplot.default}
\alias{modelplot.tanova}
\title{Plot results of t-test, ANOVA or TANOVA modeling functions}
\usage{
modelplot(...)

\method{modelplot}{default}(results, what = c("statistic", "p-value"),
  type = c("corrected", "uncorrected"), pcrit = c(0.001, 0.01, 0.05),
  grid = NULL, wrap = NULL, time_label = "Time (ms)",
  channel_label = "Channels", raster = TRUE, cluster_order = FALSE,
  low = scales::muted("blue"), mid = "white", high = scales::muted("red"),
  midpoint = 0, ...)

\method{modelplot}{tanova}(results, what = c("statistic", "p-value"),
  type = c("corrected", "uncorrected"), grid = NULL, wrap = NULL,
  time_label = "Time (ms)", ...)
}
\arguments{
\item{...}{arguments passed to \code{\link{extract}}; e.g. use
arguments 'time_window' and 'term' to display only a subset of the results}

\item{results}{a list; the return value of the corresponding modeling
functions}

\item{what}{a character string or vector indicating what should be plotted:
only the test statistics ("statistic"), only the p-values ("p-values"), or
both (default). The names can be abbreviated.}

\item{type}{display "corrected" (default) or "uncorrected" statistics or
p-values. The names can be abbreviated.}

\item{pcrit}{numeric vector of significancy limits
(default: 0.001, 0.01, 0.05) for highlighting significant areas}

\item{grid}{character vector or formula defining the layout of panels}

\item{wrap}{character vector or formula defining the dimension which
separates panels (only considered if grid is NULL)}

\item{time_label}{character string; the label of the x (time) axis
(default: "Time (ms)")}

\item{channel_label}{character string; the label of the y (channel) axis
default: "Channels")}

\item{raster}{use raster image (TRUE, default) or not (FALSE)}

\item{cluster_order}{logical; if TRUE, channels are ordered with hierarchical
agglomerative clustering (default: FALSE)}

\item{low}{colour for low/mid/high end of gradient, respectively}

\item{mid}{colour for low/mid/high end of gradient, respectively}

\item{high}{colour for low/mid/high end of gradient, respectively}

\item{midpoint}{the midpoint (in data value) of the diverging scale
(default: 0)}
}
\value{
A ggplot object or a list of such objects
}
\description{
\code{modelplot} is a generic function for plotting
\code{\link{arrayTtest}}, \code{\link{arrayAnova}}, or \code{\link{tanova}}
objects, each produced by the corresponding function.

\code{modelplot.default} plots the result of the \code{\link{arrayTtest}} or
\code{\link{arrayAnova}} function.

\code{modelplot.tanova} plots the result of the \code{\link{tanova}} function.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: 

\item \code{tanova}: 
}}
\seealso{
See the examples for \code{\link{arrayAnova}} and
\code{\link{tanova}}
}
\keyword{internal}

