% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtspec.R
\name{mtspec}
\alias{mtspec}
\alias{mtchan}
\alias{mtfft}
\title{Multitaper Spectrum}
\usage{
mtspec(
  x,
  fs = 1,
  detrend = c("short-linear", "short-mean", "long-linear", "long-mean", "none"),
  w = 4,
  overlap = 0,
  tbw = 2,
  k = round(tbw * w - 1)
)

mtchan(chan, fs, detrend, n, overlap, tapers)

mtfft(dat, fs, tapers)
}
\arguments{
\item{x}{input time series, specified as a numeric or complex vector. In case
of a vector it represents a single signal; in case of a matrix each column
is a signal. Alternatively, an object of class \code{\link{ctd}}}

\item{fs}{sampling frequency of \code{x} in Hz. Default: 1. Overruled if
\code{x} is a \code{ctd} object, in which case the sampling frequency is
\code{fs(x)}}

\item{detrend}{character string specifying detrending option; one of:
 \describe{
   \item{\code{long-mean}}{remove the mean from the data before
   splitting into segments (default)}
   \item{\code{short-mean}}{remove the mean value of each segment}
   \item{\code{long-linear}}{remove linear trend from the data before
   splitting into segments}
   \item{\code{short-linear}}{remove linear trend from each segment}
   \item{\code{none}}{no detrending}
}}

\item{w}{length of sliding window (segments) in seconds. Default: 4}

\item{overlap}{proportion of overlap between the segments. Default: 0}

\item{tbw}{taper bandwidth, specified as a positive numeric value. Default: 2
Hz}

\item{k}{number of tapers to use, specified as a positive numeric value.
Default: \code{tbw * w - 1}}

\item{chan}{single channel}

\item{n}{segment length (nearest power of 2 for \code{w})}

\item{tapers}{tapers from dpss}

\item{dat}{data}
}
\value{
An object of class \code{frd}, containing the spectra
}
\description{
Compute the multitaper spectrum of a vector or matrix.
}
\details{
The multitaper method is a method for spectral density estimation, which
overcomes some of the limitations of conventional Fourier analysis. It is a
periodogram-based method that uses multiple tapers, or windows, to form
independent estimates of the spectral density to reduce variance of the
spectral density estimate. The method uses Slepian or discrete prolate
spheroidal sequences as tapers since these vectors are mutually orthogonal
and possess desirable spectral concentration properties. They average out
noise in the spectrum and reduce information loss at the edges of the window.

The package \code{multitaper} contains a more complete version of multitaper
spectral analysis (function \code{\link[multitaper]{spec.mtm}}), which
computes confidence intervals, F-tests , and more.

The present implementation uses sliding windows, i.e., overlapping segments,
tapering each segment, and then averaging over tapered segments. The tapers
are computed by the function \code{\link[multitaper]{dpss}}, with parameters
\code{n = w * fs} (number of samples in sliding window), \code{k = tbw * w -
1} (number of tapers), and \code{nw = tbw * w / 2} (time bandwidth of the
tapers).

The function \code{mtspec} computes the multitaper spectrum for a vector or
matrix. It calls the function \code{mtchan}, which computes the multitaper
spectrum for a single channel (no parameter checking).
}
\examples{
data(EEGdata)
sp <- mtspec(EEGdata[, 1:28], fs = fs(EEGdata), detrend = "short-linear")
plot(sp, yscale = "dB", main = "EEGdata")

}
\references{
\url{https://en.wikipedia.org/wiki/Multitaper}
}
\seealso{
\link[multitaper]{spec.mtm}, \link[multitaper]{dpss},
  \code{\link{frd}}
}
\author{
Geert van Boxtel, \email{G.J.M.vanBoxtel@gmail.com}, based on Matlab
  code by Tim Mullen's (2011) adaptation of the Chronux 2 toolbox
  (\url{http://chronux.org/}) for the cleanline plugin in EEGlab.
}
