% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., plot_list = NULL, cols = 1L, layout = NULL)
}
\arguments{
\item{...}{ggplot objects}

\item{plot_list}{a list of ggplot objects}

\item{cols}{the number of columns in layout}

\item{a}{matrix specifying the layout. If present, 'cols' is ignored.}
}
\description{
\code{multiplot} plots multiple ggplot objects on one page
}
\author{
Winston Chang
}
\keyword{internal}

