% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is.R
\name{negate}
\alias{!.IsFunction}
\alias{negate}
\title{Negation for the is* family of functions}
\usage{
\method{!}{IsFunction}(fn)
}
\arguments{
\item{fn}{a function of class IsFunction}
}
\description{
\code{`!.IsFunction`} implements logical negation for \code{is*} function. It
modifies the call by negating its 'negate.' argument.
}
\keyword{internal}

