% Generated by roxygen2 (4.0.2): do not edit by hand
\name{osTtest}
\alias{osTtest}
\title{Workhorse function of arrayTest (for one-sample or paired-samples t-test)}
\usage{
osTtest(dat, id_dim = "id", mu = 0, verbose = FALSE, any_NA = NULL)
}
\description{
\code{indepTtest} computes t-values and additional p-values & degrees of
freedom if requested.
}

