% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{osTtestRandomSigns}
\alias{osTtestRandomSigns}
\title{Create random signs for one-sample and paired samples t-tests}
\usage{
osTtestRandomSigns(obj, nperm, seed)
}
\arguments{
\item{obj}{a list object returned by \code{\link{preTtest}}; the number of
rows (let's denote here 'id_length') of \code{obj$.arraydat} is relevant}

\item{nperm}{integer indicating the number of permutations}

\item{seed}{an integer value which specifies a seed (default: NULL), or a 
list of arguments passed to \code{\link{set.seed}}}
}
\value{
\code{osTtestRandomSigns} returns an integer matrix (containing 
only -1L and 1L values) with 'id_length' rows and 'nperm' columns.
}
\description{
This function creates a random sample of -1L and 1L values which can be
used in randomization runs of one-sample and paired samples t-tests. 
Care is taken so that all samples are unique.
}
\examples{
\dontrun{
# this should fail: too many permutations for this sample size
obj <- list(.arraydat = matrix(8))
temp <- try(osTtestRandomSigns(obj, 1e4), silent = TRUE)
stopifnot(inherits(temp, "try-error"))

# expect an integer matrix containing unique series of -1L and 1L
obj <- list(.arraydat = matrix(10))
temp <- osTtestRandomSigns(obj, 500)
stopifnot(range(temp), c(-1L, 1L))
stopifnot(length(unique(as.vector(temp))) == 2L)
stopifnot(!anyDuplicated(temp, MARGIN = 2L))

obj <- list(.arraydat = matrix(20))
temp <- osTtestRandomSigns(obj, 1e4)
stopifnot(!anyDuplicated(temp, MARGIN = 2L))
}
}

