% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_setters.R
\name{permParams}
\alias{permParams}
\title{Setting the parameters for parallel computation}
\usage{
permParams(n = 999L, type = c("residuals", "observations"))
}
\arguments{
\item{n}{integer number of randomization runs}

\item{type}{character string which defines if model residuals ("residuals",
the default), or the raw observations ("observations") should be permuted}
}
\value{
\code{parallelParams} returns a list with two elements corresponding
to the arguments ('n' and 'type')
}
\description{
\code{permParams} sets the parameters for permutations (randomizations)
used for example in \code{arrayAnova} or \code{tanova}.
}

