% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anova.R
\name{permPvalues}
\alias{permPvalues}
\title{Permutation of the test statistic}
\usage{
permPvalues(obj, nperm, seed, stat_obs, tfce = NULL)
}
\arguments{
\item{obj}{a list returned by \code{arrayAnova} or \code{arrayTtest}}

\item{nperm}{an integer value indicating the number of permutations}

\item{seed}{an integer value which specifies a seed (default: NULL), or a
list of arguments passed to \code{\link{set.seed}}}

\item{stat_obs}{a numeric array of the observed test statistic}

\item{tfce}{NULL (default) or a list of the TFCE parameters}
}
\value{
\code{\link{permPvalues}} returns a logical array whose elements
indicate whether the absolute observed test statistic in the given chanXtime
cell is smaller than the absolute randomized test statistic
}
\description{
\code{permPvalues} is called internally by \code{arrayAnova} and
\code{arrayTtest} and performs permutation analysis.
}
\keyword{internal}

