% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot2dview}
\alias{plot2dview}
\title{Plot the topography of the signals}
\usage{
plot2dview(dat, ch_pos, r = 1, timepoint = NULL, ampl_range = NULL,
  type = c("contour", "raster"), resol = 30L, resolcol = 101L,
  projection = "laea", projref = c("pole", "equator"), origo = c(lat =
  ifelse(projref == "pole", 90, 0), long = ifelse(projref == "pole", 270, 0)),
  plot_centroid = TRUE, centroid_circle = 0.5,
  centroid_unitsphere = FALSE, plot_bar = TRUE, plot_ch = TRUE,
  plot_chnames = TRUE, title = NULL, gfp = NULL, gfp_max = NULL,
  colour_scale = function(n) gplots::colorpanel(n, "blue", "white", "red"),
  colour_midpoint = 0, ...)
}
\arguments{
\item{dat}{a numeric vector or matrix. If a matrix is provided, it must 
contain the participants in rows and the channels in columns}

\item{ch_pos}{a data.frame of electrode positions, 
see \code{\link{coordinates}}}

\item{r}{a numeric value; the radius of the head (not used if ch_pos contains
a column \code{r})}

\item{timepoint}{an integer value representing the time point at which the 
amplitude values were recorded. This is used as the title of the topoplot if 
the 'title' argument is not provided, and for highlighting the time point on
the GFP curve if 'gfp' data are provided (with proper names or dimnames 
attribute).}

\item{ampl_range}{numeric vector, the minimum and maximum value of the 
amplitudes. If NULL (default), it is computed from the data. Data values
outside the \code{ampl_range} interval are winsorized (set to equal to the
minimum or maximum limit).}

\item{type}{the type of the topoplot, either "contour" (the default) or 
"raster". The latter is faster for very high resolution topoplots.}

\item{resol}{integer value, the resolution of the projection locations for
each axis (default: 30L)}

\item{resolcol}{integer value, the number of levels for colouring (default: 
101L)}

\item{projection}{character value (default = "laea"). See 
\url{http://www.remotesensing.org/geotiff/proj_list/} for common projections}

\item{projref}{projection reference (pole [default] or equator)}

\item{origo}{a named character vector of lat and long coordinates of the 
origo}

\item{plot_centroid}{should centroids be plotted (default: TRUE)}

\item{centroid_circle}{numeric scalar (default=0.5), the ratio of individual
centroid coordinates which fall into the area of a shaded circle}

\item{centroid_unitsphere}{should centroids be projected onto the 2d plane as
if they were located on the surface of a unit sphere (default: false)}

\item{plot_bar}{put color bar on the plot (default: TRUE)}

\item{plot_ch}{plot channels as dots (default: TRUE)}

\item{plot_chnames}{plot channel names (default: TRUE)}

\item{title}{title of the plot}

\item{gfp}{a numeric vector of the GFP values in the whole time window}

\item{gfp_max}{the upper limit if gfp is provided}

\item{colour_scale}{function which creates a colour scale with \code{n} values
(\code{n} is defined by \code{resolcol}). The default is to use the 
\code{\link[gplots]{colorpanel}} function, with blue > white > red 
colour scale.}

\item{colour_midpoint}{a number to which the middle colour of the colour scale
corresponds}

\item{...}{arguments to \code{\link{chanInterp}}}
}
\description{
\code{plot2dview} creates 2D topoplots
}
\examples{
# example data
data(erps)
chan_pos <- attr(erps, "chan")   # channel positions

# plot the grand average scalp distribution at 200 ms, for the stimulus 
# class "A" and pairtye "ident" condition; display also the GFP curve of the 
# whole epoch
plotdat <- transformArray(y ~ chan + time, erps,
                          subset = list(stimclass = "A", pairtype = "ident"),
                          datfr = FALSE)
gfpdat <- compGfp(plotdat)
plot2dview(subsetArray(plotdat, list(time = "200")),
           chan_pos, timepoint = 200, gfp = gfpdat)
}

