% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotButterfly}
\alias{plotButterfly}
\title{Create butterfly plot with extras: highlighting and peak topographies}
\usage{
plotButterfly(dat, sig = NULL, topo_time = NULL, chan_pos = NULL,
  subset = list(), pcrit = 0.05, aspect_ratio = 0.5, scalp_ratio = 0.5,
  ampl_range = NULL, caption = TRUE, ...)
}
\arguments{
\item{dat}{the array of ERP curves which must have at least 'time' and/or 
'chan' dimensions}

\item{sig}{[optional] a corresponding array to \code{dat} which is used 
to highlight significant time windows. It must have at least 'time' and 
measure' dimensions. The 'measure' dimension must have 'p' and 'p_corr' 
levels indicating the uncorrected ('p') and corrected ('p_corr') p-values.}

\item{topo_time}{[optional] an object which describes the time points at 
which the scalp topographies should be plotted. Can be a simple atomic vector
or a data.frame if the time points are not identical across the faceting
dimension(s).}

\item{chan_pos}{a data.frame of channel positions. Obligatory if 
\code{topo_time} is not \code{NULL}.}

\item{subset}{a named list to subset the input arrays; see 
\code{\link[eegR]{subsetArray}}}

\item{pcrit}{the level of alpha to highlight significant effects}

\item{aspect_ratio}{the ratio of \code{y} and \code{x} axes (default: 0.5) 
on the figure. If NULL, it is set automatically. If 'topo_time' is provided,
user-defined 'aspect_ratio' might be adjusted to avoid overlapping scalp
maps.}

\item{scalp_ratio}{the ratio of the diameter of the scalp and the vertical
range of the ERP curves on the figure}

\item{ampl_range}{the range of amplitudes to plot. If \code{NULL} (default), 
it is computed from the data.}

\item{caption}{logical flag indicating if caption should be also returned
(default: TRUE)}

\item{...}{additional arguments passed to \code{\link{topoCoord}}}
}
\value{
\code{plotButterfly} returns a ggplot object.
}
\description{
\code{plotButterfly} creates a multi-channel (a.k.a butterfly) plot. 
Additionally, significant time windows can be highlighted, and peak 
topographies can be displayed above the ERP curves.
}
\examples{
# load example data
data(erps)

# extract channel positions and participants' group memberships
chan_pos <- attr(erps, "chan")
dat_id <- attr(erps, "id")

# collapse pairtypes and participants
tempdat <- avgDims(erps, c("pairtype", "id"))

# plot butterfly with topo-maps
plotButterfly(tempdat, topo_time = seq(24, 476, by = 50),
              chan_pos = chan_pos)

}

