% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotComplex}
\alias{plotComplex}
\title{Visualize ERP curves and scalp topographies}
\usage{
plotComplex(dat, sig = NULL, topo_time = NULL, chan_pos = NULL,
  subset = list(), plot_curves = TRUE, pcrit = 0.05,
  aspect_ratio = 0.5, scalp_ratio = 0.5, ampl_range = NULL,
  map_marker_colour = "red", map_marker_shape = 16,
  map_marker_size = 1, ggplot_theme = theme_ndys, ...)
}
\arguments{
\item{dat}{the array of ERP curves which must have at least 'time' and/or 
'chan' dimensions}

\item{sig}{[optional] a corresponding array to \code{dat} which is used 
to highlight significant time windows. It must have at least 'time' and 
measure' dimensions. The 'measure' dimension must have 'p' and 'p_corr' 
levels indicating the uncorrected ('p') and corrected ('p_corr') p-values.}

\item{topo_time}{[optional] an object which describes the time points at 
which the scalp topographies should be plotted. Can be a simple atomic vector
or a data.frame if the time points are not identical across the faceting
dimension(s).}

\item{chan_pos}{a data.frame of channel positions. Obligatory if 
\code{topo_time} is not \code{NULL}.}

\item{subset}{a named list to subset the input arrays; see 
\code{\link[eegR]{subsetArray}}}

\item{pcrit}{the level of alpha to highlight significant effects}

\item{aspect_ratio}{the ratio of \code{y} and \code{x} axes (default: 0.5) 
on the figure. If NULL, it is set automatically. If 'topo_time' is provided,
user-defined 'aspect_ratio' might be adjusted to avoid overlapping scalp
maps.}

\item{scalp_ratio}{the ratio of the diameter of the scalp and the vertical
range of the ERP curves on the figure}

\item{ampl_range}{the range of amplitudes to plot. If \code{NULL} (default), 
it is computed from the data.}

\item{map_marker_colour, map_marker_shape, map_marker_size}{the colour, shape,
and size of the marker used to mark the exact time points of the maps}

\item{ggplot_theme}{a function which produces a ggplot theme 
(default: theme_ndys)}

\item{...}{additional arguments passed to \code{\link{topoCoord}}}
}
\description{
\code{plotComplex} is a workhorse function called by 
\code{\link{plotButterfly}} and \code{\link{plotMap}}.
}
\keyword{internal}
