% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotMap}
\alias{plotMap}
\title{Plot scalp topographies}
\usage{
plotMap(dat, chan_pos, topo_time = NULL, subset = list(),
  aspect_ratio = NULL, ampl_range = NULL,
  map_marker_colour = "grey10", map_marker_shape = "|",
  map_marker_size = 1.2, ...)
}
\arguments{
\item{dat}{either a numeric vector or a numeric array of ERP amplitudes. If
'dat' is a named vector, its names are taken as channel names. If unnamed,
the order of the data must correspond to the order of channels in 'chan_pos'.
If 'dat' is an array, it must have at least 'time' and 'chan' dimensions.}

\item{chan_pos}{a data.frame of channel positions. Obligatory if 
\code{topo_time} is not \code{NULL}.}

\item{topo_time}{[optional] an object which describes the time points at 
which the scalp topographies should be plotted. Can be a simple atomic vector
or a data.frame if the time points are not identical across the faceting
dimension(s).}

\item{subset}{a named list to subset the input arrays; see 
\code{\link[eegR]{subsetArray}}}

\item{aspect_ratio}{the ratio of \code{y} and \code{x} axes on the figure. 
If \code{NULL} (the default), it is set automatically. A user-defined 
'aspect_ratio' might be adjusted to avoid overlapping scalp maps.}

\item{ampl_range}{the range of amplitudes to plot. If \code{NULL} (default), 
it is computed from the data.}

\item{map_marker_colour}{the colour, shape,
and size of the marker used to mark the exact time points of the maps}

\item{map_marker_shape}{the colour, shape,
and size of the marker used to mark the exact time points of the maps}

\item{map_marker_size}{the colour, shape,
and size of the marker used to mark the exact time points of the maps}

\item{...}{additional arguments passed to \code{\link{topoCoord}}}
}
\value{
\code{plotMap} returns a ggplot object.
}
\description{
\code{plotMap} takes an array (or a vector) of ERP amplitudes and plots 
scalp topographies at user-defined time points.
}
\examples{
# load example data
data(erps)

# extract channel positions
chan_pos <- attr(erps, "chan")

# collapse pairtypes and participants
tempdat <- avgDims(erps, c("pairtype", "id"))

# plot topo-maps of stimclass differences from 0 to 500 in 50 ms steps
map <- plotMap(compareLevels(tempdat, "stimclass"),
               topo_time = seq(0, 500, by = 50),
               chan_pos = chan_pos)

# add title
library(ggplot2)
map + ggtitle("Pairwise differences between stimulus classes")
         
}
\seealso{
\code{\link{plotButterfly}} for a more complex way of visualizaton
}
