% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plotTanova}
\alias{plotTanova}
\title{Plot TANOVA results}
\usage{
plotTanova(results, grid = NULL, wrap = NULL, plot_title = "",
  time_label = "Time (ms)", highlight_p = c("p_corr", "p"),
  only_p = FALSE)
}
\arguments{
\item{results}{a list; the return value of the \code{\link{tanova}} function}

\item{grid}{character vector or formula defining the layout of panels}

\item{wrap}{character vector or formula defining the dimension which
separates panels (only considered if grid is NULL)}

\item{plot_title}{character string; the title of the plot}

\item{time_label}{character string; the label of the x (time) axis
(default: "Time (ms)")}

\item{highlight_p}{highlight significant phases based on the corrected
("p_corr", the default) or uncorrected ("p") p-values}

\item{only_p}{logical; if TRUE, p-values are plotted instead of combined
(effect + p-value) plots (default: FALSE)}
}
\value{
A ggplot object
}
\description{
\code{plotTanova} plots the result of the \code{\link{tanova}} function
}

