% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgen_parser.R
\name{popgen_parser}
\alias{popgen_parser}
\title{Common superclass for population genetics parsers}
\description{
Common superclass for population genetics parsers

Common superclass for population genetics parsers
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{sample_ids}}{Ordered vector of sample IDs}

\item{\code{sample_pops}}{Vector of sample populations, ordered like sample_ids}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-POPGEN Parser-clone}{\code{popgen_parser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-POPGEN Parser-clone"></a>}}
\if{latex}{\out{\hypertarget{method-POPGEN Parser-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{popgen_parser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
