% Generated by roxygen2 (4.0.2): do not edit by hand
\name{preAnova}
\alias{preAnova}
\title{Parameter checks and data preparation in \code{arrayAnova}}
\usage{
preAnova(arraydat, factordef, bwdat, useparallel, ncores, par_method,
  usetfce = FALSE)
}
\arguments{
\item{arraydat}{numeric array}

\item{factordef}{a list of between-subject factors (\code{between}),
within-subject factors (\code{within}) and subject identifier (\code{w_id})}

\item{useparallel}{logical; do computations in parallel or not}

\item{ncores}{integer; number of cores for parallel computations}

\item{par_method}{character; parallelization method}

\item{usetfce}{logical; use TFCE correction or not (FALSE, default)}
}
\description{
\code{preAnova} performs parameter checking, creates a data.frame for
the ANOVA design and reshapes arraydat accordingly, and prepares parallel
computations.
}
\keyword{internal}

