% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prepare2plot}
\alias{prepare2plot}
\title{Data preparation mainly aimed at facilitating plotting in lattice or ggplot2}
\usage{
prepare2plot(dat, datid, bwFac = NULL, wiFac = NULL, collFac = NULL,
  diffFac = NULL, compGFP = TRUE, sc = NULL, datfr = TRUE,
  iaFac = NULL, ...)
}
\arguments{
\item{dat}{an array of ERP data. Must have named dimnames, one of which must
be id (corresponding to participants' identification codes)}

\item{datid}{data.frame containing information on participants. Must have an
id column (participants' identification codes)}

\item{bwFac}{named list of between-subject factors (for splitting ERP data)}

\item{wiFac}{named list of within-subject factors (for subsetting ERP data)}

\item{collFac}{character vector of dimension names: average the ERP data
across these dimensions}

\item{diffFac}{a character vector indicating which levels of which dimensions
should be subtracted. The first element is the dimension name, the 2nd and 3rd
are the levels to be compared (2nd-3rd), and the 4th element corresponds to
the label of the new level.}

\item{compGFP}{a logical variable; if TRUE (default), Global Field Power is
also computed.}

\item{sc}{if not NULL, must be a named list indicating how scaling by GFP
should be done - either individually (default) or group-based, and either
timepoint-by-timepoint or for averaged segments}

\item{datfr}{a logical variable (default: TRUE) determining if the resulting
array shall be transformed to a data.frame}

\item{iaFac}{a character vector indicating which dimensions should
be combined (ia is the abbreviation of interaction), if datfr is TRUE}

\item{...}{additional parameters to be passed to \code{\link{array2df}}}
}
\value{
A data.frame if datfr is TRUE, and an array if datfr is FALSE
}
\description{
\code{prepare2plot} provides several options to transform the eeg data array
to a data file which enables direct plotting in lattice or ggplot2 afterwards.
It can also be used for analyses purposes where compact code is desirable.
}

