% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{preSumSq}
\alias{preSumSq}
\title{Prepare sum-of-squares computations}
\usage{
preSumSq(form, form_data, keep_term_order = FALSE, scaled = FALSE,
  y = NULL, between = NULL, adjusted = TRUE)
}
\arguments{
\item{form}{model formula (or character string to \code{as.formula})}

\item{form_data}{a data.frame containing all variables of the formula}

\item{keep_term_order}{a logical value indicating whether the terms should
keep their positions. If FALSE the terms are reordered so that main effects
come first, followed by the interactions, all second-order, all third-order
and so on. Effects of a given order are kept in the order specified.}

\item{scaled}{a logical value indicating if numeric variables in
'form_data' are already scaled (TRUE) or not (FALSE, the default).}

\item{y}{a numeric matrix of response data if residuals should be computed
(usually coming from \code{preAnova})}

\item{between}{the names of between-subject factors. Only used for warning
message in case of unbalanced data.}

\item{adjusted}{adjustment for unbalanced data (default: TRUE)}
}
\value{
A list with the following components: model_matrix,
model_matrix_labels, xpx, unique_contrasts, residuals (optional).
}
\description{
\code{preSumSq} prepares the objects which are needed by
\code{\link{compSumSq}}. Thus, it returns the model.matrix (X), the X'X
matrix, and all unique rows of Xs, and also the residuals if residuals are
permuted instead of raw observations.
}
\keyword{internal}

