% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary.R
\name{print}
\alias{print}
\alias{print.summary.eegr}
\title{Print summaries of \code{\link{arrayTtest}}, \code{\link{arrayAnova}}, or
\code{\link{tanova}} objects}
\usage{
\method{print}{summary.eegr}(x, digits = 3L, quote = FALSE, fill_na = "-", ...)
}
\arguments{
\item{x}{an object of class \code{summary.arrayTtest},
\code{summary.arrayAnova}, or \code{summary.tanova}}

\item{digits}{the minimum number of digits to be printed (default: 3)}

\item{quote}{logical value indicating whether or not strings should be
printed with surrounding quotes (default: FALSE [no quotes printed])}

\item{fill_na}{a character string which is used to indicate NA values in
printed output, that is descriptive statistics which could not be computed
because no data points met the given criterion (default: "-")}

\item{...}{further arguments passed to other methods (not used yet)}
}
\value{
The function invisibly returns the input object.
}
\description{
\code{print.summary.eegr} prints summaries of \code{\link{arrayTtest}},
\code{\link{arrayAnova}}, or \code{\link{tanova}} objects
}
