% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.summary.arrayAnova}
\alias{print.summary.arrayAnova}
\title{Print summaries of \code{\link{arrayAnova}} objects}
\usage{
\method{print}{summary.arrayAnova}(obj, ...)
}
\arguments{
\item{obj}{an object of class \code{summary.arrayAnova}}

\item{...}{not used yet}
}
\value{
The function invisibly returns the input object.
}
\description{
\code{print.summary.arrayAnova} prints summaries of
\code{\link{summary.arrayAnova}} objects
}

