% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.summary}
\alias{print.summary}
\title{Print summaries of \code{\link{arrayTtest}}, \code{\link{arrayAnova}}, or
\code{\link{tanova}} objects}
\usage{
print.summary(obj, ...)
}
\arguments{
\item{obj}{an object of class \code{summary.arrayTtest},
\code{summary.arrayAnova}, or \code{summary.tanova}}

\item{...}{arguments passed to the specific methods}
}
\value{
The function invisibly returns the input object.
}
\description{
\code{print.summary} prints summaries of \code{\link{arrayTtest}},
\code{\link{arrayAnova}}, or \code{\link{tanova}} objects
}

