% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chan_pos.R
\name{project3dMap}
\alias{project3dMap}
\title{Project channel positions onto 2D plane}
\usage{
project3dMap(
  pos,
  r = 1,
  projection = "laea",
  projref = c("pole", "equator"),
  origo = c(lat = ifelse(projref == "pole", 90, 0), long = ifelse(projref == "pole",
    270, 0)),
  inverse = FALSE
)
}
\arguments{
\item{pos}{electrode positions (matrix or data.frame)}

\item{r}{radius}

\item{projection}{character string (default = "laea"). See
\url{http://www.remotesensing.org/geotiff/proj_list/} for common projections}

\item{projref}{projection reference (pole [ = default] or equator)}

\item{origo}{a named character vector of lat ( = latitude) and long (longitude)}

\item{inverse}{if set to TRUE, back-projection is performed (default = FALSE)}
}
\value{
A data.frame containing the projected coordinates
}
\description{
\code{project3dMap} projects 3D coordinates onto a 2D plane or vice versa
}
