\name{read.bdf.hdr}
\alias{read.bdf.hdr}
\title{Read Header from BDF File}
\description{Read Header from BDF File}
\usage{read.bdf.hdr(bdf.file)}
\arguments{
  \item{bdf.file}{Full file listing to BDF file}
}
\value{
A list with fields for each bit of header information. See file format in references.
}
\references{
<http://www.biosemi.com/faq/file_format.htm>
}
\author{
Timothy R. Koscik <timkoscik+eegr@gmail.com>
}