% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEDF.R
\name{readEDF}
\alias{readEDF}
\title{readEDF}
\usage{
readEDF(file, records = "all", signals = "all",
  multifreq = c("group", "separate", "downsample", "upsample"),
  physical = TRUE)
}
\arguments{
\item{file}{Character string. The name of the file which the data are to be read from. \code{file} 
can also be a complete URL. (For the supported URL schemes,see the ‘URLs’ section of the help for url.)}

\item{records}{Numeric or \code{'all'}. In EDF, signals are stored in 'records', usually containing 0.5 or 1
second of data. This function either reads all records (default: \code{all}), or a subset of the records, thus 
limiting the time interval of the signals. \code{records} can be a range (e.g., \code{1:50}) or a random sequence 
of records (e.g., \code{c(1,5,12,3)}). In the latter case, the records are returned in the specified order (only
useful for discontinuous data). If \code{length(records) == 0} or \code{is.na(records)}, then no records are read
and only the file header is returned.}

\item{signals}{Numeric, character string, or \code{'all'}. Specifies the signals to read from the input file. 
\code{signals} can be a range (e.g., \code{1:21}) a random sequence of signals (e.g., \code{c(1,5,12,3)}),
or a sequence of signal labels (e.g., \code{c("F3", "Fz", "F4")}. In the latter two cases, the signals are returned 
in the specified order. If \code{length(signals) == 0} or \code{is.na(signals)}, then no records are read
and only the file header is returned.}

\item{multifreq}{Character string. Specifies what to do when signals with different sampling frequencies are
present in the file. Options are:
\describe{
  \item{\code{"group"}}{Signals with identical sampling frequencies are grouped together in a single \code{ctd} object
  (Default)}
  \item{\code{"separate"}}{Each signal will have its own \code{ctd} object}
  \item{\code{"downsample"}}{Signals with higher sampling frequencies are downsampled to the lowest frequency in the file}
  \item{\code{"upsample"}}{Signals with lower sampling frequencies are upsampled to the highest frequency in the file}
}
Up- and downsampling are performed by the \code{resample} function in the \code{\link{signal}} package. Note that this can 
take a lot of time, especially in case of upsampling to the frequency of the annotation signal (usually 1-2 kHz).}

\item{physical}{Logical. If true, convert the digital data values to physical values. \eqn{physval = gain * digival + offset},
where \eqn{gain = (physmax - physmin) / (digimax - digimin)}, and \eqn{offset = physmax - gain * digimax}. See the description
of the header below for the definition of the variables used in these equations.}
}
\value{
\code{list (head, ctd, ..., anno)}. The function returns a \code{list} containing three possible elements: the
file header, the data, and annotations, if present. The header is always returned; the other two elements depend on the
input arguments.
\describe{
  \item{\strong{header}}{A list named \code{'header'} consisting of the following elements:
    \describe{
      \item{General header:\cr}{
        \describe{
          \item{\code{id0}}{First identification byte. For BDF: -1 (0xFF), for EDF: ASCII "0" (0x30, old) or " " (0x32, new)}
          \item{\code{id1 }}{Identification string, "" for EDF, "BIOSEMI" for BDF (7 bytes ASCII)}
          \item{\code{lsi}}{Local subject identification (80 bytes ASCII)}
          \item{\code{lri}}{Local recording identification (80 bytes ASCII)}
          \item{\code{date}}{Starting date of recording (8 bytes ASCII - "dd.mm.yy")}
          \item{\code{time}}{Starting time of recording (8 bytes ASCII - "hh.mm.ss")}
          \item{\code{nbytes}}{Number of bytes in header record (converted from 8 bytes ASCII to numeric)}
          \item{\code{version}}{Reserved. Used for version of format. Empty for EDF, "EDF+C" for EDF continuous data,
            "EDF+D" for EDF discontinuous data, "BIOSEMI" or "24BIT" for BDF (44 bytes ASCII)}
          \item{\code{nrec}}{Number of data records (converted from 8 bytes ASCII to numeric), -1 if unknown}
          \item{\code{duration}}{Duration of one data record in seconds (converted from 8 bytes ASCII to numeric)}
          \item{\code{ns}}{Number of signals in each data record (converted from 4 bytes ASCII to numeric)}
        }
      }
      \item{Signal-specific header (one for each signal in the file):\cr}{
        \describe{
          \item{\code{label}}{Signal labels, e.g. "Fpz", "Fz" (ns * 16 bytes ASCII)}
          \item{\code{transducer}}{Transducer type, e.g. "AgAgCl electrode" (ns * 80 bytes ASCII)}
          \item{\code{dimension}}{Physical dimension of signal, e.g., "uV", "Ohm" (ns * 8 bytes ASCII)}
          \item{\code{physmin}}{Physical minimum in units of dimension (ns * 8 bytes ASCII converted to numeric)}
          \item{\code{physmax}}{Physical maximum in units of dimension (ns * 8 bytes ASCII converted to numeric)}
          \item{\code{digimin}}{Digital minimum ((ns * 8 bytes ASCII converted to numeric)}
          \item{\code{digimax}}{Digital maximum (ns * 8 bytes ASCII converted to numeric)}
          \item{\code{gain}}{Gain factor used for mapping digital to physical values (ns * numeric)}
          \item{\code{offset}}{Offset value used for mapping digital to physical values (ns * numeric)}
          \item{\code{prefilt}}{Prefiltering, e.g., "TC:3s;LP:70Hz" (ns * 80 bytes ASCII)}
          \item{\code{nsamp}}{Number of samples in one data record (ns * 8 bytes ASCII converted to numeric)}
          \item{\code{resvd}}{Reserved (ns * 32 bytes ASCII)}
        }
      }
    }
  }
  \item{\strong{data}}{One or more Continuous Time Domain (\code{\link{ctd}}) objects. If the input file contains signals
    which all have the same sampling frequency, or if either the \code{"upsample"} or \code{"downsample"} options were 
    specified in the \code{multifreq} parameter, then there is a single \code{ctd} object, and it is named '\code{ctd}'.\cr\cr
    In case of multiple sampling frequencies, then if the \code{multifreq = "group"}, as many \code{ctd} objects are
    returned as there are sampling frequencies, and they are named \code{ctdF}, where F equals the sampling frequency
    of the data in that object. For instance, for a data file with sampling frequencies of 256 and 1024 Hz, two \code{ctd}
    objects will be returned, named \code{ctd256} and \code{ctd1024}.\cr\cr
    In case \code{multifreq = "separate"} is specified, there will be as many \code{ctd} objects as there are signals, 
    and they will be named \code{'ctd1...ctdN'}, where \code{N} equals the number of signals in the file.
  }
  \item{\strong{annotations}}{A (list of) data frame(s) named \code{'anno'} containing the annotations parsed from
  the EDF+/BDF+ annotation channels, or from the Biosemi 'status' channel. Each annotation consists of the following fields:
    \describe{
      \item{\code{sample}}{sample number of annotation}
      \item{\code{time}}{corresponding time, relative to beginning of file}
      \item{\code{value}}{value of annotation, as per EDF+ specification, see \url{https://www.edfplus.info/specs/edfplus.html}.
        For Biosemi this ia a 16 bit number, as only bits 0-15 of 24 are used as triggers, see
        \url{http://biosemi.com/faq/trigger_signals.htm}}.
    }
    For EDF+ and BDF+ files, the Time-stamped Annotation Lists (TALs) will also be returned. In this case, sample numbers and
    times are relative to the beginning of the file, irrespective of the records read.
  }
}
}
\description{
Read data from a European Data Format (EDF, EDF+, BDF, BDF+) file.
}
\details{
The function readEDF reads files containing electrophysiological data stored in the European Data
Format (EDF) or its variants.

EDF was published in 1992 and stores multichannel data, allowing different sample rates for each signal.
Internally it includes a header and one or more data records. The header contains some general information 
(patient identification, start time...) and technical specs of each signal (calibration, sampling rate, filtering, ...),
coded as ASCII characters. The data records contain samples as little-endian 16-bit integers.

EDF+ was published in 2003 and is largely compatible to EDF, but EDF+ files also allow coding discontinuous recordings
as well as annotations, stimuli and events in UTF-8 format.

Biosemi developed an EEG acquisition system with 24-bit A/D converters, which could not be stored in the 16-bit
EDF format. They introduced a 24-bit variant of EDF called BDF (Biosemi Data Format), and a BDF+ variant is sometimes also
found.

The function \code{readEDF} can read all of these variants and produce different output that can be used for further processing
by function in the \code{eegr} package.
}
\examples{
\dontrun{
  edf <- readEDF('test.edf', signals = 1:10, records = "all")
  summary(edf$ctd)
  header <- readEDF(file.bdf, records = NA)
  header
}
}
\references{
Kemp B., Värri, A., Rosa, A.C., Nielsen, K.D. and Gade, J. (1992). A simple format for exchange of
digitized polygraphic recordings. \emph{Electroencephalogr Clin Neurophysiol}, \emph{82(5)}, 391-393.

Kemp, B. and Olivan, J. (2003). European data format ’plus’ (EDF+), an EDF alike standard format
for the exchange of physiological data. \emph{Clin Neurophysiol}, \emph{114(9)}, 1755-1761.

European Data Format. \url{https://www.edfplus.info/}

Which file format does Biosemi use? \url{https://www.biosemi.com/faq/file_format.htm}
}
\author{
Geert van Boxtel, \email{G.J.M.vanBoxtel@gmail.com}
}
