% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rearrangeList}
\alias{rearrangeList}
\title{Rearrange two-level list}
\usage{
rearrangeList(dat, newname)
}
\arguments{
\item{dat}{the list to be rearranged}

\item{newname}{character string, the name of the highest-level factor}
}
\value{
A matrix or array, if the input is a one-level list, and a one-level
list, if the input is a two-level list
}
\description{
\code{rearrangeList} reshapes a special type of one- or two-level lists.
}
\details{
One or two level lists which contain identically shaped elements at
the base level, and elements at the base level are vectors, matrices or
arrays, can be rearranged to a matrix/array (from a one-level list) or to a
one-level list (from a two-level list) by binding identical elements.
This way the second level of the list will be represented as an additional
dimension in the ground level elements (vectors will be matrices, matrices
will become arrays, arrays recieve an extra dimension)
}

