% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Replace}
\alias{Replace}
\title{Replace elements of a vector}
\usage{
Replace(x, from, to, digits = 10L)
}
\arguments{
\item{x}{a raw, integer, character or numeric (double) vector}

\item{from}{vector of elements to replace}

\item{to}{the elements}

\item{digits}{integer indicating the number of decimal places to be used in
the comparison of double values (ignored if \code{is.double(x)} is FALSE)}
}
\value{
a vector with replaced values
}
\description{
\code{Replace} replaces the values in \code{x} which are equal to 
\code{from} to the values given in \code{to}. The values in \code{to} are
recycled to match the length of \code{from}.
}
\examples{
x <- c(NA, c("a", "a", "z", "e", "q"))

# note that 'from' might contain values which are not present in 'x'
xr <- Replace(x, from = c("w", "a", "e"), to = c("ww", "aa", "ee"))

# only 'a' and 'e' were replaced
xr
stopifnot(identical(length(x), length(xr)))
stopifnot(identical(xr[c(2, 3, 5)], c("aa", "aa", "ee")))

# the missing value is not affected
stopifnot(is.na(xr[1]))
}

