% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{repLen}
\alias{repLen}
\title{Match the length of an object to an other object with warning}
\usage{
repLen(x, len, x_name = NULL)
}
\arguments{
\item{x}{the object}

\item{len}{the desired length}
}
\description{
\code{repLen} is a wrapper around \code{rep_len} to recycle or crop a
vector so that its length matches a pre-defined length. It throws a warning
in such cases.
}
\keyword{internal}

