% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reref.R
\name{reref}
\alias{reref}
\title{Rereference Data}
\usage{
reref(
  x,
  fs = 1,
  sensors = setdiff(colnames(x), "t"),
  oldref = NULL,
  newref = "robust",
  saveref = FALSE,
  interp = c("post-reference", "pre-reference", "none"),
  estmean = c("median", "huber", "mean"),
  sl = getLocationsfromLabels(colnames(x[, sensors])),
  maxIter = 4
)
}
\arguments{
\item{x}{input time series, specified as a numeric matrix or vector. In case
of a vector it represents a single signal; in case of a matrix each column
is a signal. Alternatively, an object of class \code{\link{ctd}}}

\item{fs}{sampling frequency of \code{x} in Hz. Default: 1. Overruled if
\code{x} is a \code{ctd} object, in which case the sampling frequency is
\code{fs(x)}.}

\item{sensors}{vector of sensor names or numbers that are used for the
calculations (hence they must contain EEG data). Default: all.}

\item{oldref}{vector of sensor names or numbers that formed the original
reference. If specified, the old reference signal is reconstructed back
into the data object (added if it was not present in the data). Ignored if
\code{x} is an object of class \code{ctd} and contains a \code{"ref"}
attribute. Default: NULL.}

\item{newref}{signals that form the new reference, either specified as a
character vector of signal names (\code{colnames(ctd_obj)}), or as a vector
of numerical values. If more than one signal is specified, then the average
of these signals will form the new reference. Alternatively, specifying
\code{"robust"} (default) computes an average reference with iterative
detection and interpolation of bad channels, and \code{"average"} removes
the average of the reference channels with no interpolation.}

\item{saveref}{logical indication whether to save to new reference as a}

\item{interp}{For \code{newref = "robust"}, specifies whether to perform
final channel interpolation. If \code{"post-reference"} (Dafault), after a
final robust reference is computed, the channels are re-interpolated and
the reference is corrected. In \code{"pre-reference"}, reref incrementally
adds to the bad channel list and interpolates before computing the
reference. If the initial estimate of the reference is poor, this is not a
good approach. In the \code{"none"} option, reref removes the reference but
does not interpolate. Bad channels remain in the signal. You may choose to
remove them during post-processing.}

\item{estmean}{For \code{newref = "robust"}, statistic used for estimation of
the mean; one of \code{"median"} (default), \code{"huber"} or
\code{"mean"}. Median and mean are the obvious statistics; \code{"huber"}
uses the Huber M-estimator of location with median absolute deviation (MAD)
scale, which is computationally expensive.}

\item{sl}{For \code{newref = "robust"}, sensor locations of \code{x},
specified as a data frame according to \code{\link{sensorlocs}} format.
Alternatively, a matrix with named columns x, y, and z representing the
sensor locations. Default: obtained from the sensor names given by the
column names of \code{x}. Only used in case \code{"when"} is either
\code{"before"} or \code{"after"}.}

\item{maxIter}{maximum number of iterations used for detecting noisy channels
and recomputing the reference. Default: 4. Only used in case \code{"newref"}
is \code{"robust"}.}
}
\value{
A list containing the following elements:
\describe{
  \item{y}{An object of the same class as the input, containing the
  re-referenced data. Signals not included in \code{sensors} are copied to
  the output object. If \code{saveref = TRUE} then the new reference is added
  as a separate data channel (column). If the old reference was specified
  through the \code{oldref} parameters, and it was not also present in the
  data, then it is also added.}
  \item{fs}{the sampling frequency of \code{y}}
  \item{sensors}{the sensors that \code{reref} operated on (on input)}
  \item{oldref}{the original reference sensor(s)}
  \item{newref}{the new reference sensor(s)}
  \item{saveref}{logical indicating whether the new reference is saved}
  \item{estmean}{for robust referencing, the initial estimate of the mean}
  \item{interp}{for robust referencing the instant of interpolating noisy
  sensors, either before or after re-referencing}
  \item{iterations}{for robust referencing, the number of iterations
  performed to detect noisy reference sensors}
  \item{bad}{a list of sensors that were marked as bad by different
  criteria (see \code{\link{noisysensors}})}
}
}
\description{
Compute the average reference or a new common reference for data.
}
\details{
The function \code{reref} computes the mean of the new reference sensors and
subtracts that from the channels specified in \code{sensors}. If the data
were originally referenced with respect to a sensor that is not present in
the data, then this sensor will be added, and it can be used in the new
reference.

Robust statistics can be used to compute the new reference (median or Huber
mean instead of arithmetic mean). In addition, \code{reref} can use an
iterative procedure to remove bad channels before or after re-referencing,
which makes the re-referencing even more robust to outlier sensors. The
detection of noisy sensors is done using the \code{\link{noisysensors}}
function, using the default setup.
}
\examples{

avgref <- reref(EEGdata, sensors = 1:29, oldref = "M1",
                newref = "average")
lmref <- reref(EEGdata, sensors = 1:29, oldref = "M1", 
                newref = "M1 M2")

\dontrun{
rref <- reref(EEGdata, 1:29, "M1", "robust")
}

}
\references{
Bigdely-Shamlo, N., Mullen, T.,, Kothe, C.,, Su, K.-M., and
  Robbins, K. A. (2015). The PREP pipeline: standardized preprocessing for
  large-scale EEG analysis. Frontiers in Neuroinformatics, 9, Article 16,
  \url{https://www.frontiersin.org/article/10.3389/fninf.2015.00016}.

Huber, P.J. (1981) Robust Statistics. Wiley.
}
\seealso{
\code{\link{noisysensors}}, \code{\link[MASS]{huber}}
}
\author{
Geert van Boxtel, \email{G.J.M.vanBoxtel@gmail.com}, based on Matlab
  code of the PREP pipeline by Nima Bigdely-Shamlo and colleagues.
}
