% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerefdata.R
\name{rerefdata}
\alias{rerefdata}
\title{Rereference Data}
\usage{
rerefdata(ctd_obj, signals, new_ref, keep_ref = FALSE, old_ref)
}
\arguments{
\item{ctd_obj}{object of class \code{'ctd'} containing continuous time domain
data.}

\item{signals}{signals to compute the new reference on, either specified as a
character vector of signal names (\code{colnames(ctd_obj)}, or as a vector
of numerical values.). Alternatively, specifying \code{"all"} (default)
computes the new reference of all signals, except the signal(s) specified
in \code{ref}.}

\item{new_ref}{signals that form the new reference, either specified as a
character vector of signal names (\code{colnames(ctd_obj)}), or as a vector
of numerical values. If more than one signal is specified, then the average
of these sinals will form the new reference. Alternatively, specifying
\code{"average"} (default) computes the average reference over the signals
specified by \code{signals}.}

\item{keep_ref}{logical indicating whether to keep the reference in the output.
Default: \code{FALSE}.}

\item{old_ref}{list with the sensor locations, either ('label', 'theta',
'phi'), or ('label, 'x', 'y', 'z'), specifying the location of the original
reference. If specified, the old reference signal is reconstructed back
into the data object.}
}
\value{
A \code{ctd} object containing the re-referenced data. Signals not
  included neither in \code{signals} nor in \code{ref} are copied to the
  output object.
}
\description{
Compute the average reference or a new common reference for data in a
\code{ctd} object.
}
\examples{


#data(EEGdata)
#new <- rerefdata(EEGdata, 'all', 'average')
#plot(new)
#new <- rerefdata(EEGdata, 'all', c('M1', 'M2'))

}
\seealso{
\code{\link{sensorlocs}}
}
\author{
Geert van Boxtel, \email{G.J.M.vanBoxtel@gmail.com}
}
