% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/array_transform.R
\name{revMergeDims}
\alias{revMergeDims}
\title{Reverse mergeDims transformation}
\usage{
revMergeDims(dat)
}
\arguments{
\item{dat}{numeric matrix or array with merged dimensions}
}
\value{
An array of the same dimension attributes as the array which
 \code{\link{mergeDims}} was called on
}
\description{
\code{revMergeDims} sets back the original array transformed by
 \code{\link{mergeDims}}
}

