% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erp_preprocess.R
\name{scaleChan}
\alias{scaleChan}
\title{Scale channels}
\usage{
scaleChan(dat, keep_dimorder = TRUE)
}
\arguments{
\item{dat}{numeric matrix or array with names dimensions, one of which
must be "chan"}

\item{keep_dimorder}{logical value; if TRUE (default), the order of the 
dimensions is kept intact, otherwise the channel dimension will be 
the first dimension in the resulting matrix or array}
}
\value{
A numeric matrix or array
}
\description{
\code{scaleChan} normalizes the data across channels so that for
each sampling point, the mean of the channel amplitudes is zero and
the standard deviation is one.
}
