% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectdata.R
\name{selectdata}
\alias{selectdata}
\title{Select Data}
\usage{
selectdata(
  ctd_obj,
  seconds = NULL,
  samples = NULL,
  signals = NULL,
  add.t = FALSE
)
}
\arguments{
\item{ctd_obj}{object of class \code{'ctd'} containing continuous time domain
data.}

\item{seconds}{character string representing an expression to select the
specified seconds from the data; can only used when \code{ctd_obj} contains
a \code{'t'} variable alongside the signals.}

\item{samples}{samples to select from the data; must be between 1 and
\code{npts(ctd_obj)}.}

\item{signals}{signals to select from the data. Either a vector of numeric
values between 1 and \code{ns(ctd_obj)}, or a character vector with sensors
names from \code{colnames(ctd_obj)}.}

\item{add.t}{logical indicating whether to add or renew the time variable t.
If \code{TRUE} and \code{ctd_obj} did not contain a variable \code{t}, it
is added. If it did contain a variabe \code{t}, it is updated starting at
0. If \code{FALSE} (default), an existing variable \code{t} is left
unchanged.}
}
\value{
A \code{ctd} object containing the selected data.
}
\description{
Select data from a \code{ctd} object.
}
\examples{

data(EEGdata)
sel <- selectdata(EEGdata, seconds = 'which(t < 3)')
sel <- selectdata(EEGdata, samples = 1:600)
sel <- selectdata(EEGdata, signals = 1:28)

}
\author{
Geert van Boxtel, \email{G.J.M.vanBoxtel@gmail.com}
}
