% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{setattributes}
\alias{setattributes}
\title{Attach dimension and label attributes to test statistics}
\usage{
setattributes(x, obj = NULL, label = NULL, dimorder = NULL)
}
\arguments{
\item{x}{the vector or matrix of the test statistic}

\item{obj}{if not NULL (default), a list object returned by
\code{\link{preTtest}} or \code{\link{preAnova}}}

\item{label}{if not NULL (default), a character string with a verbose
description of the test statistic (e.g. "Traditional F-value", "Welch t",
"Generalized Eta Squared", etc.)}

\item{dimorder}{if not NULL (default), a numeric or character vector
indicating the order (or subset) of the dimensions}
}
\description{
Used internally in the workhorse functions for arrayAnova, arrayTtest, etc.
to set the dimension, dimension names and nice label of the given test
statistic
}
\keyword{internal}

