% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{setDoBackend}
\alias{setDoBackend}
\title{\code{setDoBackend} is identical to \code{\link[foreach]{setDoPar}}, but 
returns the internal of the previously registered backend.}
\usage{
setDoBackend(data)
}
\arguments{
\item{data}{internal data of a foreach \%dopar\% backend.}
}
\description{
\code{setDoBackend} is identical to \code{\link[foreach]{setDoPar}}, but 
returns the internal of the previously registered backend.
}
\author{
Renaud Gaujoux, Denes Toth
}
\keyword{internal}

