% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{setSeed}
\alias{setSeed}
\title{Set seed}
\usage{
setSeed(seed)
}
\arguments{
\item{seed}{either NULL, in which case \code{setSeed} does nothing, or an 
integer value, in which case it is interpreted as the 'seed' argument in
\code{\link{set.seed}}, or a list of arguments passed to 
\code{\link{set.seed}}}
}
\description{
\code{setSeed} is called for its side effect, namely it specifies a random 
seed by calling \code{\link{set.seed}}.
}
