% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{sigPhases}
\alias{sigPhases}
\title{Internal function to create time ranges of significant effects}
\usage{
sigPhases(dat, pcrit, ymax)
}
\arguments{
\item{dat}{a numeric array of corrected and uncorrected p-values}

\item{pcrit}{numeric value of the level of alpha (e.g., 0.05)}

\item{ymax}{y coordinate used for setting the highlighted region}
}
\description{
\code{sigPhases} is an internal function which extracts significant
time ranges from the array of p-values. Used for highlighting significant
phases in butterfly plots
}
\keyword{internal}
