% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erp_preprocess.R
\name{splitMarker}
\alias{splitMarker}
\title{Split concatenated strings.}
\usage{
splitMarker(marker, header, type = NULL, splitchar = "_")
}
\arguments{
\item{marker}{a vector which contains concatenated elements}

\item{header}{a character vector which determines the column names of the
resulting data.frame}

\item{type}{a character vector which determines the class of each column in
the resulting vector. If its length is less then the number of columns, it
will be recycled. If set to NULL (default), character vectors are transformed to
factors.}

\item{splitchar}{a character vector of length one indicating the splitting
character (default: _). Splitting characters which are special characters in
R (e.g. "|", ".", etc.) should be given as in strsplit (e.g. "\\\|")}
}
\value{
A data.frame with header and appropriate classes, containing the 
spltted substrings of the original vector elements
}
\description{
\code{splitMarker} is a conveniance function wrapping strsplit. It returns
a data.frame with columns of customizable classes. Useful for post-processing
markers.
}
\seealso{
\code{\link{strsplit}}
}
