% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssinterp.R
\name{ssinterp}
\alias{ssinterp}
\title{ssinterp}
\usage{
ssinterp(
  src,
  dest,
  lambda = 1e-05,
  order = 4,
  type = c("spline", "slap"),
  tol = 1e-07
)
}
\arguments{
\item{src}{source sensor positions, specified as a data frame (or coerceable
to a data frame), with 3 columns labeled x, y, and z.}

\item{dest}{destination sensor positions, specified as a data frame (or
coerceable to a data frame) with 3 columns labeled x, y, and z.}

\item{lambda}{regularization parameter for smoothing the estimates, specified
as a numeric scalar value. Default: 1e-5.}

\item{order}{order of the polynomial interpolation to use, specified as a
positive integer scalar value. Default: 4}

\item{type}{character string specifying which interpolation method to use.
One of \code{"spline"} - spherical spline interpolation) or \code{"slap"} -
surface Laplacian (CSD). Default: \code{"spline"}}

\item{tol}{tolerance for the Legendre polynomial approximation, specified as a
numeric scalar value. Default: 1e-7.}
}
\value{
linear mapping matrix between old and new coordinates; a matrix with
  as many rows as the number of rows in \code{dest} and as many columns as
  the number of rows in \code{src}
}
\description{
Spherical spline interpolation of sensor locations
}
\examples{
src <- matrix(runif(300), 100, 3)
src[, 3] <- abs(src[, 3])
dest <- matrix(runif(90), 30, 3)
dest[, 3] <- abs(dest[, 3])
W <- ssinterp(src, dest)

}
\references{
Perrin, F., Pernier, J., Bertrand, O., & Echallier, J. F. (1989).
  Spherical splines for scalp potential and current density mapping.
  Electroencephalography and clinical neurophysiology, 72(2), 184-187.
}
\author{
Matlab code 2009 by Jason D.R. Farquhar (\email{jdrf@zepler.org}),
  adapted 2014 by Kay Robbins. R code by Geert van Boxtel,
  \email{G.J.M.vanBoxtel@gmail.com}
}
