% Generated by roxygen2 (4.0.1): do not edit by hand
\name{subsetArray}
\alias{subsetArray}
\title{Return subset of an array}
\usage{
subsetArray(dat, subsets, keep_attr = TRUE, ...)
}
\arguments{
\item{dat}{array to be subsetted}

\item{subsets}{a named list of character vectors indicating which rows of
which dimensions to keep (see Details)}

\item{keep_attr}{a logical variable which determines if the result inherits
the custom attributes of the input (TRUE, default) or not}

\item{...}{further arguments to be passed to asub}
}
\value{
An array which contains only the selected level of the given
dimensions.
}
\description{
\code{subsetArray} is a wrapper for asub() in the abind package. It returns
a subset of an array and assumes that the array has named dimnames.
}
\details{
Dimensions which are not listed in subsets are not affected. Names
of subsets indicate which dimensions are to be subsetted in the input array,
and each list element indicates which levels of the given dimension will be
kept. If a list element is a named empty vector, the correspondig dimension
will not be affected by the function.
}

