% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_transform.R
\name{subsetIndices}
\alias{subsetIndices}
\title{Create subset indices for subsetArray}
\usage{
subsetIndices(subset., which_dims., dim., dimnames.)
}
\arguments{
\item{subset.}{a possibly named list of subset indices (either logical,
integer or is* functional indices)}

\item{which_dims.}{the indices of the dimensions which should be subsetted}

\item{dim.}{the dimensions of the data to subset on}

\item{dimnames.}{the dimension names of the data to subset on}
}
\description{
\code{subsetIndices} check arguments and creates subset indices for
subsetArray and `subsetArray<-`.
}
\keyword{internal}

