% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.arrayAnova}
\alias{summary.arrayAnova}
\title{Compute summary statistics for \code{arrayAnova} object}
\usage{
\method{summary}{arrayAnova}(obj, basis = c("p_corr", "p", "es", "stat",
  "stat_corr"), crit = 0.05, ...)
}
\arguments{
\item{obj}{an object of class \code{\link{arrayAnova}}}

\item{basis}{character value indicating the criterion; if \code{p_corr}
(the default), channel X time points below a specified corrected alpha-level
are selected (choose \code{p} for uncorrected p-values); if \code{es} or
\code{stat} (or \code{stat_corr}), the effect size or (corrected) test
statistic should exceed a given limit.}

\item{crit}{the level of criterion}

\item{...}{arguments passed to \code{\link{extract}}, e.g.
\code{time_window = c(100, 200)}}
}
\value{
a list of summaries
}
\description{
\code{summary.arrayAnova} computes the minimum, maximum and median of
F-values, generalized eta squared effect sizes and P-values for all
channel X time points which meet a certain criterion.
}

