% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary.R
\name{summary.eegr}
\alias{summary.eegr}
\title{Workhorse function for eegR-summary methods}
\usage{
\method{summary}{eegr}(object, basis, basis_dim, crit, ...)
}
\arguments{
\item{object}{an object of class \code{\link{arrayTtest}},
\code{\link{arrayAnova}}, or \code{\link{tanova}}}

\item{basis}{character value indicating the criterion node}

\item{basis_dim}{integer or character vector indicating the dimensions
which form relevant subparts for which summary statistics should be
computed}

\item{crit}{numeric scalar, the level of criterion}

\item{...}{arguments passed to \code{\link{extract}}, e.g.
\code{time_window = c(100, 200)}}
}
\description{
\code{summary.eegr} is the main function behind eegR-summary methods
(\code{\link{summary.arrayTtest}}, \code{\link{summary.arrayAnova}},
\code{\link{summary.arrayAnova}}).
}
\keyword{internal}

