% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary.R
\name{summary}
\alias{summary}
\alias{summary.arrayTtest}
\alias{summary.arrayAnova}
\alias{summary.tanova}
\title{Compute summary statistics for \code{arrayTtest}, \code{arrayAnova}, or
\code{tanova} objects}
\usage{
\method{summary}{arrayTtest}(object, basis = c("p_corr", "p", "es",
  "stat", "stat_corr"), basis_dim = c("chan", "time"), crit = NULL,
  ...)

\method{summary}{arrayAnova}(object, basis = c("p_corr", "p", "es",
  "stat", "stat_corr"), basis_dim = c("chan", "time"), crit = NULL,
  ...)

\method{summary}{tanova}(object, basis = c("p_corr", "p", "stat",
  "stat_corr"), basis_dim = "time", crit = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{\link{arrayTtest}},
\code{\link{arrayAnova}}, or \code{\link{tanova}}}

\item{basis}{character value indicating the criterion node (default:
"p_corr"), see Details}

\item{basis_dim}{integer or character vector indicating the dimensions
which form relevant subparts for which summary statistics should be
computed (default: c("chan", "time"))}

\item{crit}{numeric scalar, the level of criterion (see Details)}

\item{...}{arguments passed to \code{\link{extract}}, e.g.
\code{time_window = c(100, 200)}}
}
\value{
The methods return a list of summaries which can be printed with
the corresponding \code{print} method.
}
\description{
\code{summary.arrayTtest}, \code{summary.arrayAnova}, and
\code{summary.tanova} compute the minimum, maximum and median of
criterion-referenced test statistics, effect sizes and P-values for
particular subviews (e.g. channel X time subviews).
}
\details{
If 'basis' is "p_corr" (the default) or "p", denoting corrected and
uncorrected P-values, respectively, argument 'crit' refers to the desired
alpha-level (set to 0.05 by default). In this case the basis values should be
\emph{below} or equal to the criterion. If basis is "stat" (test statistic),
"stat_corr" (corrected test statistic) or "es" (effect size), absolute basis
values \emph{above} or equal to the 'crit' criterion are selected.\cr
The summary statistics are computed for those data points which correspond
to the selected positions in the 'basis_dim' subviews. For example,
supposing the object is the result of an uncorrected ANOVA calculation (see
\code{\link{arrayAnova}}), and separate analyses were conducted for each
channel X time data points but no other dimension, \code{summary} returns
the list of \code{min}, \code{max}, and \code{median} of F-values and effect
sizes (generalized eta squares) taking into consideration only those
channel X time values for each model term, which were significant at the
level of 0.05.\cr
The returned list also contains the descriptive statistics of the selected
P-values together with the number and ratio of significant values. Note that
if no data points meet the given criterion in a particular subview, no
summaries can be computed, resulting in \code{NA} values in the
corresponding cells.
}
