% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sumsq}
\alias{sumsq}
\title{Compute Sum of Squares}
\usage{
sumsq(form, f_dat, a_dat, dimn, keep_term_order = FALSE, whichterm = NULL,
  no_icpt = FALSE, return_means = FALSE, return_array = TRUE)
}
\description{
Compute Sum of Squares
}
\keyword{internal}

