% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matstat_utils.R
\name{sweepMatrix}
\alias{sweepMatrix}
\title{Fast replacement for \code{sweep}}
\usage{
sweepMatrix(x, MARGIN, STATS, FUN = c("-", "+", "*", "/", "^", "==",
  "!=", "<", "<=", ">", ">="), has_NA = NULL)
}
\arguments{
\item{x}{numeric matrix}

\item{MARGIN}{an integer index of the dimension to sweep over (1 for rows,
2 for columns). Can be a character value if the input matrix has named 
dimension names.}

\item{STATS}{a numeric vector; the summary statistic which is to be swept 
out. It is recycled to match the size of \code{MARGIN}.}

\item{FUN}{character value, one of "+", "-", "*", "/", "^", "==", "!=", "<", 
"<=", ">=", ">"}

\item{has_NA}{logical value indicating if there is any missing value in 
\code{x}. If NULL (default), it is explicitly checked by calling 
\code{\link{anyNA}} on \code{x}.}
}
\value{
\code{sweepMatrix} returns a matrix of the same type as
\code{FUN(x[1, 1], y[1])}
}
\description{
\code{sweepMatrix} is a faster and less memory-hungry version of 
\code{sweep(x, MARGIN, STATS, FUN = c("-", "+", "*", "/", "<", ...))}.
}
\examples{
# create a matrix and choose its first row as a statistic
mat <- matrix(1:20, 4, 5)
stat <- mat[1,]

# suppose we have missing values in the matrix and the statistic
mat[2, 3] <- NA
stat[1] <- NA

# compute and print the result
( result <- sweepMatrix(mat, 2, stat, "-") )

# check
stopifnot(all.equal(result, 
                    sweep(mat, 2, stat, "-")))
stopifnot(identical(result[1, ], c(NA, rep(0L, ncol(mat) - 1))))
}
