% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tanova}
\alias{tanova}
\title{Topographical ANOVA (TANOVA) and related methods}
\usage{
tanova(arraydat, factordef, bwdat = NULL, type = c("tanova",
  "dissimilarity", "gfp"), verbose = TRUE, nperm = 999L,
  useparallel = FALSE, ncores = NULL, par_method = c("snow", "mc"),
  cl = NULL, iaterms_last = TRUE, seed = NULL, pcrit = 0.05)
}
\arguments{
\item{arraydat}{a numeric array with named dimnames containing the EEG (or
other) data. Missing values are not allowed.}

\item{factordef}{a named list of factor definitions, containing the following
elements:
\itemize{
\item{"between"}{"character vector of between-subject factors (default: NULL)"}
\item{"within"}{"character vector of within-subject factors (default: NULL)"}
\item{"w_id"}{"name of the dimension which identifies the subjects
(default: "id")}
}}

\item{bwdat}{a data.frame which contains the identification codes
(factordef$w_id) and the group memberships (factordef$between) of the
subjects. Missing values are not allowed.}

\item{type}{a character value of "tanova" (default), "dissimilarity", or
"gfp"}

\item{verbose}{logical value indicating whether means for each factor level
should be also returned}

\item{nperm}{integer value giving the number of permutations (default: 999L)}

\item{useparallel}{logical value; if TRUE (default), computations are done
in parallel}

\item{ncores}{integer value corresponding to the number of cores;
if NULL (default), it is set to the maximum number of cores available}

\item{par_method}{parallelization method; can be "snow" (default) or "mc"
(multicore)}

\item{cl}{a cluster definition for snow-type parallelization; if NULL
(default), it is set up automatically}

\item{iaterms_last}{logival variable whether all interaction terms should
follow all main effect terms (default: TRUE)}

\item{seed}{an integer value which specifies a seed (default: NULL)}

\item{pcrit}{the significance level (or a vector of multiple levels) for the
consecutive length correction (default: 0.05)}
}
\value{
A list object with effect statistics, uncorrected p-values, and
two types of corrected p-values: length correction or global correction
}
\description{
\code{tanova} performs point-to-point topographical ANOVA on arrays. Related
methods are GFP-analysis (which is based on the intensity of the signal) and
DISS-analysis (focusing only on the global dissimilarity by comparing
normalized topographies). See References for further details.
}
\details{
The function assumes that the input array contains at least two
named dimensions: chan (corresponding to the channels [electrodes]) and time
(corresponding to time points). All dimensions which are not listed as
within-subject factors are treated in a similar way as chan and time, that is
separate TANOVA-s are computed for each level of those dimensions.
}
\note{
The function computes type I p-values - this is correct if the design
is fully balanced and orthogonal (if the number of between-subject
factors is one, it may have slightly unequal group sizes).
}
\examples{
# example dataset
data(erps)
dat_id <- attr(erps, "id") # to get group memberships

# average the data in each 12 ms time-bin to decrease the computational
# burden (not needed in serious analyses)
tempdat <- avgBin(erps, "time", 6)

# Analyze the effect of the reading group (between-subject factor) and the
# two experimental conditions (stimclass, pairtye; within-subject factors)
# for each channel and time sample, taking into account both intensity and
# distributional differences (the default: type = "tanova")
# Note that the number of permutations should be increased for serious
# purposes.
result_tanova <- tanova(tempdat,
                        list(between = "group",
                             within = c("stimclass", "pairtype"),
                             w_id = "id"),
                        bwdat = dat_id,
                        nperm = 499L,
                        useparallel = TRUE, ncores = 2)

# plot results (for now, only p-values)
plotTanova(result_tanova, only_p = TRUE)
}
\references{
Ported from the MATLAB toolbox by Koenig T, Kottlow M, Stein M,
Melie García L (2011) Ragu: A free tool for the analysis of EEG and MEG
event-related scalp field data using global randomization statistics.
Computational Intelligence and Neuroscience, 2011:938925
}
\seealso{
\code{\link{plotTanova}} plots the result of \code{tanova}.
}

