% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tfceFn}
\alias{tfceFn}
\title{Low-level TFCE function which calls C++ code}
\usage{
tfceFn(x, chn, eh, nr_steps = 50L, channel_dim = 1L)
}
\arguments{
\item{x}{numeric matrix or array}

\item{chn}{channel neighbourhood matrix}

\item{eh}{numeric vector of E and H parameters}

\item{nr_steps}{number of threshold steps (default: 50L)}

\item{channel_dim}{the dimension of x which represents channels (default: 1L)}
}
\value{
numeric matrix or array of the same dimensions as x
}
\description{
\code{tfceFn} performs TFCE correction. This function is not
intended for direct use.
}
\keyword{internal}

