% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tfceTtest}
\alias{tfceTtest}
\title{T-test with TFCE correction}
\usage{
tfceTtest(dat, ChN, EH = c(0.66, 2), type = c("o", "i", "d"),
  nperm = 100L, useparallel = TRUE, ncores = NULL,
  par_method = c("snow", "mc"), cl = NULL, groups = NULL,
  use_subset = NULL, seed = NULL, save.random2matlab = FALSE)
}
\arguments{
\item{dat}{numeric array with named dimensions (at least "id", "chan" and
"time")}

\item{ChN}{channel neighbourhood matrix}

\item{EH}{numeric vector of E and H parameters of the TFCE correction
(default: c(0.66, 2))}

\item{type}{character value; "o" for one-sample t-test (default),
"i" for independent samples t-test and "d" for dependent samples t-test}

\item{nperm}{integer value giving the number of permutations (default: 100L)}

\item{useparallel}{logical value; if TRUE (default), computations are done
in parallel}

\item{ncores}{integer value corresponding to the number of cores;
if NULL (default), it is set to the maximum number of cores available}

\item{par_method}{parallelization method; can be "snow" (default) or "mc"
(multicore)}

\item{cl}{a cluster definition for snow-type parallelization; if NULL
(default), it is set up automatically}

\item{groups}{a list of two character vectors corresponding to subject IDs;
used only if independent samples are compared}

\item{use_subset}{a list to be passed to \code{\link{subsetArray}}}

\item{seed}{an integer value which specifies a seed}
}
\value{
A list containing traditional t-values, TFCE-corrected t-values and
p-values
}
\description{
\code{tfceTtest} performs t-test with TFCE correction
}

