% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{theme_ndys}
\alias{theme_ndys}
\title{Helper function to create customized ggplot2 theme}
\usage{
theme_ndys(base_fontsize = 8, legend_direction = c("vertical",
  "horizontal"), panel_aspect_ratio = NULL)
}
\arguments{
\item{base_fontsize}{size of texts}

\item{legend_direction}{the direction of the various levels of a legend key
(default: "vertical")}

\item{panel_aspect_ratio}{the ratio of the y and x axes on the plot}
}
\description{
\code{theme_ndys} modifies the base ggplot2 theme for figures in Leppanen et 
al.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg, colour = factor(cyl))) + 
    geom_point() + facet_grid(~ am) + 
    theme_ndys(12, "h", 1.5)
    
}
