% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{topoCoord}
\alias{topoCoord}
\title{Compute projections for plotting scalp topographies}
\usage{
topoCoord(
  dat,
  peak_df,
  ch_pos,
  size_x,
  size_y,
  shift_y,
  r = 1,
  ampl_range = NULL,
  resol = 50L,
  resolcol = 101L,
  projection = "laea",
  projref = c("pole", "equator"),
  origo = c(lat = ifelse(projref == "pole", 90, 0), long = ifelse(projref == "pole",
    270, 0)),
  ...
)
}
\arguments{
\item{dat}{array of ERP data}

\item{peak_df}{data.frame of peaks (time and potentially other facetting
variables)}

\item{ch_pos}{data.frame of channel positions}

\item{size_x}{the width of the topoplot in time units}

\item{size_y}{the height of the topoplot in amplitude units}

\item{shift_y}{the vertical center of the topoplot in amplitude units}

\item{r}{radius of the channel positions if \code{ch_pos} does not contain 
it (default: 1)}

\item{ampl_range}{numeric vector of length 2 containing the minimum and 
maximum values. If NULL, it is computed from the data. Values outside the
\code{ampl_range} interval are set equal to the corresponding limit.}

\item{resol}{an integer value; the resolution of the grid topographic map in
horizontal and vertical direction (default: 30)}

\item{resolcol}{integer value, the number of levels (resolution) for 
colouring (default: 101L)}

\item{projection}{character value (default = "laea"). See 
\url{http://www.remotesensing.org/geotiff/proj_list/} for common projections.}

\item{projref}{projection reference (pole [default] or equator)}

\item{origo}{a named character vector of lat and long coordinates of the 
origo}

\item{...}{arguments passed to \code{\link{chanInterp}}. You might consider
setting the argument \code{N} for dense electrode caps.}
}
\value{
The function returns a named list of two data.table objects (topo
and boundary).
}
\description{
\code{topoCoord} computes the topographic projection of the amplitudes at 
the given peaks.
}
\note{
This function is called by \code{\link{plotButterfly}} and 
\code{\link{plotMap}}.
}
\keyword{internal}
