% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoplot.R
\name{topoplot}
\alias{topoplot}
\title{Topographic Plot}
\usage{
topoplot(x, sl, res = 200, scale = "auto", plot = c("sensorlocs",
  "legend", "contour", "labels", "all"), ...)
}
\arguments{
\item{x}{Numeric array (coerced). The data to be interpolated and plotted. The number of data points should
equal the number of sensors in the \code{sens} parameter.}

\item{sl}{\code{\link{sensorlocs}} object, or data frame with sensor coordinates. Should at least contain the variables
'label', 'x2d' and 'y2d'.}

\item{res}{Numeric. Resolution of the output grid, Default 200.}

\item{scale}{Plot using this scale. Use "auto" (default) for automatic scaling; otherwise pass a pair of c(min, max),
e.g., c(-40,40).}

\item{plot}{Character. Specifies how the plot will look like. Levels are always plotted, "sensorlocs" (default) plots sensor
locations as dots, "legend" plots a legend at the right side of the plot, "contour" specifies contour lines with values in the plot,
"labels" plots sensor labels, and "all" plots all of the above.}

\item{...}{Additional parameters passed to the 'image' function, e.g. main=}
}
\value{
list with 4 components:
\describe{
  \item{x,y}{vectors of x- and y- coordinates of the output grid, each 'res' values ranging from -200 to + 200.}
  \item{z}{matrix of fitted z-values, dimensions 400 x 400. The fitted value at an sensor location with coordinates
  x2d and y2d can be found at z[(x2d * 360/pi), (y2d * 360/pi)].}
  \item{zlim}{range of z-values used in the interpolation}
}
}
\description{
Draw a topographical plot of EEG data using spherical spline interpolation
}
\details{
The function \code{topoplot} draws a topographical map of EEG data \code{x} recorded at sensors \code{sens}.
Values in between sensor locations are interpolated using spherical splines, as proposed by Perrin et al. (1989,
see also the Corrigendum). The plot may optionally include contour lines, sensor positions, sensor labels, and a
legend. Scaling can be manual or automatic.

The function assumes the internal sensor coordinate system, which originates in the middle of the head (0,0,0). See
\code{\link{sensorlocs}}

Multiple topoplots can also be combined on a single page or display, using appropriate par(mfrow = c()), par(mfcol = c()),
or split.screen() commands. In such cases, you may want to draw a single legend for all graphs. To do so, do not specify plotting
a legend in the constituent plots, but add the legands after plotting, as in the example below.
}
\note{
Spline interpolation is implemented through the \code{sspline} package, which implements the methods proposed by Wahba (1981),
on which the work of Perrin et al. (1989) is also based. The \code{sspline} package works with latitudes and longitudes in degrees
instead of sensor locations as defined internally in the \code{eegr} package. This i s normally hidden from the user calling this
function, and only important for users who want to work with the underlying code.
}
\examples{
\dontrun{
data <- ReadEDF ("yourfile")
sensors <- getLocationsfromLabels(labels = data$head$label)
opar <- par(xpd=NA, mfrow=c(1,2));
ret <- topoplot (data$ctd[200,], sensors, scale = c(-10, 10), plot = c("contour","sensorlocs"), main="Plot1");
ret <- topoplot (data$ctd[400,], sensors, scale = c(-10, 10), plot = c("contour","sensorlocs"), main="Plot2");
usr <- par("usr")
plotrix::color.legend(xl = usr[2], yb = (usr[3] + 23), xr = (usr[2] + 23), yt = (usr[4] - 23)
                      rect.col = rev(rainbow(250,start=0,end=.7)),
                      gradient = "y", align = "rb",
                      legend = sprintf("\%+2.0f", seq(-10, 10, length=5)))
par(opar)
# Note: use appropriate scaling min and max scaling limits
}

}
\references{
Perrin F., Pernier J., Bertrand O., and Echallier JF. (1989). Spherical splines for scalp potential and current
density mapping. \emph{Electroencephalogr. Clin. Neurophysiol.}, \emph{72(2)}, 184-187.
\url{https://doi.org/10.1016/0013-4694(89)90180-6}

Perrin, F., Pernier, J., Bertrand, O. and Echallier, J.F. (1990) Corrigenda: EEG 02274.
\emph{Electroenceph. Clin. Neurophysiol.}, \emph{76}, 565. \url{https://doi.org/10.1016/0013-4694(90)90009-9}

Wahba, G., (1981). Spline interpolation and smoothing on the sphere. \emph{SIAM J. Sci. Stat. Comput.},
\emph{2(1)}, 5-16. \url{https://doi.org/10.1137/0902002}
}
\author{
Geert van Boxtel, \email{G.J.M.vanBoxtel@gmail.com}.
}
