% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Ttest}
\alias{Ttest}
\alias{isTtest}
\alias{osTtest}
\title{Workhorse functions of arrayTest}
\usage{
isTtest(dat, groups, id_dim = "id", mu = 0, var_equal = TRUE,
  verbose = FALSE, has_NA = NULL)

osTtest(dat, id_dim = "id", mu = 0, verbose = FALSE, has_NA = NULL)
}
\arguments{
\item{dat}{numeric matrix or array with named dimension names}

\item{groups}{integer vector with values 1 and 2, corresponding to the group
membership of the subjects (see \code{id_dim})}

\item{id_dim}{character; the dimension of \code{dat} which refers to the
subjects (default: "id")}

\item{mu}{a number indicating the true value of the mean (or difference in
means if one is performing a two sample test)}

\item{var_equal}{a logical variable indicating whether to treat the two
variances as being equal. If TRUE then the pooled variance is used to
estimate the variance otherwise the Welch (or Satterthwaite) approximation
to the degrees of freedom is used.}

\item{verbose}{logical value; if TRUE, degrees of freedom and p-values are
also returned (default: FALSE)}

\item{has_NA}{logical indicating whether there is any missing value in
\code{dat}. If NULL (default), it is explicitly checked by calling
\code{\link{anyNA}} on \code{dat}.}
}
\value{
An array of the same dimensions as \code{dat} except for the
\code{id_dim} dimension. Auxiliary results (degrees of freedom and p-values)
are added as attributes (if requested).
}
\description{
\code{isTtest} and \code{osTtest} compute t-values and additional p-values &
degrees of freedom if requested for independent samples (\code{isTtest}) and
one-sample or paired-samples (\code{osTest}).
}
\keyword{internal}

