% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warp.R
\name{warp}
\alias{warp}
\title{Compute time-warping weights}
\usage{
warp(
  erp,
  align_dim,
  time_dim = "time",
  lambda = 0,
  method = c("mean", "median"),
  smooth_data = FALSE,
  smooth_times = 25L,
  verbose = FALSE,
  time_points = NULL,
  ...
)
}
\arguments{
\item{erp}{a numeric matrix or array of ERPs}

\item{align_dim}{numeric or character index of the dimension which groups
the individual time series. If missing, and 'erp' is a matrix, it is the
dimension which is not 'time_dim'. See Details.}

\item{time_dim}{numeric or character index of the time dimension (default: 
"time")}

\item{lambda}{numeric scalar which controls the elasticity (default: 0; 
higher lambda means less elastic curves)}

\item{method}{warp and calculate to Karcher Mean ("mean", the default) or 
Karcher Median ("median"). The latter option seems buggy; do not use it
in serious analyses.}

\item{smooth_data}{logical value whether to smooth data using box filter 
(default: FALSE)}

\item{smooth_times}{numeric scalar; number of times to apply box filter if 
'smooth_data' is TRUE (default: 25L)}

\item{verbose}{logical value whether iteration messages should be displayed
and plots should be produced}

\item{time_points}{numeric vector of time points at which the ERPs were 
sampled. If NULL (default), it is derived from the dimnames attribute of 
time dimension of the ERP array.}

\item{...}{further arguments passed to \code{\link{fnDims}} (e.g. parallel)}
}
\description{
\code{warp} computes time-warping weights which can be used for 
time-alignment of ERP curves (see \code{\link{align}}). It is a wrapper 
around \code{\link[fdasrvf]{time_warping}}. See its documentation for 
references to the time-warping algorithm.
}
\details{
In the standard case, \code{warp} takes a matrix as input which 
represent multiple time series (where the individual time series are either 
the columns or the rows of the matrix). However, \code{warp} also accepts 
multidimensional array as input which is treated as a collection of 
independent time X align_dim matrices (the order of the dimensions is not
relevant).\cr
\code{warp} returns an array of the same size as \code{erp} consisting
of the warping weights. The weights can be used for time-alignment of the 
same or an other ERP array in \code{\link{align}}.
}
\note{
The defaults of \code{warp} are not necessarily identical to the 
defaults of the underlying \code{\link[fdasrvf]{time_warping}} function and
are subject to change in the future.
}
\seealso{
\code{\link{align}} for time-aligning transformation and examples, 
and \code{\link{fnDims}} for additional parameters
}
