% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_admixture.R
\name{admixture_parser}
\alias{admixture_parser}
\title{Class for ADMIXTURE parsers}
\value{
A dplyr tibble.
}
\description{
An object of class admixture_parser helps you easily parse
results from an ADMIXTURE analysis.
}
\section{Super class}{
\code{eegr::POPGEN Parser} -> \code{Admixture Parser}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{k_min_error}}{Singleton vector holding K of least cross-validation error}

\item{\code{k_values}}{Vector of K-values}

\item{\code{cv_errors}}{Vector of cross-validation errors}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Admixture Parser-new}{\code{admixture_parser$new()}}
\item \href{#method-Admixture Parser-get_assignments}{\code{admixture_parser$get_assignments()}}
\item \href{#method-Admixture Parser-clone}{\code{admixture_parser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Admixture Parser-new"></a>}}
\if{latex}{\out{\hypertarget{method-Admixture Parser-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{admixture_parser$new(admixture_dir, sample_ids, popmap_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{admixture_dir}}{Character singleton. Directory of Admixture output.}

\item{\code{sample_ids}}{Character vector. IDs of samples input to Admixture.}

\item{\code{popmap_path}}{Path to .tsv with headerless columns for ID and pop}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Admixture Parser-get_assignments"></a>}}
\if{latex}{\out{\hypertarget{method-Admixture Parser-get_assignments}{}}}
\subsection{Method \code{get_assignments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{admixture_parser$get_assignments(k = self$k_min_error)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{k}}{Integer singleton. Value of K.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Admixture Parser-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Admixture Parser-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{admixture_parser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
