% Generated by roxygen2 (4.0.1): do not edit by hand
\name{array2df}
\alias{array2df}
\title{Transforms an array to a data.frame}
\usage{
array2df(dat, response_name = "values", dim_types = "character",
  na_omit = FALSE, ...)
}
\arguments{
\item{dat}{a matrix or array with named dimnames}

\item{response_name}{the name of the variable in the data.frame which
contains the values of the input (default: "values")}

\item{dim_types}{a character value or a named list which specifies
the type of the variable in the data.frame corresponding to the dimension of
the input array. If a character value is given, all variables are
transformed to that type (default: "character"). Possible types are
"logical", "character", "numeric", "double", "integer", "factor".}

\item{na_omit}{if TRUE, omit all rows from the data.frame which have missing
values (default: FALSE)}

\item{...}{additional parameters to be passed to \code{\link{as.data.frame.table}}}
}
\value{
A data.frame
}
\description{
\code{array2df} transforms an array with dimnames to a data.frame
}

