% Generated by roxygen2 (4.0.1): do not edit by hand
\name{arrayAnova}
\alias{arrayAnova}
\title{Perform ANOVA (potentially with TFCE correction) on arrays}
\usage{
arrayAnova(arraydat, factordef, bwdat = NULL, verbose = TRUE,
  nperm = 999L, useparallel = FALSE, ncpus = NULL, par_method = "snow",
  cl = NULL, usetfce = TRUE, tfce_options = NULL, seed = NULL)
}
\arguments{
\item{arraydat}{a numeric array with named dimnames containing the EEG (or
other) data. Missing values are not allowed.}

\item{factordef}{a named list of factor definitions, containing the following
elements:
\itemize{
\item{"between"}{"character vector of between-subject factors (default: NULL)"}
\item{"within"}{"character vector of within-subject factors (default: NULL)"}
\item{"w_id"}{"name of the dimension which identifies the subjects
(default: "id")}
}}

\item{bwdat}{a data.frame which contains the identification codes
(factordef$w_id) and the group memberships (factordef$between) of the
subjects. Missing values are not allowed.}

\item{verbose}{logical value indicating if p-values and effect sizes should
be computed for the traditional ANOVA results}

\item{nperm}{integer value giving the number of permutations (default: 999L)}

\item{useparallel}{logical value; if TRUE (default), computations are done
in parallel}

\item{ncpus}{integer value corresponding to the number of cores;
if NULL (default), it is set to the maximum number of cores available}

\item{par_method}{parallelization method; can be "snow" (default) or "mc"
(multicore)}

\item{cl}{a cluster definition for snow-type parallelization; if NULL
(default), it is set up automatically}

\item{usetfce}{logical value whether TFCE (threshold-free cluster enhancement)
correction should also be computed (default: TRUE)}

\item{tfce_options}{a named list containing the channel neighbourhood matrix
(named ChN) and the vector of the E/H parameters (named EH)}

\item{seed}{an integer value which specifies a seed (default: NULL)}
}
\value{
A list object with F values, TFCE-corrected F-values and
permutation-based p-values (if requested)
}
\description{
\code{arrayAnova} performs point-to-point ANOVAs on arrays. Permutation-based
p-values and Threshold-free Cluster Enhancement (TFCE) correction can be requested.
}
\details{
The function assumes that the input array contains at least two
named dimensions: chan (corresponding to the channels [electrodes]) and time
(corresponding to time points). All dimensions which are not listed as
within-subject factors are treated in a similar way as chan and time, that is
separate ANOVA-s are computed for each level of those dimensions.
}
\note{
The function computes type I p-values - this is correct if the design
is fully balanced and orthogonal (if the number of between-subject
factors is one, it may have slightly unequal group sizes).
}

