% Generated by roxygen2 (4.0.1): do not edit by hand
\name{arrayAnovaSub}
\alias{arrayAnovaSub}
\title{Workhorse function of arrayAnova}
\usage{
arrayAnovaSub(a_dat, f_def, d_names, f_dat, verbose = TRUE)
}
\description{
\code{arrayAnovaSub} computes F-values in ANOVA designs, and provides
additional p-values, degrees of freedom, generalized effect sizes if
requested.
}

