% Generated by roxygen2 (4.0.1): do not edit by hand
\name{arrayTtest}
\alias{arrayTtest}
\title{Generalization of \code{t.test} to arrays}
\usage{
arrayTtest(dat1, dat2 = NULL, id_dim = "id", paired = FALSE, mu = 0,
  var_equal = TRUE, verbose = TRUE)
}
\arguments{
\item{dat1}{matrix or array with named dimnames; if dat2 is NULL (default),
the function computes one-sample t-tests.}

\item{dat2}{matrix or array with named dimnames; if dat2 is given and
paired is TRUE, dat2 must be of equal size as dat1}

\item{id_dim}{the name of dimension with identifies observational units
(usually participants)}

\item{paired}{logical variable, only used if dat2 is provided; if paired is FALSE
(default), the function computes independent samples t-tests, otherwise
paired samples t-tests are performed}

\item{mu}{a number indicating the true value of the mean (or difference in
means, if paired tests are performed)}

\item{var_equal}{a logical variable whether the variances are equal. If TRUE
(default), the pooled variance is used to estimate the variance otherwise
the Welch (or Satterthwaite) approximation to the degrees of freedom is used.}
}
\value{
A named list of two vectors/matrices/arrays containing t- and
p-values, and the corresponding degrees of freedom. The dimension of the
t- and p-values corresponds to the dimension of the input data, apart from
the id_dim dimension, which is dropped.
}
\description{
\code{arrayTtest} generalizes \code{t.test} to arrays
}

