% Generated by roxygen2 (4.0.1): do not edit by hand
\name{avgBin}
\alias{avgBin}
\title{Compute averages of bins on a dimension of an array}
\usage{
avgBin(dat, target_dim, bin_length = NULL, bin_ind = NULL,
  rolling = FALSE, newnames = "avg")
}
\arguments{
\item{dat}{matrix or array}

\item{target_dim}{name of binned dimension of dat}

\item{bin_length}{number of data points in one bin. If rolling parameter is
FALSE (default), bin_length must match the length of the target dimension}

\item{bin_ind}{an abbreviation for bin indices; a numeric or character vector
or a factor which provides the bin membership for each data point of
target_dim. If bin_ind is provided, bin_length is not used.}

\item{rolling}{logical variable; if yes, avgBin computes rolling (or moving)
window averages (the size of the moving window is determined by bin_length)}

\item{newnames}{one of "avgs" (default), NULL, or a character vector defining
the new dimension names of the binned dimension. If "avgs", the original
names of target_dim are averaged per each bin. Not relevant if rolling is
TRUE (which reuses the original labels).}
}
\value{
A matrix or array, depending on the input
}
\description{
\code{avgBin} computes averages of bins on a dimension of an array. The
size of bins or directly the bins themselves can be specified by the user.
}

