% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erp_preprocess.R
\name{avgTrials}
\alias{avgTrials}
\title{Average single-trials}
\usage{
avgTrials(dat, markers, which_factors = NULL)
}
\arguments{
\item{dat}{numeric array, the segmented ERPs}

\item{markers}{data.frame containing marker definitions}

\item{which_factors}{numeric or character vector indicating which columns
from markers should be used for averaging. If NULL (default), all columns
are used.}
}
\value{
numeric array
}
\description{
\code{avgTrials} performs single-trial averaging
}
