% Generated by roxygen2 (4.0.1): do not edit by hand
\name{chanInterp}
\alias{chanInterp}
\title{Spherical spline interpolation}
\usage{
chanInterp(dat, ch_pos, interp_pos = NULL, maxNA = 0.3, m = 4L, N = 7L,
  lambda = 1e-10, type = c("voltage", "laplacian", "scd"),
  alarm_tolerance = 0.01)
}
\arguments{
\item{dat}{a numeric vector, matrix, data.frame or array with named dimnames}

\item{ch_pos}{channel (electrode) positions; should be a matrix or data.frame
with the following column names (order is not important): "theta", "phi"
(spherical coordinates), or "x", "y", "z" (cartesian coordinates)}

\item{interp_pos}{interpolation positions, if they differ from ch_pos;
should be in the same format as ch_pos (default: NULL)}

\item{maxNA}{numeric value given as ratio (0 < maxNA < 1) or integer
(maxNA >= 1 ); it determines the maximum ratio or number of non-missing
channels in a time sample. If exceeded, no interpolation occurs in the given
time sample.}

\item{m}{integer value (default: 4); the flexibility of the spline}

\item{N}{integer value (default: 7); number of terms in the Legendre
polynomial}

\item{lambda}{numeric value (default: 1e-10); smoothing factor}

\item{type}{character value, do not use yet}

\item{alarm_tolerance}{numeric value (default: 1e-2); if the maximal absolute
interpolation error at any time sample exceeds this limit, a message is
shown. If set to NULL, no check is performed.}
}
\value{
An object having the same attributes as dat
}
\description{
\code{chanInterp} performs spherical spline interpolation. It can impute
bad channels (i.e. channels with missing values) or interpolate to
artbitrary positions on the scalp.
}

