% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{colorize}
\alias{colorize}
\title{Make image plots more colorful (only for ggplot objects)}
\usage{
colorize(
  obj,
  low = scales::muted("blue"),
  mid = "white",
  high = scales::muted("red"),
  ...
)
}
\arguments{
\item{obj}{a ggplot object created by \code{\link{imageValues}}}

\item{low, mid, high}{colors to produce a low < mid < high color scale}

\item{...}{additional arguments to be passed to \code{link[ggplot2]{scale_fill_gradient2}}}
}
\value{
a ggplot object
}
\description{
\code{colorize} changes the color scale of \code{\link{imageValues}} plots
}
