% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compGfp}
\alias{compGfp}
\title{Compute Global Field Power}
\usage{
compGfp(dat, keep_channels = FALSE)
}
\arguments{
\item{dat}{numeric matrix or array with named dimensions (one of which
must be "chan")}

\item{keep_channels}{logical value; if TRUE, the original channels are
retained, and the GFP values are added with a channel code of "GFP"
(default=FALSE)}
}
\value{
The function returns a matrix or an array. Note that if
the original channels are not retained, the channel dimension is dropped.
}
\description{
\code{compGfp} computes Global Field Power (the standard deviation of
channel values for each sampling point)
}

