% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{compTanovaEffect}
\alias{compTanovaEffect}
\title{Compute generalized dissimilarity}
\usage{
compTanovaEffect(obj, new_indices = NULL, attribs = FALSE)
}
\arguments{
\item{obj}{a list object as returned by \code{\link{preAnova}}}

\item{new_indices}{if not NULL (default), it must be a vector of numeric
indices to be passed to the model matrix. This parameter is used for
permutation-based analyses.}

\item{attribs}{a logical value if dimension attributes should be attached to
the resulting array (default: FALSE)}
}
\value{
This function returns a numeric matrix of effects if 'attribs' is
FALSE, and an array of effects if 'attribs' is TRUE.
}
\description{
\code{compTanovaEffect} computes the generalized effect size measure as
given in Koenig & Melie-Garcia, 2009, p177. This is an internal function
called by the high-level function \code{\link{tanova}}.
}
\keyword{internal}
