% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anova.R
\name{compZTanova}
\alias{compZTanova}
\title{Compute effect size in TANOVA}
\usage{
compZTanova(effect_perm)
}
\arguments{
\item{effect_perm}{an array of permuted effects (including the observed
effect as the first level of the 'perm' dimension)}
}
\description{
\code{compZTanova} is a helper function to compute corrected test statistic
in the high-level \code{\link{tanova}} function. It is actually the z-scored
generalized GFP statistic.
}
\keyword{internal}

