% Generated by roxygen2 (4.0.2): do not edit by hand
\name{decorateDims}
\alias{decorateDims}
\title{Add dimension names to a matrix or array with (partially) missing dimension
names}
\usage{
decorateDims(dat, names = TRUE, new_names = NULL, new_dimnames = NULL,
  in_place = FALSE)
}
\arguments{
\item{dat}{a matrix or array}

\item{names}{logical; if TRUE (default), named list of dimnames are returned.
Ignored if new_names is provided.}

\item{new_names}{character vector for named dimnames}

\item{new_dimnames}{list of new dimension names. If new_names is not
provided but new_dimnames is a named list, its names are considered as
new_names}

\item{in_place}{logical; if TRUE, dat is modified in place (without copy)
and it is returned only invisibly (default: FALSE). Use in_place with extra
care because it modifies in place all objects which dat refers to.}
}
\value{
The original matrix or array with non-null dimension names
}
\description{
\code{decorateDims} is not unlike \code{\link{provideDimnames}}. The main
difference is that \code{decorateDims} can modify the input data in place and
it has a more convenient interface for providing names of dimensions
}
\seealso{
\code{\link{provideDimnames}} for a slightly different solution
}

