% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{eegR-package}
\alias{eegR-package}
\alias{eegR}
\title{A package to analyze EEG signals}
\description{
The eegR package has been developed to process
electroencephalography (EEG) signals. This is only a pre-alpha version
which has basic support for pre-processing EEG signals and provides
some more advanced tools to plot and analyze averaged event related
potentials (ERPs).
}
\author{
Denes Toth <\email{toth.denes@ttk.mta.hu}>
}
