% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{eegR-package}
\alias{eegR}
\alias{eegR-package}
\title{A package to analyze EEG signals}
\description{
The package 'eegR' has been developed to process electroencephalography (EEG)
signals. Beyond common EEG signal processing functionalities, eegR provides 
advanced tools to analyze single-trial and averaged event related potentials 
(ERPs).
IMPORTANT NOTE: This is only a pre-alpha version which will be restructured
substantially in short term. Do not rely on any of the functions.
}
\author{
Denes Toth <\email{toth.denes@ttk.mta.hu}>
}

