\name{eegr-package}
\alias{eegr-package}
\alias{eegr}
\docType{package}
\title{
\packageTitle{eegr}
}
\description{
\packageDescription{eegr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{eegr}
\packageIndices{eegr}
Utilites for EEG analysis.  Specifically, utilities that allow read headers, events, and timeseries data from BDF files from BIOSEMI EEG systems.
}
\author{
\packageAuthor{eegr}

Maintainer: \packageMaintainer{eegr}
}
\references{
<http://www.biosemi.com/faq/file_format.htm>
}