% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print_summary.R
\name{extract}
\alias{extract}
\alias{extract.arrayAnova}
\alias{extract.arrayTtest}
\alias{extract.tanova}
\title{Extract slots from arrayAnova, arrayTtest, or tanova object}
\usage{
extract(...)

\method{extract}{arrayTtest}(object, what = c("all", "stat", "p", "es",
  "stat_corr", "p_corr", "means", "df"), time_window = NULL, term = NULL,
  chan = NULL, drop = FALSE, ...)

\method{extract}{arrayAnova}(object, what = c("all", "stat", "p", "es",
  "stat_corr", "p_corr", "means", "df"), time_window = NULL, term = NULL,
  chan = NULL, drop = FALSE, ...)

\method{extract}{tanova}(object, what = c("all", "stat", "p", "stat_corr",
  "p_corr", "means"), time_window = NULL, term = NULL, drop = FALSE, ...)
}
\arguments{
\item{...}{not used yet}

\item{object}{an object of class \code{arrayTtest}, \code{arrayAnova}, or
\code{tanova}}

\item{what}{a character vector indicating which slot(s) to extract. If
'what' is 'all' (default), all relevant slots is extracted (see the Usage
section for the relevant slot names in each particular method). Otherwise,
the slots to extract are found by partial matching to the following choices:
\itemize{
\item{stat: }{Test statistic (e.g. F values in \code{arrayAnova})}
\item{stat_corr: }{Corrected test statistic (e.g. TFCE values in
\code{arrayAnova})}
\item{p: }{Uncorrected P-values}
\item{p_corr: }{Corrected P-values (e.g. TFCE-corrected P-values in
\code{arrayAnova}, or consecutive length correction in \code{tanova})}
\item{es: }{Effect size statistic (e.g. Generalized Eta Squared in
\code{arrayAnova})}
\item{means: }{Marginal means for each model term} (TODO! Temporarily removed...)
}}

\item{time_window}{numeric vector of length two indicating the time window
which should be extracted. If NULL (default), all time points are returned.}

\item{term}{character vector indicating the name of the model terms which
should be extracted. If NULL (default), all model terms are returned.}

\item{chan}{character or integer indices of channels to extract. If NULL
(default), all channels are returned.}

\item{drop}{logical value; if TRUE, singleton dimensions are dropped from
the arrays after subsetting on time, term, and/or chan (default: FALSE)}
}
\value{
The function returns an array of the requested component or a named
list of arrays if multiple components were requested.
}
\description{
\code{extract} is a generic function for extracting slots from
\code{\link{arrayTtest}}, \code{\link{arrayAnova}}, or \code{\link{tanova}}
objects, each produced by the corresponding function.
}
\section{Methods (by class)}{
\itemize{
\item \code{arrayTtest}: 

\item \code{arrayAnova}: 

\item \code{tanova}: 
}}
\seealso{
See the examples of \code{\link{arrayAnova}}
}

