% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matstat_utils.R
\name{fastUnique}
\alias{fastUnique}
\title{Fast version of unique for matrices}
\usage{
fastUnique(x, margin = 1L, freq = FALSE)
}
\arguments{
\item{x}{matrix}

\item{margin}{1 (default) or 2, referring to rows and columns of the 
matrix, respectively. Can be a character string which is interpreted
as the name of the dimension if \code{x} has named dimnames.}

\item{freq}{logical value if the frequency of the given rows or columns 
should also be returned (default: FALSE)}
}
\value{
\code{fastUnique} returns a matrix without duplicated units. If 
requested, the matrix has a \code{freq} attribute containing the frequency
of the given rows or columns.
}
\description{
\code{fastUnique} finds the unique rows or columns of a matrix. It is 
much faster and also more memory-efficient than 
\code{\link{unique.matrix}}, especially if the input matrix is not a 
character matrix.
}
\examples{
# create a matrix of random integer values ranging from 1 to 3
x <- matrix_(sample(1:3, 4e5, TRUE), 1e5, 4)

# compare computation times
system.time(un_x <- unique(x))
system.time(un_x_fast <- fastUnique(x))

# the same if x has dimension names
decorateDims_(x)
system.time(un_x <- unique(x))
system.time(un_x_fast <- fastUnique(x))

# the results are identical
stopifnot(identical(un_x, un_x_fast))

# fastUnique can also return frequency counts
system.time(un_x_withfreq <- fastUnique(x, freq = TRUE))

# check that frequencies add up to the number of rows of x
stopifnot(sum(attr(un_x_withfreq, "freq")) == nrow(x))
}
