% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/array_transform.R
\name{fillMissingDimnames}
\alias{fillMissingDimnames}
\title{Find and fill missing dimension names and identifiers}
\usage{
fillMissingDimnames(dimn, .dim, names = TRUE, new_names = NULL,
  new_dimnames = NULL)
}
\arguments{
\item{dimn}{a list of the original dimension names}

\item{.dim}{an integer vector of dimensions}

\item{names}{logical; if TRUE (default), named list of dimnames are returned.
Ignored (or more precisely, considered TRUE) if new_names is provided.}

\item{new_names}{character vector for named dimnames}

\item{new_dimnames}{list of new dimension names. If 'new_names' is not
provided but 'new_dimnames' is a named list, its names are considered as
'new_names'.}
}
\description{
\code{fillMissingDimnames} is the workhorse function for
\code{\link{decorateDims}} and \code{\link{decorateDims_}}. It replaces
missing dimension names and identifiers and returns the modified 'dimnames'
attribute.
}
\keyword{internal}

