% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{imagePvalues}
\alias{imagePvalues}
\title{Image plot of p-values}
\usage{
imagePvalues(pvalues, pcrit = c(0.001, 0.01, 0.05), grid = NULL,
  wrap = NULL, time_label = "Time (ms)", channel_label = "Channels",
  raster = TRUE, cluster_order = FALSE)
}
\arguments{
\item{pvalues}{numeric matrix or array of p-values with named dimensions (at
least chan and time)}

\item{pcrit}{numeric vector of significancy limits 
(default: 0.001, 0.01, 0.05)}

\item{grid}{character vector or formula defining the layout of panels}

\item{wrap}{character vector or formula defining the dimension which 
separates panels (only considered if grid is NULL)}

\item{time_label}{character string; the label of the x (time) axis 
(default: "Time (ms)")}

\item{channel_label}{character string; the label of the y (channel) axis 
default: "Channels")}

\item{raster}{use raster image (TRUE, default) or not (FALSE)}

\item{cluster_order}{logical; if TRUE, channels are ordered with hierarchical
agglomerative clustering (default: FALSE)}
}
\value{
A ggplot object
}
\description{
\code{imagePvalues} creates an image plot from a matrix or array of p-values
}
\seealso{
\code{\link{imageValues}} for plotting effects or raw amplitudes
}

