% Generated by roxygen2 (4.0.2): do not edit by hand
\name{imageValues}
\alias{imageValues}
\title{Image plot of channel values}
\usage{
imageValues(dat, grid = NULL, wrap = NULL)
}
\arguments{
\item{dat}{numeric matrix or array of values with named dimensions (at least
chan and time)}

\item{grid}{character vector or formula defining the layout of panels}

\item{wrap}{character vector or formula defining the dimension which
separates panels (only considered if grid is NULL)}
}
\value{
A ggplot object
}
\description{
\code{imageValues} creates an image plot from a matrix or array of channel x
time values
}

