% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mergeDims}
\alias{mergeDims}
\title{Reshape array by merging specific dimensions}
\usage{
mergeDims(dat, dims, sep = ".", return_dimattributes = TRUE)
}
\arguments{
\item{dat}{array to reshape}

\item{dims}{list of dimension names or dimension indices to merge. Can be a vector if
only one group of dimensions should be merged. If any of the list elements is a
character vector, dat must have named dimnames.}

\item{sep}{character which separates dimension names after merging (default ".")}

\item{return_attributes}{logical value (default: TRUE) whether original
attributes should be appended to the resulting array}
}
\value{
Array with merged dimensions
}
\description{
\code{mergeDims} reshapes an array by merging user-specified dimensions
}

