% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is.R
\name{negate}
\alias{!}
\alias{!.IsFunction}
\alias{negate}
\title{Negation for the is* family of functions}
\usage{
\method{!}{IsFunction}(fn)
}
\arguments{
\item{fn}{a function of class IsFunction}
}
\description{
\code{`!.IsFunction`} implements logical negation for \code{is*} function.
The same effect can be achieved be setting the 'negate.' argument of the
corresponding function.
}
\keyword{internal}

