% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_plink_pca.R
\name{parse_plink_pca}
\alias{parse_plink_pca}
\title{Instantiate new PLINK PCA parser}
\usage{
parse_plink_pca(eigenval_path, eigenvec_path, id_name = "ID")
}
\arguments{
\item{eigenval_path}{Path to PLINK output with .eigenval extension}

\item{eigenvec_path}{Path to PLINK output with .eigenvec extension}

\item{id_name}{Name to use for sample ID column in output}
}
\value{
An object of class plink_pca_parser
}
\description{
Instantiate new PLINK PCA parser
}
