% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot2dview}
\alias{plot2dview}
\title{Plot the topography of the signals}
\usage{
plot2dview(dat, ch_pos, r = 1, timepoint = NULL, ampl_range = c(-5, 5),
  resol = 100L, resolcol = 1000L, projection = "laea",
  projref = c("pole", "equator"), origo = c(lat = ifelse(projref == "pole",
  90, 0), long = ifelse(projref == "pole", 270, 0)), plot_centroid = TRUE,
  centroid_circle = 0.5, centroid_unitsphere = FALSE, plot_bar = TRUE,
  plot_ch = TRUE, plot_chnames = TRUE, title = NULL, gfp = NULL,
  gfp_max = NULL, ...)
}
\arguments{
\item{dat}{a numeric vector or matrix. If a matrix is provided, it must
contain the participants in rows and the channels in columns}

\item{ch_pos}{a data.frame with the channel coordinates}

\item{r}{a numeric value; the radius of the head}

\item{timepoint}{an integer value}

\item{ampl_range}{numeric vector, the minimum and maximum value of the
amplitudes (default: c(-5, 5))}

\item{resol}{integer value, the resolution of the projection (default: 100L)}

\item{resolcol}{integer value, the resolution of the colours (default: 1000L)}

\item{projection}{character value (default="laea"). See
\link{http://www.remotesensing.org/geotiff/proj_list/} for common projections}

\item{projref}{projection reference (pole [=default] or equator)}

\item{gfp}{a numeric vector of the GFP values in the whole time window}
}
\description{
\code{plot2dview} creates 2D topoplots
}

