% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotTanova}
\alias{plotTanova}
\title{Plot TANOVA results}
\usage{
plotTanova(results, grid = NULL, wrap = NULL, plot_title = "",
  only_p = FALSE)
}
\arguments{
\item{results}{a list; the return value of the \code{\link{tanova}} function}

\item{grid}{character vector or formula defining the layout of panels}

\item{wrap}{character vector or formula defining the dimension which
separates panels (only considered if grid is NULL)}

\item{plot_title}{character string; the title of the plot}

\item{only_p}{logical; if TRUE, p-values are plotted instead of combined
(effect + p-value) plots (default: FALSE). See note.}

\item{time_label}{character string; the label of the x (time) axis
(default: "Time (ms)")}
}
\value{
A ggplot object
}
\description{
\code{plotTanova} plots the result of the \code{\link{tanova}} function
}
\note{
Use only_p = TRUE if you want to check the uncorrected p-values. If
only_p is set to FALSE (which is the default), \code{plotTanova} highlights
the significant phases of the effect curves based on the corrected p-values.
}

