% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{preAnova}
\alias{preAnova}
\title{Parameter checks and data preparation in \code{arrayAnova}}
\usage{
preAnova(.arraydat, factordef, bwdat, verbose, tfce, perm,
  tanova_type = NULL)
}
\arguments{
\item{.arraydat}{numeric array}

\item{factordef}{a list of between-subject factors (named 'between'),
within-subject factors ('within'), subject identifier ('w_id'), and a
character vector of observed (i.e., not manipulated) variables ('observed').
The last element ('observed') is only used in effect size computations.}

\item{bwdat}{a data.frame with the subject-level variables (usually factors)}

\item{tfce}{logical value whether TFCE is requested}

\item{perm}{logical value whether permutation is requested}

\item{tanova_type}{type of tanova ("tanova", "dissimilarity", "gfp") if this
function is called from \code{tanova}}
}
\description{
\code{preAnova} performs parameter checking, creates a data.frame for
the ANOVA design and reshapes .arraydat accordingly, prepares the computation
of sum-of-squares. It is called by \code{\link{arrayAnova}} and
\code{\link{tanova}}.
}
\seealso{
\code{\link{arrayAnova}} and \code{\link{tanova}} are the
higher-level function exported to the user, \code{\link{preSumSq}},
\code{\link{compF}} and \code{\link{compTanovaEffect}} are internal
functions relying on the returned object of \code{preAnova}.
}
\keyword{internal}
