% Generated by roxygen2 (4.0.2): do not edit by hand
\name{preAnova}
\alias{preAnova}
\title{Parameter checks and data preparation before arrayAnovaSub}
\usage{
preAnova(arraydat, factordef, bwdat, useparallel, ncores, par_method,
  usetfce = FALSE)
}
\description{
Parameter checks and data preparation before arrayAnovaSub
}
\keyword{internal}

