% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{pvalueConsec}
\alias{pvalueConsec}
\title{Correct p-values in time-series}
\usage{
pvalueConsec(dat, sig_level = 0.05, min_length = 10)
}
\arguments{
\item{dat}{a numeric matrix or array containing p-values, or a list of such
matrices/arrays. dat must have a named time dimension.}

\item{sig_level}{numeric value (default: 0.05), the level of significance}

\item{min_length}{numeric value (default: 10), the minimum number of
consecutive significant time points}
}
\value{
An object of the same dimension as the input data
}
\description{
\code{pvalueConsec} sets a minimal consecutive length criterion on the
p-values in a time series
}
