% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rollFun}
\alias{rollFun}
\title{Compute rolling (a.k.a. moving) window statistics}
\usage{
rollFun(dat, width, FUN, force_rollapply = FALSE, ...)
}
\arguments{
\item{dat}{a numeric vector, matrix or data.frame. In the latter cases
rolling statistics are computed column-wise.}

\item{width}{width of moving window; can be an integer value or vector.}

\item{FUN}{the function to be applied to compute moving window statistics.
See details.}

\item{force_rollapply}{logical variable; if yes, zoo:::rollapply is called
(default = FALSE).}

\item{...}{optional arguments to the corresponding function in caTools or
zoo:::rollapply}
}
\value{
An object having the same attributes as dat.
}
\description{
\code{rollFun} computes rolling window statistics on vectors or matrices.
}
\details{
If FUN is one of min, max, mean, sd, mad, quantile (OR "min", "max",
"mean", etc.) rollFun calls the corresponding function from the caTools
package (e.g. caTools:::runmin). Otherwise, or if force_rollapply is TRUE,
zoo:::rollapply is called.
}

