% Generated by roxygen2 (4.0.2): do not edit by hand
\name{splitArray}
\alias{splitArray}
\title{Splits an array along a given dimension}
\usage{
splitArray(dat, whichdim)
}
\arguments{
\item{dat}{numeric array (preferably with named dimnames)}

\item{whichdim}{numeric or character vector, the dimension of the array
to split along}
}
\value{
A list of subsets of the original data matrix/array
}
\description{
\code{splitArray} splits an array along a given dimension into a list of
sub-arrays
}

