% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/array_transform.R
\name{subsetArray}
\alias{subsetArray}
\alias{subsetArray<-}
\title{Extract or replace a part of an array}
\usage{
subsetArray(dat, subsets, which_dims = NULL, drop = NULL, keep_attributes = TRUE)

subsetArray(dat, subsets, which_dims = NULL) <- value
}
\arguments{
\item{dat}{array to be subsetted}

\item{subsets}{a (named) list of character, numeric, or logical vectors
indicating which levels of which dimensions to subset (see Details). If
'subsets' is an unnamed list, the argument 'which_dims' must be provided.}

\item{which_dims}{numeric or character indices of the dimensions which
should be subsetted. If 'which_dims' is not NULL, 'which_dims' is used and
the names of 'subsets' is ignored.}

\item{drop}{either 1) NULL (the default), or 2) a logical value
(TRUE or FALSE), or 3) numeric or character indices of the dimensions which
should be dropped if they become singleton dimensions (i.e. have only one
level) after subsetting (see Details)}

\item{keep_attributes}{a logical variable which determines if the result
inherits the custom attributes of the input (TRUE, default) or not}

\item{value}{a vector, matrix or array of the new values}
}
\value{
The function returns a subset of the array or the array with replaced
values.
}
\description{
\code{subsetArray} is a convenience function for extracting or replacing a
part of an array which has dimension names
}
\details{
Names of 'subsets' or the indices of dimensions as given in
'which_dim' indicate which dimensions are to be subsetted in
the input array, and each list element indicates which levels of the given
dimension will be selected. If a list element is an empty vector, all levels
of the correspondig dimension will be selected.\cr
The argument 'drop' defines the procedure if a dimension becomes a singleton
dimension after subsetting. The default behaviour (\code{drop = NULL}) is to
drop all subsetting dimensions but no others. If 'drop' is FALSE, all
dimensions are kept, if TRUE, all singleton dimensions are dropped. If 'drop'
is a numeric or character vector, its elements define which dimensions to
drop.
}
\examples{
# example data (see ?erps)
data(erps)
str(erps)

# subsetting without knowing the exact order of dimensions
sub1 <- subsetArray(erps,
                    list(time = "0", chan = "Fp2",
                         stimclass = c(TRUE, FALSE, FALSE)),
                    keep_attributes = FALSE)

# traditional subsetting
sub2 <- erps["A", , "Fp2", "0", ]

# the results are identical
stopifnot(identical(sub1, sub2))

# the same for replacement
subsetArray(sub1, list(id = 1)) <- NA
sub2[, 1] <- NA
stopifnot(identical(sub1, sub2))
}
\seealso{
\code{\link[abind]{asub}}
}

