% Generated by roxygen2 (4.0.2): do not edit by hand
\name{subsetArray}
\alias{subsetArray}
\alias{subsetArray<-}
\title{Extract or replace a part of an array}
\usage{
subsetArray(dat, subsets, keep_attr = TRUE, ...)

subsetArray(dat, subsets) <- value
}
\arguments{
\item{dat}{array to be subsetted}

\item{subsets}{a named list of character, numeric, or logical vectors
indicating which levels of which dimensions to subset (see Details)}

\item{keep_attr}{a logical variable which determines if the result inherits
the custom attributes of the input (TRUE, default) or not}

\item{...}{further arguments to be passed to asub}
}
\value{
A subset of an array or the array with replaced values
}
\description{
\code{subsetArray} is a convenience function for extracting or replacing a
part of an array which has dimension names
}
\details{
Names of subsets indicate which dimensions are to be subsetted in
the input array, and each list element indicates which levels of the given
dimension will be selected. If a list element is a named empty vector, or
the name of a dimension does not appear in subsets, all levels of the
correspondig dimension will be selected.
}

