% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tanova}
\alias{tanova}
\title{Topographical ANOVA (TANOVA) and related methods}
\usage{
tanova(arraydat, factordef, bwdat = NULL, type = c("tanova",
  "dissimilarity", "gfp"), verbose = TRUE, nperm = 999L,
  useparallel = FALSE, ncpus = NULL, par_method = c("snow", "mc"),
  cl = NULL, iaterms_last = TRUE, seed = NULL, pcrit = 0.05)
}
\arguments{
\item{arraydat}{a numeric array with named dimnames containing the EEG (or
other) data. Missing values are not allowed.}

\item{factordef}{a named list of factor definitions, containing the following
elements:
\itemize{
\item{"between"}{"character vector of between-subject factors (default: NULL)"}
\item{"within"}{"character vector of within-subject factors (default: NULL)"}
\item{"w_id"}{"name of the dimension which identifies the subjects
(default: "id")}
}}

\item{bwdat}{a data.frame which contains the identification codes
(factordef$w_id) and the group memberships (factordef$between) of the
subjects. Missing values are not allowed.}

\item{type}{a character value of "tanova" (default), "dissimilarity", or
"gfp"}

\item{verbose}{logical value indicating whether means for each factor level
should be also returned}

\item{nperm}{integer value giving the number of permutations (default: 999L)}

\item{useparallel}{logical value; if TRUE (default), computations are done
in parallel}

\item{ncpus}{integer value corresponding to the number of cores;
if NULL (default), it is set to the maximum number of cores available}

\item{par_method}{parallelization method; can be "snow" (default) or "mc"
(multicore)}

\item{cl}{a cluster definition for snow-type parallelization; if NULL
(default), it is set up automatically}

\item{iaterms_last}{logival variable whether all interaction terms should
follow all main effect terms (default: TRUE)}

\item{seed}{an integer value which specifies a seed (default: NULL)}

\item{pcrit}{the significance level for duration and global count p-value
correction (default: 0.05)}
}
\value{
A list object with effect statistics, uncorrected p-values, and
two types of corrected p-values: length correction or global correction
}
\description{
\code{tanova} performs point-to-point topographical ANOVA on arrays. Related
methods are GFP-analysis (which is based on the intensity of the signal) and
DISS-analysis (focusing only on the global dissimilarity by comparing
normalized topographies). See References for further details.
}
\details{
The function assumes that the input array contains at least two
named dimensions: chan (corresponding to the channels [electrodes]) and time
(corresponding to time points). All dimensions which are not listed as
within-subject factors are treated in a similar way as chan and time, that is
separate TANOVA-s are computed for each level of those dimensions.
}
\note{
The function computes type I p-values - this is correct if the design
is fully balanced and orthogonal (if the number of between-subject
factors is one, it may have slightly unequal group sizes).
}
\references{
Koenig (2011) RAGU
}

