% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tfce.fnc}
\alias{tfce.fnc}
\title{Low-level TFCE function which calls C code}
\usage{
tfce.fnc(x, chn, eh)
}
\arguments{
\item{x}{numeric matrix or array}

\item{chn}{channel neighbourhood matrix}

\item{eh}{numeric vector of E and H parameters}
}
\value{
numeric matrix or array of the same dimensions as x
}
\description{
\code{tfce.fnc} performs TFCE correction. This function is not
intended for direct use.
}
\keyword{internal}

