% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matstat.R
\name{tfceFn}
\alias{tfceFn}
\title{Low-level TFCE function which calls C++ code}
\usage{
tfceFn(x, chn, eh, nr_steps = 50L, channel_dim = 1L, has_neg = TRUE,
  has_pos = TRUE)
}
\arguments{
\item{x}{numeric matrix or array}

\item{chn}{channel neighbourhood matrix}

\item{eh}{numeric vector of E and H parameters}

\item{nr_steps}{number of threshold steps (default: 50L)}

\item{channel_dim}{the dimension of x which represents channels (default: 1L)}

\item{has_neg}{logical value if there are (or can be) negative values in x 
(default: TRUE)}

\item{has_pos}{logical value if there are (or can be) positive values in x
(default: TRUE)}
}
\value{
numeric matrix or array of the same dimensions as x
}
\description{
\code{tfceFn} performs TFCE correction. This function is not 
intended for direct use.
}
\author{
The original C code was written by Christian Gaser, further modified
by Pau Coma and adopted to EEG analysis by Armand Mensen. Denes Toth ported
the C code to C++ (Rcpp) with minor improvements and added the R wrapper.
}
\keyword{internal}
