% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoplot.R
\name{topoplot}
\alias{topoplot}
\title{Topographic Plot}
\usage{
topoplot(
  x,
  sl,
  res = 200,
  scale = "auto",
  plot = c("sensorlocs", "legend"),
  col = (grDevices::colorRampPalette(c("#00007F", "blue", "#007FFF", "cyan", "#7FFF7F",
    "yellow", "#FF7F00", "red", "#7F0000")))(100),
  xlab = "",
  ylab = "",
  main = "",
  ...
)
}
\arguments{
\item{x}{Numeric array (coerced). The data to be interpolated and plotted.
The number of data points should equal the number of sensors in the
\code{sl} parameter.}

\item{sl}{\code{\link{sensorlocs}} object, or data frame with sensor
coordinates. Should at least contain the variables 'label', 'x2d' and
'y2d'.}

\item{res}{Numeric. Resolution of the output grid, Default 200.}

\item{scale}{Plot using this scale. Use "auto" (default) for automatic
scaling; otherwise pass a pair of c(min, max), e.g., c(-40,40).}

\item{plot}{Character. Specifies how the plot will look like. Levels are
always plotted, "sensorlocs" plots sensor locations as dots, "legend" plots
a legend at the right side of the plot, "contour" specifies contour lines
with values in the plot, "labels" plots sensor labels, and "all" plots all
of the above. Default: \code{c("sensorlocs", "Legend")}.}

\item{col}{color palette used for the map. Default: jet colors (as in
Matlab), interpolated using \code{\link[grDevices]{colorRampPalette}}
using 100 levels.}

\item{xlab, ylab, main}{labels passed too plotting function. Default: "".}

\item{...}{Additional parameters passed to the plotting function.}
}
\value{
A list with 4 components, returned invisibly: 
\describe{
  \item{x, y}{vectors of x- and y- coordinates of the output grid, each 'res'
  values ranging from -200 to + 200.}
  \item{z}{matrix of fitted z-values, dimensions 400 x 400. The fitted value
  at an sensor location with coordinates x2d and y2d can be found at
  \code{z[x2d, y2d]}.}
  \item{zlim}{range of z-values used in the interpolation} }
}
\description{
Draw a topographical plot of EEG data using spherical spline interpolation.
}
\details{
The function \code{topoplot} draws a topographical map of EEG data \code{x}
recorded at sensors \code{sens}. Values in between sensor locations are
interpolated using spherical splines. The plot may optionally include contour
lines, sensor positions, sensor labels, and a legend. Scaling can be manual
or automatic.

The function assumes the internal sensor coordinate system, which originates
in the middle of the head (0,0,0). See \code{\link{sensorlocs}}

Multiple topoplots can also be combined on a single page or display, using
appropriate par(mfrow = c()), par(mfcol = c()), or split.screen() commands.
In such cases, you may want to draw a single legend for all graphs. To do so,
do not specify plotting a legend in the constituent plots, but add the
legends after plotting, as in the example below.
}
\examples{
data("EEGdata")
ncoi <- 28
coi <- colnames(EEGdata)[1:ncoi]
sensors <- getLocationsfromLabels(labels = coi)

## scalp distributuon of two time points, Matlab-style 'jet' colors
op <- par(mfrow = c(1, 2))
topoplot (EEGdata[400, coi], sensors, scale = c(-10, 10), main="Sample #400");
topoplot (EEGdata[450, coi], sensors, scale = c(-10, 10), main="Sample #450");
par(op)
title(line = -2, "Example displaying two scalp maps",
      sub = "Matlab-style jet colors")

## this color scheme might be more appropriate for scalp plots
topoplot(EEGdata[400, coi], sensors, scale = c(-20, 20),
         col = grDevices::colorRampPalette(c("blue", "white", "red"))(50))

}
\seealso{
\code{\link[grDevices]{colorRampPalette}}
}
\author{
Geert van Boxtel, \email{G.J.M.vanBoxtel@gmail.com}
}
